% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{addGroupCol}
\alias{addGroupCol}
\title{Add Group Col for Corrections}
\usage{
addGroupCol(data, newColumnName, isNewCol, newGroupValue = NULL,
  groupChildValue = NULL, vars = NULL)
}
\arguments{
\item{data}{existing table data (for corrections)}

\item{newColumnName}{name of column to place new data derived}

\item{isNewCol}{function to determine if new column should be created for derived data}

\item{newGroupValue}{function for deriving new data from existing data}

\item{groupChildValue}{function to ??? TODO}
}
\value{
mutated data table (with added columns/rows)
}
\description{
helper function to create table of information for placing corrections labels
}

