% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{addMeasurementsAndError}
\alias{addMeasurementsAndError}
\title{Add Measurements and Errors
Given a gsplot object, will add measurements to the plot. Measurements points and errors styled differently if they are historical or not}
\usage{
addMeasurementsAndError(vplot, measurements, styles)
}
\arguments{
\item{vplot}{the gsplot object for the vdiagram}

\item{measurements}{measurements information}

\item{styles}{a list of styles to be used for styling the points}
}
\description{
Add Measurements and Errors
Given a gsplot object, will add measurements to the plot. Measurements points and errors styled differently if they are historical or not
}
