% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vdiagram-fnc.R
\name{addMinMax}
\alias{addMinMax}
\title{add min max horizontal lines to plot}
\usage{
addMinMax(gsplot, minStage, maxStage, ...)
}
\arguments{
\item{minStage}{y values for min and max lines}

\item{maxStage}{x values for min and max lines}

\item{...}{additional arguments passed to \code{lines}}
}
\description{
add min max horizontal lines to plot
}
\keyword{internal}

