% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{addRatingShifts}
\alias{addRatingShifts}
\title{Add Rating Shifts
Given a gsplot object, will add rating shift points, arrows, lines with callouts to the plot (The "V's")}
\usage{
addRatingShifts(vplot, shifts, styles)
}
\arguments{
\item{vplot}{the gsplot object for the vdiagram}

\item{shifts}{shift information}

\item{styles}{a list of styles to be used for styling the points/callouts}
}
\description{
Add Rating Shifts
Given a gsplot object, will add rating shift points, arrows, lines with callouts to the plot (The "V's")
}
