% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{addToGsplot}
\alias{addToGsplot}
\title{Add a feature to a gsplot object}
\usage{
addToGsplot(gsplot, plotConfig)
}
\arguments{
\item{gsplot}{A gsplot, plot object.}

\item{plotConfig}{list of gsplot calls to make}
}
\value{
A modified gsplot, plot object, with everything in the plot config included.
}
\description{
helper function to add arbitrary list of gsplot calls to a gsplot. Has added feature of extending
the Y limits for any errors bars found.
}
