% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{addYData}
\alias{addYData}
\title{Add Y Data}
\usage{
addYData(allData, height, overlapInfo, dateLim)
}
\arguments{
\item{allData}{the parsed data for the CORR report}

\item{height}{vector of calculated heights for the data for the CORR report data types}

\item{overlapInfo}{vector of any overlap details for each of the CORR report data types}

\item{dateLim}{vector of the date range for each of the CORR report data types}
}
\value{
allData containing the Y values needed for plotting and a list of
labels if the dataset was not empty and if there is room in the chart to plot it
}
\description{
Calculates the top and bottom Y values for the row based on the input data
}

