% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{adjustCorrectionTypes}
\alias{adjustCorrectionTypes}
\title{adjust correction types}
\usage{
adjustCorrectionTypes(corrections)
}
\arguments{
\item{corrections}{a corrections json object}
}
\value{
corrections formatted appropriately for TSS report
}
\description{
takes the corrections data from the api and formats
the correction types how the hydrologist expects them
}
