% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{applyApprovalBarStyles}
\alias{applyApprovalBarStyles}
\title{Apply Styles (and some properties) to Approval Bar Rectangles}
\usage{
applyApprovalBarStyles(object, data)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{data}{A list of gsplot objects to display on the plot.}
}
\value{
gsplot object with approval bar rectangle styles applied.
}
\description{
Apply Styles (and some properties) to Approval Bar Rectangles
}

