% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{applyDataGaps}
\alias{applyDataGaps}
\title{Break apart timeseries data frames for any gaps.}
\usage{
applyDataGaps(timeValueDF, startGaps, endGaps, timezone, isDV)
}
\arguments{
\item{timeValueDF}{data.frame to split with at least two columns, value and time}

\item{startGaps}{vector of dates giving the beginning of a gap period}

\item{endGaps}{vector of dates giving the beginning of a gap period}

\item{isDV}{logical saying whether or not the time series is made of daily values}

\item{timeZone}{string giving the timezone that the date time values are given}
}
\description{
Return a list of new data frames given a single data frame and the 
start/end times of gaps.
}

