% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{applyQualifiersToValues}
\alias{applyQualifiersToValues}
\title{Apply Qualifiers to Points}
\usage{
applyQualifiersToValues(points, qualifiers)
}
\arguments{
\item{points}{data frame of time/value points}

\item{qualifiers}{of qualifiers (qualifier comment per time range)}
}
\value{
updated point list with qualifiers prepended
}
\description{
given data frames of points and qualifiers, will prepended all applicable qualifiers (comma separated) to each point.
}
