% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{ApprovalBarY}
\alias{ApprovalBarY}
\title{Compute Top or Bottom Vertical Position of Approval Bars}
\usage{
ApprovalBarY(lim, ylog = NULL, reverse, ratio)
}
\arguments{
\item{lim}{The \emph{y}-axis real interval, as two element vector.}

\item{ylog}{A Boolean, indicating whether the \emph{y}-axis is \eqn{log_{10}} scale:
TRUE => \eqn{log_{10}}; FALSE => linear.}

\item{reverse}{A Boolean, indicating whether the \emph{y}-axis is inverted:
TRUE => inverted \emph{y}-axis; FALSE => not inverted.}

\item{ratio}{A scaling ratio to adjust top or bottom of approval bar rectangle.}
}
\value{
Approval bar, top or bottom \emph{y}-axis point, in world coordinates.
}
\description{
Compute Top or Bottom Vertical Position of Approval Bars
}
\author{
Andrew Halper
}

