% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{ApprovalBarYBottom}
\alias{ApprovalBarYBottom}
\title{Compute Bottom Position of Approval Bars}
\usage{
ApprovalBarYBottom(lim, ylog, reverse)
}
\arguments{
\item{lim}{The \emph{y}-axis real interval, as two element vector.}

\item{ylog}{A Boolean, indicating whether the \emph{y}-axis is \eqn{log_{10}} scale:
TRUE => \eqn{log_{10}}; FALSE => linear.}

\item{reverse}{A Boolean, indicating whether the \emph{y}-axis is inverted:
TRUE => inverted \emph{y}-axis; FALSE => not inverted.}
}
\value{
Approval bar, vertical bottom extent, in world coordinates.
}
\description{
Compute Bottom Position of Approval Bars
}
\author{
Andrew Halper
}

