% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{as.character.repgendate}
\alias{as.character.repgendate}
\title{Setup special repgendate as.character method}
\usage{
\method{as.character}{repgendate}(x, ...)
}
\arguments{
\item{x}{a date vector with class "repgendate"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Need this so that whisker.render does not drop HH:SS when there 
is only one value in a vector and it has 00:00 as the time. A 
class of repgendate is added in flexibleTimeParse.
}
