% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{as.repgendate}
\alias{as.repgendate}
\title{Setup special repgen date as method}
\usage{
as.repgendate(x)
}
\arguments{
\item{x}{a date vector}
}
\description{
For non-daily values, use a special class for the dates with 
times, so that they print correctly.
See print.repgendate and as.character.repgendate
}
