% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{attachFullDataToSubFrame}
\alias{attachFullDataToSubFrame}
\title{Attach Full Data to Sub Frame}
\usage{
attachFullDataToSubFrame(mainFrame, subFrameIdentifier)
}
\arguments{
\item{mainFrame}{The primary data frame to work on}

\item{subFrameIdentifier}{the name of the column containing the sub-frame}
}
\description{
Given a data frame with some sub-data frame will take the contents of the row of the
main frame and create a new row for each row of the sub data frame with the main data attached.
Example: 

   x  y  sub-frame
1  1  2  [1 z=3, 2 z=4]
2  3  4  [1 z=5, 2 z=6]
-----------------------
   x  y  z
1  1  2  3
2  1  2  4
3  3  4  5
4  3  4  6
}
