% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authenticateUser.R
\name{authenticateUser}
\alias{authenticateUser}
\title{authenticate a user for AQCU services}
\usage{
authenticateUser(username, password)
}
\arguments{
\item{username}{an Active Directory user name}

\item{password}{an Active Directory password}
}
\value{
an auth token from the authentication service
}
\description{
authenticate a user for AQCU services
}

