% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{boundLaneDates}
\alias{boundLaneDates}
\title{Bound Lane Dates}
\usage{
boundLaneDates(startDates, endDates, dateRange, padDays = 1)
}
\arguments{
\item{startDates}{The start dates to bound}

\item{endDates}{The end dates to bound}

\item{dateRange}{The date range to bound the dates to}

\item{padDays}{[DEFAULT: 1] The number of days to pad dates with}
}
\description{
Given start and end dates, bound them to the provided date range
}
