% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{bufferLims}
\alias{bufferLims}
\title{Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of these other points}
\usage{
bufferLims(lims, buffer.value.sequence)
}
\arguments{
\item{lims}{current limits}

\item{buffer.value.sequence}{A sequence of y values from another time series.}
}
\value{
The y-lim
}
\description{
Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of these other points
}
