% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{bufferUncorrLim}
\alias{bufferUncorrLim}
\title{Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of the uncorrected points}
\usage{
bufferUncorrLim(corr.value.sequence, uncorr.value.sequence)
}
\arguments{
\item{corr.value.sequence}{A sequence of y values from corrected series.}

\item{uncorr.value.sequence}{A sequence of y values from an uncorrected time series.}
}
\value{
The y-lim
}
\description{
Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of the uncorrected points
}
