% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{calculateLimitsAndSides}
\alias{calculateLimitsAndSides}
\title{Calculate Limits And Sides}
\usage{
calculateLimitsAndSides(primarySeriesList)
}
\arguments{
\item{primarySeriesList}{list of timeseries available for reporting}
}
\value{
named list with two items, sides and ylims. Each item has a named list with items for each timeseries. (EG: side for reference would be returnedObject$sides$reference)
}
\description{
Depending on the configuration, reference and comparison series will be plotted on different sides. This constructs ylims and what side each series should be on
}

