% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{calculateLims}
\alias{calculateLims}
\title{Calculate Lims
For a data frame of points, will calculate a lims object. X and Y field names can be configured for the points.}
\usage{
calculateLims(pts = NULL, xMinField = "time", xMaxField = "time",
  yMinField = "value", yMaxField = "value")
}
\arguments{
\item{pts}{data frame of points}

\item{xMinField}{name of the field which will contain the min x value (default "time")}

\item{xMaxField}{name of the field which will contain the max x value (default "value")}

\item{yMinField}{name of the field which will contain the min y value (default "time")}

\item{yMaxField}{name of the field which will contain the max y value (default "value")}
}
\description{
Calculate Lims
For a data frame of points, will calculate a lims object. X and Y field names can be configured for the points.
}

