% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{calculatePrimaryLims}
\alias{calculatePrimaryLims}
\title{Calculate Primary Lims}
\usage{
calculatePrimaryLims(primarySeriesList, isDischarge)
}
\arguments{
\item{primarySeriesList}{list of UV series that may appear on plot}

\item{isDischarge}{whether or not the primary series is discharge}
}
\value{
the lims object that to be applied to the primary plot
}
\description{
Will provide the lims of the primary plot given relevant data, should cover all series
}

