% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{calculateTotalDays}
\alias{calculateTotalDays}
\title{Calculate the number of days between two dates}
\usage{
calculateTotalDays(startDate, endDate)
}
\arguments{
\item{startDate}{the start date for calculating the difference in days}

\item{endDate}{the end date for calculating the difference in days}
}
\value{
days numeric number of days between the two dates
}
\description{
For two dates passed into the function returns the number of days
between the two dates
}

