% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{calculateYLim}
\alias{calculateYLim}
\title{Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of the uncorrected points}
\usage{
calculateYLim(corr.value.sequence, uncorr.value.sequence)
}
\arguments{
\item{corr.value.sequence}{A sequence of corrected time series points.}

\item{uncorr.value.sequence}{A sequence of uncorrected time series points.}
}
\value{
The y-lim
}
\description{
Compute the y lims, y-lims will ensure all of the corrected points are shown, but not necessarily all of the uncorrected points
}

