% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{checkRequiredFields}
\alias{checkRequiredFields}
\title{Returns a list of missing required fields from the given data}
\usage{
checkRequiredFields(data, requiredFields)
}
\arguments{
\item{data}{The data retrieved from a fetch function}

\item{requiredFields}{The list of fields that are required to be present}
}
\description{
Given some fetched data and required fields this function
checks the data for the existance of all requiredFields. It also will
check data returned as a data frame from a JSON array to ensure that all
array entries have the required fields.
}

