% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{cleanTempSpace}
\alias{cleanTempSpace}
\title{Clean up temporary disk space used when rendering reports}
\usage{
cleanTempSpace()
}
\description{
deletes report components after they are used in render function
and in case any old files are lingering (render crashed or errored for some reason) 
also removes anything older than 5 minutes when the function is called
}
