% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{completeQualifiers}
\alias{completeQualifiers}
\title{Complete qualifiers}
\usage{
completeQualifiers(reportObject)
}
\value{
consolidated qualifiers for all point types
}
\description{
adds qualifier metadata into qualifier section
}
