% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{constructAdvOptions}
\alias{constructAdvOptions}
\title{Parse Advanced Options settings}
\usage{
constructAdvOptions(reportData)
}
\arguments{
\item{reportData}{The full report JSON object}
}
\value{
advOptions List of applied settings and options to print on the report
}
\description{
Makes a list of user applied advanced options to print on the report
}
