% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{containsOutsideUncertainty}
\alias{containsOutsideUncertainty}
\title{Create a note on report about corrected value}
\usage{
containsOutsideUncertainty(diffData)
}
\arguments{
\item{diffData}{list of peak differences}
}
\value{
boolean of where peak differences are >0.05
}
\description{
Create a note on report about corrected value
}
