% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{convertStringToTableDisplay}
\alias{convertStringToTableDisplay}
\title{Convert a String to the Equivalent HTML Code}
\usage{
convertStringToTableDisplay(characters)
}
\arguments{
\item{characters}{The string to convert.}
}
\value{
The equivalent HTML codes for the string.
}
\description{
Convert a String to the Equivalent HTML Code
}

