% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-render.R
\name{correctionsataglanceReport}
\alias{correctionsataglanceReport}
\title{Corrections at a glance report}
\usage{
correctionsataglanceReport(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given the full report JSON object, creates a CORR report
}
