% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{createApprovalLane}
\alias{createApprovalLane}
\title{Create Approval Lane}
\usage{
createApprovalLane(approvalData, height, initialHeight, dateRange, startSeq,
  endSeq)
}
\arguments{
\item{approvalData}{The approvals to create thr lane using}

\item{height}{The height to use for lane rectangles}

\item{initialHeight}{The height to start building this lane from}

\item{dateRange}{The date range of the report}

\item{startSeq}{The sequence of all month start dates on the report}

\item{endSeq}{The sequence of all month end dates on the report}
}
\description{
Get the lane data for the approval lane. This logic
differs slightly from other lanes because the approval lane has
some unique behaviors.
}

