% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{createControlConditionsString}
\alias{createControlConditionsString}
\title{Create Control Condition String}
\usage{
createControlConditionsString(controlConditions)
}
\arguments{
\item{controlConditions}{The vector of control conditions to create the string from}
}
\description{
Creates a comma-separated string of the control conditions
}
