% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{createDataRows}
\alias{createDataRows}
\title{Create a Set of Rows for One Data Parameter}
\usage{
createDataRows(reportObject, param, rowName, isUpchain = FALSE,
  isDv = FALSE, includeRelated = TRUE, doMerge = TRUE)
}
\arguments{
\item{reportObject}{A set of extremes report data for either upchain, primary, or DV.}

\item{param}{Either "min" or "max" to specify if we are generating minimum or
maximum rows.}

\item{rowName}{The name to use for the specified row.}

\item{isUpchain}{Whether or not this is the upchain dataset and it needs to
compare to primary.}

\item{isDv}{Context is daily values when TRUE; not-daily-values otherwise.}

\item{includeRelated}{Whether or not there is a second column of
corresponding data.}

\item{doMerge}{Whether or not we should merge duplicate rows.}
}
\value{
list dataRows
}
\description{
Create a Set of Rows for One Data Parameter
}
