% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{createDVHydrographPlot}
\alias{createDVHydrographPlot}
\title{Create DV Hydrograph Plot}
\usage{
createDVHydrographPlot(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given a full report JSON object, extracts
relevant data, formats it, and then creates a DV Hydrogrpah
plot from it.
}

