% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{createDVHydrographRefPlot}
\alias{createDVHydrographRefPlot}
\title{Create DV Hdyrograph Reference Plot}
\usage{
createDVHydrographRefPlot(reportObject, series, descriptions)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{series}{the series field name to extract from the JSON}

\item{descriptions}{the description field name to extract from the JSON}
}
\description{
Given the full report object, a series field name, and
a series description field name, creates a DV Hydrograph Reference
plot using the specific parameters.
}
