% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{createGapsFromEstimatedPeriods}
\alias{createGapsFromEstimatedPeriods}
\title{Create gaps in data if there is estimated time series.}
\usage{
createGapsFromEstimatedPeriods(timeSeries, timezone, isDV, inverted = FALSE)
}
\arguments{
\item{timeSeries}{list of points, values for estimated data}

\item{timezone}{string giving the timezone}

\item{isDV}{logical, does this time series have daily values}

\item{inverted}{logical, are treating estimated
time periods as gaps (FALSE, default) or you are treating
non-estimated time periods as gaps (TRUE).}
}
\description{
If there is estimated data, then the estimated 
time series should treat non-estimated periods as gaps and 
non-estimated time-series should treat estimated periods as gaps.
}

