% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{createLabelTable}
\alias{createLabelTable}
\title{Create Label Table}
\usage{
createLabelTable(labels)
}
\arguments{
\item{labels}{The list of labels to store in the table}
}
\description{
Creates the table that holds the labels that were removed
from the plot because they didn't fit in their respective rectangles.
}

