% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{createLabelTable_old}
\alias{createLabelTable_old}
\title{Creates table to display removed labels}
\usage{
createLabelTable_old(allData, labelData, empty_nms)
}
\arguments{
\item{allData}{All the rows for the CORR report}

\item{empty_nms}{A list of the unnamed CORR report sections}
}
\value{
allData with new numText parameter identifying the label number 
for cross referencing and the table of labels/numbers to print below 
the CORR report chart
}
\description{
If labels are marked as need to be move due to not enough space, they're 
placed into a table below the CORR chart and given a number label in the chart to 
reference in the table below
}

