% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{createLane}
\alias{createLane}
\title{Create Lane}
\usage{
createLane(data, height, initialHeight, dateRange, bgColor, laneName = NULL,
  overlapInfo = NULL)
}
\arguments{
\item{data}{The dataset to create a lane for}

\item{height}{The height to use for lane rectangles}

\item{initialHeight}{The height to start building this lane from}

\item{dateRange}{The date range of the report}

\item{bgColor}{The background color to use for this lane}

\item{laneName}{[DEFAULT: NULL] The display name to use for this lane}

\item{overlapInfo}{[DEFAULT: NULL] Calculated rectangle overlap data to use}
}
\description{
Get the lane data for the provided data set
}

