% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{createOutputDir}
\alias{createOutputDir}
\title{createOutputDir}
\usage{
createOutputDir(name)
}
\arguments{
\item{name}{an identifier to use in creating temp dirs}
}
\value{
the output directory created
}
\description{
Creates a unique output director in the current working director
}
