% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{createPrimaryPlot}
\alias{createPrimaryPlot}
\title{TODO documentation}
\usage{
createPrimaryPlot(uvInfo, refInfo, compInfo, timeInformation, primarySeriesList,
  dailyValues, meas_Q, water_qual, gw_level, readings, approvalBars,
  corrections, lims)
}
\description{
TODO documentation
}

