% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{createSecondaryPlot}
\alias{createSecondaryPlot}
\title{Create Secondary Plot
Will create and configure gsPlot object using provided data}
\usage{
createSecondaryPlot(uvInfo, secondarySeriesList, approvals, effective_shift_pts,
  meas_shift, gage_height, readings, corrections, timezone,
  excludeZeroNegativeFlag, tertiary_label = "")
}
\arguments{
\item{uvInfo}{main timeseries information for the plot}

\item{secondarySeriesList}{named list of timeseries to be plotted. Also includes log/invert axis info.}

\item{approvals}{bars to be plotted which show approval level of data}

\item{effective_shift_pts}{effective shift data to be plotted (list of points)}

\item{meas_shift}{measured shift data to be plotted (list of points)}

\item{gage_height}{measured gage height data to be plotted (list of points)}

\item{readings}{named list of different readings series (ref/csg/hwm readings)}

\item{corrections}{data correction information be plotted (time/correction pairs)}

\item{timezone}{timezone to plot/display in}

\item{excludeZeroNegativeFlag}{flag to exclude ploting of values <= 0}

\item{tertiary_label}{label for the 3rd data item plotted}
}
\value{
fully configured gsPlot object ready to be plotted
}
\description{
Create Secondary Plot
Will create and configure gsPlot object using provided data
}
