% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{DelineateYearBoundaries}
\alias{DelineateYearBoundaries}
\title{Delineate Year Boundaries on Five Year GW Summary Plots, and DV Hydrographs, 
Having Time Intervals of One Year or More}
\usage{
DelineateYearBoundaries(object, years)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{years}{A sequence of year-point-type begin dates to draw the lines at.}
}
\value{
The passed-in gsplot object, with year boundaries delineated.
}
\description{
Delineate Year Boundaries on Five Year GW Summary Plots, and DV Hydrographs, 
Having Time Intervals of One Year or More
}

