% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-render.R
\name{doAddToPlot}
\alias{doAddToPlot}
\title{Do add to plot}
\usage{
doAddToPlot(data)
}
\arguments{
\item{data}{The lane data to check}
}
\description{
Returns whether or not the provided lane data should be added to the plot
}

