% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{DVHydrographRescaleYTop}
\alias{DVHydrographRescaleYTop}
\title{Rescale top of y-axis to create ~4% margin between vertical top extent of 
plot objects and top edge of plot. This is an inaccurate emulation of (the 
top-end-of-plot behavior of) R graphics::par's "yaxs = 'r'" state, because we
have to use "yaxs = 'i'" to get the approval bars aligned correctly, but
still want the ~4% margin at the top of the plot, so we adjust the y-axis
endpoint accordingly after we do what we need.}
\usage{
DVHydrographRescaleYTop(object)
}
\arguments{
\item{object}{A gsplot, plot object.}
}
\value{
The passed-in gsplot object, with y-axis top augmented (upwards).
}
\description{
Rescale top of y-axis to create ~4% margin between vertical top extent of 
plot objects and top edge of plot. This is an inaccurate emulation of (the 
top-end-of-plot behavior of) R graphics::par's "yaxs = 'r'" state, because we
have to use "yaxs = 'i'" to get the approval bars aligned correctly, but
still want the ~4% margin at the top of the plot, so we adjust the y-axis
endpoint accordingly after we do what we need.
}

