% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{enableLog}
\alias{enableLog}
\title{Safely Enable logging a side of a plot}
\usage{
enableLog(gsplot, side, minValue = 0.0095)
}
\arguments{
\item{gsplot}{The gsplot object to modify}

\item{side}{The side to enable logging on}

\item{minValue}{(Optional) [DEFAULT: 0.0095] The value to move axis lims that are at or below 0 to}
}
\description{
Helper function to enable logarithmic scaling of a side of plot safely by
forcing the limits of that side to a region tha tis usable for logs.
}
