% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{excludedConditionsMessage}
\alias{excludedConditionsMessage}
\title{Excluded Conditions Message}
\usage{
excludedConditionsMessage(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Generates the excluded conditions message to show on the plot
}
