% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{extendStep}
\alias{extendStep}
\title{use the actual last date-time value for the plot and have the straight part of the step plot extend to that point
the points do not have times, but the x limit is extended with a time to show the whole day
the step needs to be extended to meet this time}
\usage{
extendStep(toPlot, endDate)
}
\arguments{
\item{toPlot}{list of items that will be called in the do.call}

\item{endDate}{upper x limit of the plot}
}
\description{
use the actual last date-time value for the plot and have the straight part of the step plot extend to that point
the points do not have times, but the x limit is extended with a time to show the whole day
the step needs to be extended to meet this time
}

