% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{extendYaxisLimits}
\alias{extendYaxisLimits}
\title{Extend y-Axis Limits}
\usage{
extendYaxisLimits(gsplot, error_bar_args)
}
\arguments{
\item{gsplot}{The gsplot object to extend limits for}

\item{error_bar_args}{the error bar arguments to check the limits against}
}
\description{
Extends the Y-Axis limits if the error bars extend further than the current limits
}

