% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extremes.R
\name{extremes}
\alias{extremes}
\title{extremes report}
\usage{
extremes(data, output, ...)
}
\arguments{
\item{data}{}

\item{output}{}

\item{...}{additional arguments (e.g., username & password for authentication)}
}
\description{
#
}

