% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesQualifiersTable}
\alias{extremesQualifiersTable}
\title{create flat text 'qualifiers table' type output table}
\usage{
extremesQualifiersTable(data, table)
}
\arguments{
\item{data}{report data}
}
\value{
string table
}

