% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-render.R
\name{extremesReport}
\alias{extremesReport}
\title{Render Extremes report}
\usage{
extremesReport(reportObject)
}
\arguments{
\item{reportObject}{extremes report json object}
}
\description{
given a report object representing an extremes report, will create an HTML report
}

