% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesTable}
\alias{extremesTable}
\title{create a flat text 'extremes table' type output table}
\usage{
extremesTable(rawData)
}
\arguments{
\item{rawData}{extremes report json string}
}
\value{
string table
}

