% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchApprovals}
\alias{fetchApprovals}
\title{Fetch Approvals (TSS)}
\usage{
fetchApprovals(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the approvals
}
