% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchApprovalsForSeries}
\alias{fetchApprovalsForSeries}
\title{Fetch Approvals for a given Time Series}
\usage{
fetchApprovalsForSeries(reportObject, seriesName)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}

\item{field}{The specific field to select from the metadata}
}
\description{
Given a full report object this will extract the
approvals for the supplied series name.
}

