% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchCorrections}
\alias{fetchCorrections}
\title{Fetch corrections for a time series}
\usage{
fetchCorrections(reportObject, seriesCorrName)
}
\arguments{
\item{reportObject}{the full report data}

\item{seriesCorrName}{the time series name to pull corrections for}
}
\description{
Given a report object and series name, will pull the corrections
list for the specified time series
}

