% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchCorrReportURL}
\alias{fetchCorrReportURL}
\title{Fetch CORR Report URL (TSS)}
\usage{
fetchCorrReportURL(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the URL for the associated CORR report
}
