% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchCurveNumber}
\alias{fetchCurveNumber}
\title{Fetch the curve number/rating}
\usage{
fetchCurveNumber(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the curve number
}

