% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchDownchainSeries}
\alias{fetchDownchainSeries}
\title{Fetch Downchain Series (TSS)}
\usage{
fetchDownchainSeries(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the downchain related series
}
