% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchExcludedControlConditions}
\alias{fetchExcludedControlConditions}
\title{Fetch Excluded Control Conditions (V-Diagram)}
\usage{
fetchExcludedControlConditions(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the excluded control conditions
}
