% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchExtremesSeriesQualifiers}
\alias{fetchExtremesSeriesQualifiers}
\title{Fetch Extremes Primary Series Qualifiers (Extremes)}
\usage{
fetchExtremesSeriesQualifiers(reportObject, qualType)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}

\item{qualType}{the qualifier type we're looking to fetch}
}
\description{
Given a full report object this will extract the data 
for primary series qualifiers.
}
