% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchFieldVisitMeasurements}
\alias{fetchFieldVisitMeasurements}
\title{Fetch field visit measurements}
\usage{
fetchFieldVisitMeasurements(reportObect)
}
\arguments{
\item{reportObect}{the full report data}
}
\description{
Given a report object, will pull the field visit measurements
}

