% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchFieldVisitReadings}
\alias{fetchFieldVisitReadings}
\title{Fetch field visit readings}
\usage{
fetchFieldVisitReadings(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the field visit readings
}

