% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchFieldVists}
\alias{fetchFieldVists}
\title{Fetch Field Visits (CORR)}
\usage{
fetchFieldVists(reportObject)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}
}
\description{
Given a full report object this will extract the field
vists data.
}
