% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchGradeMetadata}
\alias{fetchGradeMetadata}
\title{Fetch Grade Metadata (TSS)}
\usage{
fetchGradeMetadata(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the grade metadata
}
