% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchGroundWaterLevels}
\alias{fetchGroundWaterLevels}
\title{Fetch ground water levels}
\usage{
fetchGroundWaterLevels(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the ground water levels
}
