% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMaxStage}
\alias{fetchMaxStage}
\title{Fetch maximum stage height}
\usage{
fetchMaxStage(reportObject)
}
\arguments{
\item{reportObject}{a report object}
}
\value{
numeric value for max stage
}
\description{
Given a report object will pull the max stage value.
}
