% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMeasurements}
\alias{fetchMeasurements}
\title{Fetch Discharge measurements}
\usage{
fetchMeasurements(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\value{
The list of measurements attached to the report. If none, will be NULL.
}
\description{
Given a report object, will attempt to pull the measurements list.
}

