% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMeasurementsField}
\alias{fetchMeasurementsField}
\title{Fetch Measurements Field}
\usage{
fetchMeasurementsField(reportObject, field)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}

\item{field}{The specific field to select from the measurements}
}
\description{
Given a full report object this will extract the data
associated with the specified field.
}

