% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMethods}
\alias{fetchMethods}
\title{Fetch Methods (TSS)}
\usage{
fetchMethods(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the primary TS methods
}
