% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMinMaxIVsDV}
\alias{fetchMinMaxIVsDV}
\title{Fetch Min/Max IV for DV Hydro Report}
\usage{
fetchMinMaxIVsDV(reportObject, stat)
}
\arguments{
\item{reportObject}{the full report data}

\item{stat}{'min' or 'max' data to select}
}
\description{
Given a report object, will pull the min or max IV points data
}
