% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchMinStage}
\alias{fetchMinStage}
\title{Fetch minimum stage height}
\usage{
fetchMinStage(reportObject)
}
\arguments{
\item{reportObject}{a report object}
}
\value{
numeric value for min stage
}
\description{
Given a report object will pull the min stage value.
}
