% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchNotes}
\alias{fetchNotes}
\title{Fetch Notes (TSS)}
\usage{
fetchNotes(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the notes
}
