% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchNumberOfShifts}
\alias{fetchNumberOfShifts}
\title{Fetch number of Shifts}
\usage{
fetchNumberOfShifts(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the number of shifts
}

