% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchPrimarySeriesApprovals}
\alias{fetchPrimarySeriesApprovals}
\title{Fetch Primary Series Approvals (DV and Five YR)}
\usage{
fetchPrimarySeriesApprovals(reportObject)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}
}
\description{
Given a full report object this will extract the data 
for primary series approvals.
}

