% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchPrimarySeriesQualifiers}
\alias{fetchPrimarySeriesQualifiers}
\title{Fetch Primary Series Qualifiers (DV and Five YR)}
\usage{
fetchPrimarySeriesQualifiers(reportObject)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}
}
\description{
Given a full report object this will extract the data 
for primary series qualifiers.
}
