% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchPrimaryTSMetadata}
\alias{fetchPrimaryTSMetadata}
\title{Fetch Primary TS Metadata (TSS)}
\usage{
fetchPrimaryTSMetadata(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the primary TS metadata
}
