% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchProcessingCorrections}
\alias{fetchProcessingCorrections}
\title{Fetch Processing Corrections (CORR)}
\usage{
fetchProcessingCorrections(reportObject, processOrder)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}

\item{processOrder}{The processing order to get corrections for. Valid
choices: "pre", "post", and "normal"}
}
\description{
Given a full report object this will extract the corrections
data for the specified processing order.
}

