% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchRatingShifts}
\alias{fetchRatingShifts}
\title{Fetch Rating Shifts}
\usage{
fetchRatingShifts(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\value{
The list of ratingShifts attached to the report. If none, will be NULL.
}
\description{
Given a report object, will attempt to pull the rating shifts list.
}
