% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchRatingShiftsField}
\alias{fetchRatingShiftsField}
\title{Fetch Rating Shifts Field}
\usage{
fetchRatingShiftsField(reportObject, field)
}
\arguments{
\item{reportObject}{The full report data loaded from the report JSON}

\item{field}{The specific field to select from the rating shifts}
}
\description{
Given a full report object this will extract the data
associated with the specified field.
}

