% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchReportField}
\alias{fetchReportField}
\title{Fetch Field}
\usage{
fetchReportField(reportObect, field)
}
\arguments{
\item{reportObect}{The full report data loaded from the report JSON}

\item{field}{The field to fetch from the data}
}
\description{
Generalize fetch function that fetches the provided field
}

