% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchShiftPoints}
\alias{fetchShiftPoints}
\title{Fetch the shift points}
\usage{
fetchShiftPoints(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the shift points
}

