% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchSRSQualifiers}
\alias{fetchSRSQualifiers}
\title{Fetch Qualifiers (SRS)}
\usage{
fetchSRSQualifiers(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a reading this will extract the qualifiers
}
