% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchStagePoints}
\alias{fetchStagePoints}
\title{Fetch the stage points}
\usage{
fetchStagePoints(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the stage points
}

