% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchStartTime}
\alias{fetchStartTime}
\title{Fetch the start time for the data}
\usage{
fetchStartTime(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the start time
}

