% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchThresholds}
\alias{fetchThresholds}
\title{Fetch Threshold Data (CORR)}
\usage{
fetchThresholds(reportObject)
}
\description{
Given a full report object this will extract the threshold extremes-data.R
}

