% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchTimeSeries}
\alias{fetchTimeSeries}
\title{Fetch time series}
\usage{
fetchTimeSeries(reportObject, seriesName)
}
\arguments{
\item{reportObject}{the full report data}

\item{seriesName}{the time series name to fetch}
}
\description{
Given a report object, will pull time series with given name
}
