% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchTSSThresholds}
\alias{fetchTSSThresholds}
\title{Fetch Threshold Data (TSS)}
\usage{
fetchTSSThresholds(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the thresholds
}
