% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{fieldExists}
\alias{fieldExists}
\title{fieldExists}
\usage{
fieldExists(data, field)
}
\arguments{
\item{data}{list object}

\item{field}{string for name to look for}
}
\value{
true/false
}
\description{
given any list data will detrminte if the named item exists
}

