% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{filterAndMarkDuplicates}
\alias{filterAndMarkDuplicates}
\title{Filter and Mark Duplicates}
\usage{
filterAndMarkDuplicates(extremesRows, note, includeRelated, fieldToCheck)
}
\arguments{
\item{extremesRows}{the data frame of extremes data as characters (fields: name, date, time, primary, related)}

\item{note}{a note to append to the date value of the remaining duplicate row}

\item{includeRelated}{true to include the related field in the returned data frame}

\item{fieldToCheck}{the field compare for duplicate values}
}
\description{
Given a list of rows, will remove duplicates. The first row of a set of duplicates will be kept, and it's date value marked with the note provided
}
