% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{findDefinedGaps}
\alias{findDefinedGaps}
\title{Find the start and end times for any possible data gaps.}
\usage{
findDefinedGaps(timeSeries, timezone)
}
\arguments{
\item{timeSeries}{list of points, values, gaps}

\item{timezone}{string giving the timezone}
}
\description{
Find the start and end times
}

