% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{findOverlap}
\alias{findOverlap}
\title{Checks for overlapping positions in order to make new lines when needed}
\usage{
findOverlap(dataList)
}
\arguments{
\item{dataList}{The data to check for overlapping dates}
}
\value{
The number of lines to add and a list of the data that need to be
shifted
}
\description{
The function checks for overlapping date ranges and if there
is overlap, adds a line to the chart and puts the overlapping data there
}

