% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{findTextLocations}
\alias{findTextLocations}
\title{Find location for labels}
\usage{
findTextLocations(dataIn, yTop, yBottom, isDateData = FALSE)
}
\arguments{
\item{dataIn}{The dataset to identify text label locations for}

\item{yTop}{The upper Y-bound of the lane}

\item{yBottom}{The lower Y-bound of the lane}

\item{isDateData}{A flag indicating if the incoming label data are dates or not}
}
\description{
For each corr report section, determine where to place the label
}
