% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-fiveyeargwsum.R
\docType{methods}
\name{fiveyeargwsum}
\alias{fiveyeargwsum}
\alias{fiveyeargwsum,character,character-method}
\alias{fiveyeargwsum,list,character-method}
\alias{fiveyeargwsum,list,missing-method}
\title{Five-year Groundwater Summary report}
\usage{
fiveyeargwsum(data, ...)

\S4method{fiveyeargwsum}{list,character}(data, ...)

\S4method{fiveyeargwsum}{character,character}(data, ...)

\S4method{fiveyeargwsum}{list,missing}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET or authenticateUser}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
data <- fromJSON(system.file('extdata','fiveyeargwsum','fiveyeargwsum-example.json', package = 'repgen'))
fiveyeargwsum(data, 'Author Name')

}

