% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{flexibleTimeParse}
\alias{flexibleTimeParse}
\title{will attempt to parse a DV, UTC time, or offset time}
\usage{
flexibleTimeParse(x, timezone)
}
\arguments{
\item{x}{the date/time}

\item{timezone}{a timezone code}
}
\value{
time vector
}
\description{
convienence that will attempt to parse a DV, UTC time, or offset time
extremes json
}

