% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatAdvReportOptions}
\alias{formatAdvReportOptions}
\title{Format TSS Advanced Report Options}
\usage{
formatAdvReportOptions(advancedReportOptions)
}
\arguments{
\item{advancedReportOptions}{The param to format}
}
\value{
advOptions List of applied options to print on the report
}
\description{
Format user applied advanced options to print on the report
}
