% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{formatComments}
\alias{formatComments}
\title{Takes comments formatted with special escape characters and converts them to HTML breaks
and includes a break at the end between comments}
\usage{
formatComments(comments)
}
\arguments{
\item{comments}{which have \\r\\n indicating line breaks}
}
\value{
comments formatted with html breaks instead of escape pattern
}
\description{
Formats comments more effectively for including them in HTML table
}

