% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionParameters}
\alias{formatCorrectionParameters}
\title{Format TSS Correction Parameters}
\usage{
formatCorrectionParameters(corrections, timezone)
}
\arguments{
\item{corrections}{the corrections for the report}

\item{timezone}{the timezone to parse data into}
}
\value{
corrections with formattedParameters field to print in report
}
\description{
formats the parameters so they are prettier to print
}
