% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamAdjustableTrim}
\alias{formatCorrectionsParamAdjustableTrim}
\title{formats the adjustable trim correction parameters}
\usage{
formatCorrectionsParamAdjustableTrim(upperThresholdPoints, lowerThresholdPoints,
  timezone)
}
\arguments{
\item{upperThresholdPoints}{upper threshold points date/times and values as a list}

\item{timezone}{the timezone for the drift parameters
#' @return formatted string of adjustable trim parameters for report display}
}
\description{
formats the adjustable trim correction parameters
}
