% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamDeviation}
\alias{formatCorrectionsParamDeviation}
\title{formats the deviation correction parameters}
\usage{
formatCorrectionsParamDeviation(deviationValue, deviationType,
  windowSizeInMinutes)
}
\arguments{
\item{deviationValue}{the value for the deviation}

\item{deviationType}{a description of the type of deviation for this correction}

\item{windowSizeInMinutes}{the window for the deviation in minutes}
}
\value{
formatted string of deviation parameters for report display
}
\description{
formats the deviation correction parameters
}
