% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamDrift}
\alias{formatCorrectionsParamDrift}
\title{formats the drift correction parameters}
\usage{
formatCorrectionsParamDrift(driftPoints, timezone)
}
\arguments{
\item{driftPoints}{the driftPoints values in a list including
date/times and the difference values in feet}

\item{timezone}{the timezone for the drift parameters}
}
\value{
formatted string of drift parameters for report display
}
\description{
formats the drift correction parameters
}
