% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamFillGaps}
\alias{formatCorrectionsParamFillGaps}
\title{formats the fill gaps correction parameters}
\usage{
formatCorrectionsParamFillGaps(resamplePeriod, gapLimit)
}
\arguments{
\item{resamplePeriod}{a description of the resample period}

\item{gapLimit}{a description of the gap limits}
}
\value{
formatted string of fill gaps parameters for report display
}
\description{
formats the fill gaps correction parameters
}
