% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamOffset}
\alias{formatCorrectionsParamOffset}
\title{formats the offset correction parameters}
\usage{
formatCorrectionsParamOffset(offset)
}
\arguments{
\item{offset}{the offset value for the parameter offset}
}
\value{
formatted string of offset parameters for report display
}
\description{
formats the offset correction parameters
}
