% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamSinglePoint}
\alias{formatCorrectionsParamSinglePoint}
\title{formats the single point correction parameters}
\usage{
formatCorrectionsParamSinglePoint(value)
}
\arguments{
\item{value}{the value for the parameter single point}
}
\value{
formatted string of single point parameter for report display
}
\description{
formats the single point correction parameters
}
