% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatCorrectionsParamUSGSMultiPoint}
\alias{formatCorrectionsParamUSGSMultiPoint}
\title{formats the USGSMultiPoint correction parameters}
\usage{
formatCorrectionsParamUSGSMultiPoint(startShiftPoints, endShiftPoints, usgsType)
}
\arguments{
\item{startShiftPoints}{startShift points values and offset as a list}

\item{endShiftPoints}{endShift points values and offset as a list}

\item{usgsType}{the USGSMultiPoint type}
}
\value{
formatted string of USGSMultiPoint parameters for report display
}
\description{
formats the USGSMultiPoint correction parameters
}
