% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{formatDataList}
\alias{formatDataList}
\title{Formats data coming into CORR report}
\usage{
formatDataList(dataIn, type, timezone, ...)
}
\arguments{
\item{dataIn}{The data type}

\item{type}{Whether the data are approvals, metadata, or pre, post 
or normal processing data}

\item{timezone}{The timezone for the dataset}

\item{...}{Additional arguments passed into function}
}
\value{
The data formatted appropriate for the its type
}
\description{
Formats data for downstream CORR report generation
}

