% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatDataTable}
\alias{formatDataTable}
\title{Format Data Table}
\usage{
formatDataTable(inputData)
}
\arguments{
\item{inputData}{the data to format}
}
\description{
Formats a given dataframe into a whisker table
}
