% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{formatMinMaxLabel}
\alias{formatMinMaxLabel}
\title{Format Min Max Labels (DV hYdro and Five YR)}
\usage{
formatMinMaxLabel(ml, units)
}
\arguments{
\item{ml}{The min / max IV label object to format}

\item{units}{the units to use for the IV label}
}
\description{
Formats a min/max IV as a lable to be put on the
top of the plot.
}
