% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{formatOpenDateLabel}
\alias{formatOpenDateLabel}
\title{Format Open Date Label}
\usage{
formatOpenDateLabel(dates)
}
\arguments{
\item{dates}{The date to format}
}
\description{
Replace open ended dates (1700 or less and 9999) with "Open"
}
