% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{formatPersistenceGapFill}
\alias{formatPersistenceGapFill}
\title{formats the persistence fill gaps correction parameters}
\usage{
formatPersistenceGapFill(resamplePeriod, gapLimit,
  resampleInterpolationType)
}
\arguments{
\item{resamplePeriod}{a description of the resample period}

\item{gapLimit}{a description of the gap limits}

\item{resampleInterpolationType}{a description of the resample interpolation type}
}
\value{
formatted string of persistence fill gaps parameters for report display
}
\description{
formats the persistence fill gaps correction parameters
}
