% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{formatQualifiersTable}
\alias{formatQualifiersTable}
\title{Create Flat Text, "qualifiers table" Type Output Table}
\usage{
formatQualifiersTable(inQualifiers)
}
\arguments{
\item{inQualifiers}{data frame of filtered (for SVP) or all (for SRS) qualifiers.}
}
\value{
list of deduplicated qualifiers with column names.
}
\description{
Create Flat Text, "qualifiers table" Type Output Table
}

