% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{formatSensorData}
\alias{formatSensorData}
\title{Creates the formatted data.frame for the report}
\usage{
formatSensorData(readings, columnNames, includeComments, timezone,
  qualifierMetadata)
}
\arguments{
\item{readings}{Sensor reading report readings JSON string.}

\item{columnNames}{list of column names for the report}

\item{includeComments}{flag for TRUE or FALSE depending on user selection on 
whether they want comments included in the report output}

\item{timezone}{the timezone of the report}

\item{qualifierMetadata}{the metadata for all qualifiers found in readings}
}
\value{
data.frame table
}
\description{
Takes a JSON data string, a list of column names and a flag 
for comments and returns formatted data.frame for the report
}
