% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{formatSplitTimeSeriesForPlotting}
\alias{formatSplitTimeSeriesForPlotting}
\title{Format split time series plotting}
\usage{
formatSplitTimeSeriesForPlotting(seriesList,
  excludeZeroNegativeFlag = NULL)
}
\arguments{
\item{seriesList}{The time series list to format for plotting}

\item{excludeZeroNegativeFlag}{Whether or not zero and negative values should be removed}
}
\value{
A list of formated time series data frames
}
\description{
Helper function for formating a spearated list of time series
that are all part of the same overall time series. This is primarily used for
time series that have been split up because of gaps caused by estimated periods
or actual gaps in data.
}
