% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{formatSrsTime}
\alias{formatSrsTime}
\title{Formats the date/time data for the report columns Date and Time or CorrectedDate and CorrectedTime 
in the Field Visit Info section of the SRS report}
\usage{
formatSrsTime(displayTime, time)
}
\arguments{
\item{displayTime}{the JSON data field containing the yyyy-mm-ddTHH:MM:SS data}

\item{time}{The JSON data field containing the time if it exists}
}
\value{
the date formatted as mm/dd/yyyy and the timeFormatted as HH:MM:SS (UTC -XX:XX)
}
\description{
accepts the display time and time from the JSON and formats it for the report
}

