% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{formatSVPData}
\alias{formatSVPData}
\title{create a flat text 'sitevisitpeak table' type output table}
\usage{
formatSVPData(data, columnNames)
}
\arguments{
\item{rawData}{sitevisitpeak report json string}
}
\value{
data.frame table
}
\description{
create a flat text 'sitevisitpeak table' type output table
}

