% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{formatThresholdsData}
\alias{formatThresholdsData}
\title{Formats the threshold correction type}
\usage{
formatThresholdsData(thresholds)
}
\arguments{
\item{thresholds}{the threshold data from the JSON}
}
\value{
the formatted threshold data for the chart
}
\description{
Takes threshold data and if not suppressed
they will then be plotted on the chart
}

