% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{formatUTCTimeLabel}
\alias{formatUTCTimeLabel}
\title{Format Time as a label with the UTC Offset}
\usage{
formatUTCTimeLabel(time)
}
\arguments{
\item{time}{The time datum to format as a text label}
}
\description{
Formats a time datum as a text
label including the UTC offset at the end.
}
