% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getAppliedCorrection}
\alias{getAppliedCorrection}
\title{get applied correction}
\usage{
getAppliedCorrection(raw, corrected)
}
\arguments{
\item{raw}{The raw reading value}

\item{corrected}{The corrected reading value}
}
\value{
The rounded difference between the raw and corrected values or 
empty character if the raw and corrected value passed in are null or empty
}
\description{
Takes a raw and corrected value and calculates the applied correction
if they are not null or empty
}
