% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{getApprovalBarConfig}
\alias{getApprovalBarConfig}
\title{Get Approval Bar Configuration}
\usage{
getApprovalBarConfig(approvals, ylim, ylog, reverse)
}
\arguments{
\item{approvals}{A list of data objects relevant to plotting approval
bars.}

\item{ylim}{The \emph{y}-axis interval, as ordered pair vector.}

\item{ylog}{A Boolean truth value. \code{TRUE} indicates the \emph{y}-axis is
referenced to log10; linear \emph{y}-axis otherwise.}

\item{reverse}{A Boolean truth value. \code{TRUE} indicates the \emph{y}-axis
is inverted; not inverted otherwise.}
}
\value{
A named list of gsplot calls. The name is the plotting call to make,
  and it points to a list of configuration parameters for that call.
}
\description{
Given approval bars, return a named list of gsplot elements to
  call.
}
\author{
Andrew Halper
}
