% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{getApprovalColors}
\alias{getApprovalColors}
\title{Returns the approval colors}
\usage{
getApprovalColors(approvalLevels)
}
\arguments{
\item{approvalLevels}{A list of approval levels from the data}
}
\value{
a list of colors in order of the approvalLevels passed in
}
\description{
for approval present in the data passed in, returns
the approval colors needed in order 
0=Working (red), 1=In Review (yellow), 2=Approved (green), 
any other value is colored as 'grey'
}

