% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-auth.R
\name{getAuth}
\alias{getAuth}
\title{Get authorization header.}
\usage{
getAuth()
}
\value{
NULL if no auth token is set; authorization header as string
       otherwise.
}
\description{
Get authorization header.
}
\keyword{internal}

