% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{getComments}
\alias{getComments}
\title{Returns a list of comments or an empty character if there are no comments}
\usage{
getComments(comments)
}
\arguments{
\item{comments}{The text comments from the JSON data}
}
\value{
comments as they were passed or an empty string if empty or null
}
\description{
Accepts the comments string and checks to see if it's null or empty,
and if it is, returns an empty string
}

