% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getCorrectedRef}
\alias{getCorrectedRef}
\title{Are the corrected values within the reference values accounting for 
their specified uncertainty?}
\usage{
getCorrectedRef(value, nearestcorrectedValue, uncertainty)
}
\arguments{
\item{value}{The reading value}

\item{nearestcorrectedValue}{The nearest corrected value to the reading value}

\item{uncertainty}{The uncertainty specified for the reading value}
}
\value{
Yes or No if the corrections applied are within the reference
values for their specified uncertainty
}
\description{
This function returns a yes or no when comparing whether the 
corrections applied to the recorder reading are within the reference values
accounting for their specified uncertainty?
}

