% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getCorrectionAbLinesPositions}
\alias{getCorrectionAbLinesPositions}
\title{Get AB line positions for Corrections
Given a list of corrections, will return the time(x) position of each}
\usage{
getCorrectionAbLinesPositions(corrections)
}
\arguments{
\item{corrections}{a list of corrections}
}
\value{
list of time/x for each correction
}
\description{
Get AB line positions for Corrections
Given a list of corrections, will return the time(x) position of each
}

