% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getCorrectionArrows}
\alias{getCorrectionArrows}
\title{Get Correction Arrows
For a set of correction labels, will return a list of arrows to connect label to correction line.}
\usage{
getCorrectionArrows(correctionLabels)
}
\arguments{
\item{correctionLabels}{list of labels with positioning information already calculated/attached}
}
\value{
list of data describing how to draw lines to corresponding labels
}
\description{
Get Correction Arrows
For a set of correction labels, will return a list of arrows to connect label to correction line.
}
