% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getCorrectionPositions}
\alias{getCorrectionPositions}
\title{Get positions for Corrections
Given a list of corrections, will return a list of times (with duplicates removed) which are the x position of vertical lines}
\usage{
getCorrectionPositions(corrections)
}
\arguments{
\item{corrections}{a list of corrections}
}
\value{
list of time/x for each correction
}
\description{
Get positions for Corrections
Given a list of corrections, will return a list of times (with duplicates removed) which are the x position of vertical lines
}

