% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getCorrections}
\alias{getCorrections}
\title{Extract corrections from a time series data structure.}
\usage{
getCorrections(ts, field)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{field}{The name of the field in "ts" parameter, to extract as
corrections.}
}
\description{
Extract corrections from a time series data structure.
}

