% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{getDate}
\alias{getDate}
\title{get a date from json list in specified outputformat}
\usage{
getDate(ts, param, format = "\%Y/\%m/\%d")
}
\arguments{
\item{ts}{a timeseries list that can be extracted w/ \code{\link{getRaw}}}

\item{param}{the json parameter name to match to (passed to \code{\link{getRaw}})}

\item{format}{string: the date format output.}
}
\value{
a string date, or empty if the value returned from \code{\link{getRaw}} is empty
}
\description{
get a date from json list in specified outputformat
}
\seealso{
\code{\link{getRaw}}
}

