% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{getDVHydrographPlotConfig}
\alias{getDVHydrographPlotConfig}
\title{Get DV Hydrograph Plot Config}
\usage{
getDVHydrographPlotConfig(plotItem, plotItemName, yLabel = "",
  lineOffset = 1, ...)
}
\arguments{
\item{plotItem}{the item to fetch styles and plot features for}

\item{plotItemName}{the string to use for matching the configuration and styles}

\item{yLabel}{the string to use for the Y-Axis label for this object (if applicable) (Default: "")}

\item{...}{any additional parameters to pass into the function}
}
\description{
Given an item to plot, fetch the associated
plot feature(s) and their styles.
}
