% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getEffectiveShiftPlotConfig}
\alias{getEffectiveShiftPlotConfig}
\title{Get Effective Shift Plot Config}
\usage{
getEffectiveShiftPlotConfig(effective_shift, secondary_lbl, tertiary_lbl)
}
\arguments{
\item{effective_shift}{list of effective shift points}

\item{secondary_lbl}{label of secondary time series}

\item{tertiary_lbl}{label of tertiary time series}
}
\description{
Given a report object, some information about the plot to build, will return a named list of gsplot elements to call
}

