% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{getErrorBarYLims}
\alias{getErrorBarYLims}
\title{Get Error Bar Y-Limits}
\usage{
getErrorBarYLims(error_bar_args)
}
\arguments{
\item{error_bar_args}{The error bar arguments to return.}
}
\description{
Gets the min and max y-value of error bars and returns a side if there
is a specific side the bars are associated with.
}

