% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{getEstimatedEdges}
\alias{getEstimatedEdges}
\title{Create vertical step edges between estimated and non-estimated series}
\usage{
getEstimatedEdges(stat, est)
}
\arguments{
\item{stat}{the parsed non-estimated time series}

\item{est}{the parsed estimated time series}
}
\value{
a list of vertical lines connecting steps between stat and est

a list of the edges with the folowing columns time, y0, y1,
and newSet. Time is the x position of the edge, y0 is the first y
value and y1 is the second. newSet is what the next time series is
}
\description{
Given a stat TS and an estimated TS this function creates
vertical lines connecting the steps between those TS.
}

