% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{getExtremesTableQualifiers}
\alias{getExtremesTableQualifiers}
\title{Extract Qualifiers Used}
\usage{
getExtremesTableQualifiers(table, primaryHeaderTerm, upchainHeaderTerm)
}
\arguments{
\item{table}{table/list representation (already processed data) of an extremes report.}

\item{primaryHeaderTerm}{a unique term to help determine which column title is for primary data}

\item{upchainHeaderTerm}{a unique term to help determine which column title is for upchain data}
}
\value{
unique list of qualifier strings found in Primary/Unpchain columns
}
\description{
Given an extremes table, look through the primary/upchain columns for qualifiers
}
