% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getFieldVisitMeasurementsShifts}
\alias{getFieldVisitMeasurementsShifts}
\title{Extract field visit measurement shifts from a time series data structure.}
\usage{
getFieldVisitMeasurementsShifts(ts)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}
}
\description{
Extract field visit measurement shifts from a time series data structure.
}

