% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveyeargwsum-data.R
\name{getFiveYrTimeSeries}
\alias{getFiveYrTimeSeries}
\title{Get Time Series}
\usage{
getFiveYrTimeSeries(reportObject, series, description, timezone,
  estimated = FALSE)
}
\arguments{
\item{reportObject}{the full report JSON}

\item{series}{the name of the series to load}

\item{description}{the name of the field to use for the TS label}

\item{timezone}{the timezone to parse the time series points into}
}
\description{
DVHydrograph readTimeSeries() tryCatch wrapper to
to properly read in time series and handle read errors.
}

