% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getGroundWaterLevels}
\alias{getGroundWaterLevels}
\title{Extract and Restructure Groundwater Levels Data}
\usage{
getGroundWaterLevels(ts, ...)
}
\arguments{
\item{ts}{A time series data frame.}

\item{...}{Unknown. Possibly obsolete.}
}
\description{
Extract and Restructure Groundwater Levels Data
}

