% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getGwPlotConfig}
\alias{getGwPlotConfig}
\title{Get GW level Plot Config}
\usage{
getGwPlotConfig(gw_level)
}
\arguments{
\item{gw_level}{list of GW level objects}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a gw level readings, will return a named list of gsplot elements to call
}
