% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{getInput}
\alias{getInput}
\title{get input from extremes json list}
\usage{
getInput(ts, param, ...)
}
\arguments{
\item{ts}{a list, can be the output of \code{\link[jsonlite]{fromJSON}}.}

\item{param}{the field name (e.g., 'endDate')}

\item{...}{additional arguments passed to \code{repgen:::validParam},
such as \code{required}, or \code{as.numeric}}
}
\value{
a value or array corresponding to the field specified by \code{param}
}
\description{
convienence function for accessing from the "inputs" block in
extremes json
}

