% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{getLaneLabelData}
\alias{getLaneLabelData}
\title{Get Lane Label Data}
\usage{
getLaneLabelData(data, laneYTop, laneYBottom, dateRange,
  isDateData = FALSE)
}
\arguments{
\item{data}{The lane data to get label positions and text for}

\item{laneYTop}{The upper Y-bound of the lane}

\item{laneYBottom}{The lower Y-bound of the lane}

\item{dateRange}{The date range of the report}

\item{isDateData}{[DEFAULT: FALSE] Whether or not the data is just dates}
}
\value{
data frame of labels
}
\description{
Get the label positions and text for the specified lane
}
