% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getMaxStage}
\alias{getMaxStage}
\title{Compute Maximum Stage Value}
\usage{
getMaxStage(ts, ...)
}
\arguments{
\item{ts}{A time series data structure.}

\item{...}{Optional parameters forwarded to \code{validParam()}.}
}
\description{
Compute Maximum Stage Value
}

