% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getMeasurements}
\alias{getMeasurements}
\title{Extract measurements from a time series data structure.}
\usage{
getMeasurements(ts, param, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{param}{A field name.}

\item{...}{Optional parameters.}
}
\description{
Extract measurements from a time series data structure.
}

