% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{getMinMaxIV}
\alias{getMinMaxIV}
\title{Get the Min/Max IV Data}
\usage{
getMinMaxIV(reportObject, stat, timezone, tsType, inverted)
}
\arguments{
\item{reportObject}{the full report data}

\item{stat}{the stat to look up (MAX or MIN)}

\item{timezone}{the timezone to parse the times into}

\item{tsType}{the type of the TS (to use for the legend name)}

\item{inverted}{whether or not the TS is inverted}
}
\description{
Reads the Max/Min IV Data from the reportObject then takes the
first entry and formats it properly for use on the DV Hydrograph report.
}

