% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getMinStage}
\alias{getMinStage}
\title{Compute Minimum Stage Value}
\usage{
getMinStage(ts, ...)
}
\arguments{
\item{ts}{A time series data structure.}

\item{...}{Optional parameters forwarded to \code{validParam()}.}
}
\description{
Compute Minimum Stage Value
}

