% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getMonths}
\alias{getMonths}
\title{Get months}
\usage{
getMonths(reportObject)
}
\arguments{
\item{reportObject}{a UV Hydrograph report}
}
\value{
list of months, sorted, as 'YYMM'
}
\description{
For a report, return a list of all months with data (sorted) in format 'YYMM'.
}

