% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getRatingShifts}
\alias{getRatingShifts}
\title{Extract rating shifts from a time series data structure.}
\usage{
getRatingShifts(ts, param, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{param}{A field name.}

\item{...}{Optional parameters.}
}
\description{
Extract rating shifts from a time series data structure.
}

