% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getReadingsPlotConfig}
\alias{getReadingsPlotConfig}
\title{Get Readings Plot Config}
\usage{
getReadingsPlotConfig(reading_type, readings)
}
\arguments{
\item{reading_type}{name of reading type to get style for (ref, csg, hwm)}

\item{readings}{list of readings objects relavant to primary plot}
}
\description{
Given a readings list, will pull the desired plotting calls and styling
}

