% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getRecorderWithinUncertainty}
\alias{getRecorderWithinUncertainty}
\title{calculate the recorder w/in uncertainty}
\usage{
getRecorderWithinUncertainty(uncertainty, value, recorderValue)
}
\arguments{
\item{uncertainty}{The value for uncertainty for the reading}

\item{value}{The value for the reading}

\item{recorderValue}{The recorderValue for the reading}
}
\value{
recorderWithin Yes or No response on whether the recorderValue is within the
uncertainty value known
}
\description{
This takes the uncertainty values, the reading values and the recorderValues
and returns a Yes or No repsonse if the reading value is within the uncertainty level. 
If the recorderValue is not available, it compares the reading value with the recorderValue
and if the values match, returns Yes, if not, No. The uncertainty range is inclusive.
}
