% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getReportMetadata}
\alias{getReportMetadata}
\title{get value from extremes json list}
\usage{
getReportMetadata(ts, param, ...)
}
\arguments{
\item{ts}{a list, can be the output of \code{\link[jsonlite]{fromJSON}}.}

\item{param}{the field name (e.g., 'locationNumber')}

\item{...}{additional arguments passed to \code{repgen:::validParam},
such as \code{required}, or \code{as.numeric}}
}
\value{
a value or array corresponding to the field specified by \code{param}
}
\description{
convienence function for accessing from the "values" block in
extremes json
}

