% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getSecondaryPlotConfig}
\alias{getSecondaryPlotConfig}
\title{Get Secondary Plot Config}
\usage{
getSecondaryPlotConfig(name, x, y, legend_label)
}
\arguments{
\item{name}{name of style to be applied to given x/y points (corrected, estimated, or uncorrected)}

\item{x}{the x/time values to put into the gsplot calls}

\item{y}{the y/time values to put into the gsplot calls}

\item{legend_label}{label to be applied to points in legend}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a report object, some information about the plot to build, will return a named list of gsplot elements to call
}

