% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getSecondarySeriesList}
\alias{getSecondarySeriesList}
\title{Depending on conditions, might be ref series or upchain series}
\usage{
getSecondarySeriesList(reportObject, month, timezone)
}
\description{
Depending on conditions, might be ref series or upchain series
}

