% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveyeargwsum-render.R
\name{getSides}
\alias{getSides}
\title{Sort data and sides FY}
\usage{
getSides(stat1TimeSeries, stat2TimeSeries, stat3TimeSeries, stat4TimeSeries,
  stat1TimeSeriesEst, stat2TimeSeriesEst, stat3TimeSeriesEst,
  stat4TimeSeriesEst)
}
\value{
named list with two items, sides and ylims. Each item has a named list with items for each timeseries. (EG: side for reference would be returnedObject$sides$reference)
}
\description{
Depending on the configuration, stat-derived time series will be plotted on different sides. This constructs ylims and what side each series should be on
}
