% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{getSimsUrl}
\alias{getSimsUrl}
\title{Inserts the SIMS URL (if it exists) into the base of the report}
\usage{
getSimsUrl(reportObject)
}
\arguments{
\item{reportObject}{coming in to create a plot which may have SIMS info}
}
\value{
the HTML link for SIMS URL
}
\description{
Takes the SIMS URL and formats it for including in the report as
  a link.
}

