% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getSrsPrecision}
\alias{getSrsPrecision}
\title{Sets a precision value for some known numbers rather
than having a hardcode precision number sprinkled out.}
\usage{
getSrsPrecision()
}
\value{
The precision identified in the function
}
\description{
Provides a precision value which can be used any time we want to 
control precision of values
}

