% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getSRSQualifiers}
\alias{getSRSQualifiers}
\title{Get the qualifiers for use in the main table}
\usage{
getSRSQualifiers(inQualifiers)
}
\arguments{
\item{inQualifiers}{The qualifiers coming in from the JSON data string}
}
\value{
Formatted string of qualifiers
}
\description{
takes the qualifiers and parses them to format for use in the table
}

