% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getSrsTableQualifiers}
\alias{getSrsTableQualifiers}
\title{Takes a list of data and extracts the qualifiers from it}
\usage{
getSrsTableQualifiers(table)
}
\arguments{
\item{table}{The table of data created for the SRS report}
}
\value{
Extracts the unique qualifiers and returns a list
}
\description{
when passed a list of data, extracts the qualifiers from it
}

