% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{getStatDerived}
\alias{getStatDerived}
\title{Extract Derived Statistics From a Time Series Data Structure}
\usage{
getStatDerived(data, chain_nm, legend_nm, estimated, rmZeroNeg)
}
\arguments{
\item{data}{A structure of time series data, as list of fields.}

\item{chain_nm}{A chain name.}

\item{legend_nm}{A legend name.}

\item{estimated}{Extract estimated values when TRUE; don't extract estimated
values otherwise.}

\item{rmZeroNeg}{Exclude zero-or-negative values when not NULL, NA, or the
empty string; otherwise, include zero-or-negative values.}
}
\description{
Extract Derived Statistics From a Time Series Data Structure
}

