% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{getSVPColumns}
\alias{getSVPColumns}
\title{Return a list of columns for the Site Visit Peak report}
\usage{
getSVPColumns(includeComments)
}
\arguments{
\item{includeComments}{boolean value about whether to include comments or not}
}
\value{
list of columns
}
\description{
Return a list of columns for the Site Visit Peak report
}
