% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Extract and Restructure a Time Series}
\usage{
getTimeSeries(ts, field, estimatedOnly = FALSE, shiftTimeToNoon = TRUE)
}
\arguments{
\item{ts}{A time series data structure, as R list.}

\item{field}{A field name.}

\item{estimatedOnly}{Extract estimated time series only when TRUE.}

\item{shiftTimeToNoon}{Reference time to 12:00 p.m. when TRUE.}
}
\description{
Extract and Restructure a Time Series
}

