% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-legend.R
\name{getTimeSeriesLabel}
\alias{getTimeSeriesLabel}
\title{Create Time Series Labels For the Legend}
\usage{
getTimeSeriesLabel(ts, field)
}
\arguments{
\item{ts}{The time series object which contains the time series information}

\item{field}{The type of time series data requested}
}
\value{
The type of time series requested as well as the units
}
\description{
Returns the desired time series label for the time series
requested formatted for inclusion in the legend
}

