% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{getUniqueComments}
\alias{getUniqueComments}
\title{Checks comment to see if it already printed the comment for the same date 
otherwise if the same prints empty char}
\usage{
getUniqueComments(comments, date, lastDate, lastComm)
}
\arguments{
\item{comments}{a single string of comments}

\item{date}{the date for the current comments param formatted mm/dd/yyyy}

\item{lastDate}{the last date for which it printed comments to compare}

\item{lastComm}{the last comment it printed to compare}
}
\value{
selectedComm which is the new comment to print or empty character
if the comments are the same as what it just printed
}
\description{
Takes the comments and compares the last comment and date to 
decide whether or not to print the comments
}

