% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-styles.R
\name{getUvStyles}
\alias{getUvStyles}
\title{Get UV Styles}
\usage{
getUvStyles()
}
\value{
list of named styling elements
}
\description{
Get styling and label information for UV hydrograph report elements
}
