% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getVerticalFlagArrows}
\alias{getVerticalFlagArrows}
\title{Get Vertical Flag Arrows
For a set of vertical flag labels, will return a list of arrows to connect label to vertical line.}
\usage{
getVerticalFlagArrows(flagLabels)
}
\arguments{
\item{flagLabels}{list of labels with positioning information already calculated/attached}
}
\value{
list of data describing how to draw lines to corresponding labels
}
\description{
Get Vertical Flag Arrows
For a set of vertical flag labels, will return a list of arrows to connect label to vertical line.
}
