% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getVerticalFlagPositions}
\alias{getVerticalFlagPositions}
\title{Get positions for vertical flags
Given a list of flags, will return a list of times (with duplicates removed) which are the x position of vertical lines}
\usage{
getVerticalFlagPositions(flags)
}
\arguments{
\item{flags}{a list of vertical flags}
}
\value{
list of time/x for each flag
}
\description{
Get positions for vertical flags
Given a list of flags, will return a list of times (with duplicates removed) which are the x position of vertical lines
}
