% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{getWaterDataUrl}
\alias{getWaterDataUrl}
\title{Put the waterdata.usgs.gov url (if it exists) into the base of the report}
\usage{
getWaterDataUrl(data)
}
\arguments{
\item{data}{coming in to create a plot which may have waterdata info}
}
\description{
Put the waterdata.usgs.gov url (if it exists) into the base of the report
}

