% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getWaterQualityMeasurements}
\alias{getWaterQualityMeasurements}
\title{Extract Water Quality measurements from a time series data structure.}
\usage{
getWaterQualityMeasurements(ts, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{...}{Optional arguments.}
}
\description{
Extract Water Quality measurements from a time series data structure.
}

