% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getWqPlotConfig}
\alias{getWqPlotConfig}
\title{Get Water Quality Plot Config}
\usage{
getWqPlotConfig(water_qual)
}
\arguments{
\item{water_qual}{list of water_qual objects}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a list of wq objects, will return a named list of gsplot elements to call
}

