% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{hasEnoughVdiagramData}
\alias{hasEnoughVdiagramData}
\title{Has Enough Vdiagram Data}
\usage{
hasEnoughVdiagramData(shifts)
}
\arguments{
\item{shifts}{the data in this report
return true or false if enough data exists to render a meaningful plot}
}
\description{
returns true if we have enough data to plot
}
