% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{hasReferenceSeries}
\alias{hasReferenceSeries}
\title{Has Refernce Series
Determines if the report has a reference series attached}
\usage{
hasReferenceSeries(reportObject)
}
\arguments{
\item{reportObject}{UV Hydro report object}
}
\value{
true/false
}
\description{
Has Refernce Series
Determines if the report has a reference series attached
}
