% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{hasUpchainSeries}
\alias{hasUpchainSeries}
\title{Has Upchain Series
Determines if the report has a upchain series attached}
\usage{
hasUpchainSeries(reportObject)
}
\arguments{
\item{reportObject}{UV Hydro report object}
}
\value{
true/false
}
\description{
Has Upchain Series
Determines if the report has a upchain series attached
}

