% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-render.R
\name{hasValidDataToPlot}
\alias{hasValidDataToPlot}
\title{Has Valid Data to Plot}
\usage{
hasValidDataToPlot(allLaneData)
}
\arguments{
\item{allLaneData}{A list containing all of the created data lanes}
}
\description{
Returns whether or not there is any lane data to plot
}

