% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{historyMeasurementsLabel}
\alias{historyMeasurementsLabel}
\title{History Measurements Label}
\usage{
historyMeasurementsLabel(reportObject)
}
\arguments{
\item{reportObject}{The V Diagram report}
}
\description{
A function that checks to see if there are any historic measurements and, if so, creates a label
that describes how long the historic data goes back.
}

