% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{is.repgendate}
\alias{is.repgendate}
\title{Setup special repgen date is method}
\usage{
is.repgendate(x)
}
\arguments{
\item{x}{a date vector}
}
\description{
For non-daily values, use a special class for the dates with 
times, so that they print correctly.
See print.repgendate and as.character.repgendate
}
