% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isEmpty}
\alias{isEmpty}
\title{Check for empty values}
\usage{
isEmpty(val)
}
\arguments{
\item{val}{any R object.}
}
\description{
Check if a value is NULL or NA.
}
\details{
Be careful what you pass in as \code{val},
e.g. a list will return a vector of logicals.
}

