% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isEmptyOrBlank}
\alias{isEmptyOrBlank}
\title{Test if a Value is the NULL Indicator, the NA Indicator, or the Empty String}
\usage{
isEmptyOrBlank(val = NULL, listObjects = NULL, objectName = NULL)
}
\arguments{
\item{val}{A value.}

\item{listObjects}{An R list.}

\item{objectName}{An object name.}
}
\description{
Test if a Value is the NULL Indicator, the NA Indicator, or the Empty String
}

