% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isEmptyVar}
\alias{isEmptyVar}
\title{Test Whether a Variable is Empty}
\usage{
isEmptyVar(variable)
}
\arguments{
\item{variable}{A variable's value.}
}
\description{
Test Whether a Variable is Empty
}

