% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{isFirstDayOfMonth}
\alias{isFirstDayOfMonth}
\title{Checks to see if the date passed in is the first of the month}
\usage{
isFirstDayOfMonth(date)
}
\arguments{
\item{date}{the date to check to see if it's the first of the month}
}
\value{
TRUE or FALSE depending on if the date is the first day of the month
}
\description{
For a date passed into the function, returns true or false 
depending on if that date is the first of the month
}
