% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{isLogged}
\alias{isLogged}
\title{Determines whether a time series should be plotted relative to a log10
vertical axis.}
\usage{
isLogged(all_data, ts_data, series)
}
\arguments{
\item{all_data}{A set of time series data, as list of fields.}

\item{ts_data}{A single time series, as ordinal sequence of (time,value)
pairs.}

\item{series}{The field name of the time series.}
}
\description{
Determines whether a time series should be plotted relative to a log10
vertical axis.
}

