% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{isPrimaryDischarge}
\alias{isPrimaryDischarge}
\title{Is Primary Discharge
Determines if the primary series in the report is Discharge parameter}
\usage{
isPrimaryDischarge(reportObject)
}
\arguments{
\item{reportObject}{UV Hydro report object}
}
\value{
true/false
}
\description{
Is Primary Discharge
Determines if the primary series in the report is Discharge parameter
}

