% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{isTextLong}
\alias{isTextLong}
\title{Check label length to make sure it will fit}
\usage{
isTextLong(labelText, dateLim = NULL, startD, endD, totalDays = NULL)
}
\arguments{
\item{labelText}{the label to apply to the CORR block}

\item{dateLim}{the start and end dates for the block section}

\item{startD}{a list of the start dates for the month sequence}

\item{endD}{a list of the end dates for the month sequence}

\item{totalDays}{the number of days for the section of the block}
}
\description{
Checks to see if the label to apply to the section block on the CORR report 
is too large for the space alotted
}

