% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{isTimeSeriesInverted}
\alias{isTimeSeriesInverted}
\title{Test Whether a Time Series is to be Plotted vs. a Reverse \emph{y}-axis}
\usage{
isTimeSeriesInverted(timeseries)
}
\arguments{
\item{timeseries}{A time series data structure.}
}
\description{
Test Whether a Time Series is to be Plotted vs. a Reverse \emph{y}-axis
}

