% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{labelDateSeq}
\alias{labelDateSeq}
\title{Provides the label for the corr report sections}
\usage{
labelDateSeq(startSeq, endSeq, numdays)
}
\arguments{
\item{startSeq}{a list of start dates for months}

\item{endSeq}{a list of end dates for months}

\item{numdays}{the full number of days requested in the report}
}
\value{
the list of months to label in the corr report
}
\description{
Given the start and end dates for the sections and the
total number of days, provides the list of month labels to aply to the report
removing those labels where there isn't enough room
}

