% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadCommonCssBase64}
\alias{loadCommonCssBase64}
\title{Load Common CSS}
\usage{
loadCommonCssBase64()
}
\value{
base64 string CSS
}
\description{
will load a CSS file that is common to all reports and base64 encode it
}
