% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadCommonTemplate}
\alias{loadCommonTemplate}
\title{Load Common Template}
\usage{
loadCommonTemplate(templateName)
}
\arguments{
\item{templateName}{name of tempalte not including extension (all templates must have the .mustache extension)}
}
\value{
string template
}
\description{
will load a template with the given name from the common set of templates
}
