% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadCustomCssBase64}
\alias{loadCustomCssBase64}
\title{Load Report CSS}
\usage{
loadCustomCssBase64(reportName)
}
\arguments{
\item{reportName}{name of report to load for}
}
\value{
base64 string CSS
}
\description{
will load the main CSS file and base64 encode it for the given report. Reports should have files provided in common file structure.
}
