% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadCustomJsBase64}
\alias{loadCustomJsBase64}
\title{Load Report JS}
\usage{
loadCustomJsBase64(reportName)
}
\arguments{
\item{reportName}{name of report to load for}
}
\value{
base64 string JS
}
\description{
will load the main JS file and base64 encode it for the given report. Reports should have files provided in common file structure.
}
