% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadImportsIntoListBase64}
\alias{loadImportsIntoListBase64}
\title{loadImportsIntoListBase64}
\usage{
loadImportsIntoListBase64(reportName, filename)
}
\arguments{
\item{reportName}{name of report to load imports for}

\item{filename}{the file to load with the report's directory structure}
}
\value{
a list of loaded imports in base64
}
\description{
will load a list of imports specified in the reports filename file. the file is a line separted list of file paths relative from the shared/lib folder that will be loaded.
}
