% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationchain-data.R
\name{loadLegendSymbologyDerivationChain}
\alias{loadLegendSymbologyDerivationChain}
\title{load legend symbology for derivation chain report}
\usage{
loadLegendSymbologyDerivationChain()
}
\value{
base64 string representation of legend images
}
\description{
will load the legend icons needed and convert to base64 text
}
