% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadPartialsListForReport}
\alias{loadPartialsListForReport}
\title{loadPartialsListForReport}
\usage{
loadPartialsListForReport(reportName)
}
\arguments{
\item{reportName}{name of report (folder) to load partials from}
}
\value{
named list of loaded templates
}
\description{
Will return a named list of loaded template strings for given report. The name wills be the template file name without the .mustache extensions
}
