% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadReportTemplate}
\alias{loadReportTemplate}
\title{Load Report Template}
\usage{
loadReportTemplate(reportName, templateName)
}
\arguments{
\item{reportName}{name of report to load for}

\item{templateName}{name of tempalte not including extension (all templates must have the .mustache extension)}
}
\value{
string template
}
\description{
will load a template with the given name from the report specified
}
