% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{loadUsgsLogoBase64}
\alias{loadUsgsLogoBase64}
\title{loadUsgsLogoBase64}
\usage{
loadUsgsLogoBase64()
}
\value{
base64 string representation of an image
}
\description{
will load the logo image and convert to base64 text
}
