% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{log_tick_marks}
\alias{log_tick_marks}
\title{Add tick marks to logarithmic axis}
\usage{
log_tick_marks(min, max)
}
\arguments{
\item{min}{ylim function returning minimum axis range}

\item{max}{ylim function returning maximum axis range}
}
\description{
Add tick marks to logarithmic axis
}

