% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{makeGapsInclusiveIfStatistic}
\alias{makeGapsInclusiveIfStatistic}
\title{Make Gaps Inclusive If Statistic}
\usage{
makeGapsInclusiveIfStatistic(reportData, timezone, gaps, isStatDerived)
}
\arguments{
\item{reportData}{The report data}

\item{timezone}{the time zone to parse data into}

\item{gaps}{the gaps data to look at}

\item{isStatDerived}{boolean if the ts is stat-derived}
}
\description{
Make gap date range inclusive if the TS Processor is a statistic
}
