% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{makeVDiagramRmd}
\alias{makeVDiagramRmd}
\title{Starting point; creates R Markdown and runs rendering.}
\usage{
makeVDiagramRmd(rmd_dir, data, wd)
}
\arguments{
\item{rmd_dir}{Path to R Markdown directory.}

\item{data}{Report data structure.}

\item{wd}{Path to working directory.}
}
\description{
Starting point; creates R Markdown and runs rendering.
}

