% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{mergeLists}
\alias{mergeLists}
\title{Merge two lists}
\usage{
mergeLists(list1, list2)
}
\arguments{
\item{list1}{The first list to merge}

\item{list2}{The second list to merge}
}
\description{
Merges the two lists passed in. The lists must have the exact same structure.
}
