% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{negValues}
\alias{negValues}
\title{Finds if the plot data has any zero values.}
\usage{
negValues(data, val_nm)
}
\arguments{
\item{data}{A time series data structure, as a sequence of ordinal
(time,value) pairs.}

\item{val_nm}{The field name of the "value" component, in the time series
data structure.}
}
\description{
Finds if the plot data has any zero values.
}

