% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{nullMask}
\alias{nullMask}
\title{if a value is null, returns the empty string instead of null,
and if a value is present, the value is returned}
\usage{
nullMask(val)
}
\arguments{
\item{val}{the value you want to check for null and mask}
}
\value{
either the original value or a null empty character object
}
\description{
makes sure that the slot in the data frame is not missing by
exchanging null values as empty character or the original value if not null
}

