% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/padTable.R
\name{padTable}
\alias{padTable}
\title{pad a table in mono text}
\usage{
padTable(data, align = "left", space = 16)
}
\arguments{
\item{data}{a data.frame with character data only.
If \code{row.names} are present, they will be injected
into the table output.}

\item{align}{a character vector of length 1 or of equal length to the
number of columns in \code{data}}

\item{space}{a numeric vector of length 1 or of equal length to the
number of columns in \code{data}. Used to layout the table. Is nchar
width of a single column}
}
\value{
a character output for the table
}
\description{
create a simple text table
}
\examples{
rwNames <- c(character(0),'MAXIMUM INSTANTANEOUS DISCHARGE AND CORRESPONDING Gage height',
'MINIMUM INSTANTANEOUS DISCHARGE AND CORRESPONDING Gage height')
df <- data.frame('DATE'=c("-----","06/01/2013", "01/29/2013"),
                'HH:MM:SS' = c("-----",'21:30:00','13:15:00'),
                '(CFS)' = c("-----","457000","26600"),
                '(CMS)' = c("-----",'13000', '753'),
                '(FT)' = c("-----",'33.14','0.93'),
                '(M)' = c("-----",'10.10', '0.28'),
                stringsAsFactors = FALSE, check.names = FALSE,
                row.names = rwNames)
cat(padTable(df))
}

