% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrApprovals}
\alias{parseCorrApprovals}
\title{Parse CORR Approvals}
\usage{
parseCorrApprovals(timeSeries, timezone, dateSeq)
}
\arguments{
\item{timeSeries}{The time series to get approvals from}

\item{timezone}{The timezone to parse data into}

\item{dateSeq}{The sequence of dates from the report start date to the report end date}
}
\description{
Reads and formats the approvals for use on the CORR report
}
