% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseCorrectionsAsTable}
\alias{parseCorrectionsAsTable}
\title{Corrections as table}
\usage{
parseCorrectionsAsTable(corrections)
}
\arguments{
\item{corrections}{list of corrections}
}
\value{
table structure with unique row entries for each correction type found
}
\description{
Given a list of corrections, will create a table structure with unique entries for all corrections that appear in the list.
}

