% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrectionsData}
\alias{parseCorrectionsData}
\title{Reads and formats data for the CORR Report}
\usage{
parseCorrectionsData(reportObject)
}
\arguments{
\item{reportObject}{A corrections at a glance report JSON string}
}
\value{
Returns all the data, formatted correctly for the CORR report
}
\description{
Given the entire JSON data object, reads and formats the data 
into the format needed to create the CORR report
}

