% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrFieldVisits}
\alias{parseCorrFieldVisits}
\title{Parse Corr Field Visits}
\usage{
parseCorrFieldVisits(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone of the report}
}
\description{
Returns the field visit data and formats it properly for plotting
}

