% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrGrades}
\alias{parseCorrGrades}
\title{Parse CORR Grades}
\usage{
parseCorrGrades(timeSeries, timezone)
}
\arguments{
\item{timeSeries}{The time series to get grades for}

\item{timezone}{The timezone to parse data into}
}
\description{
Retrieves and formats grades for the CORR report
}
