% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrNotes}
\alias{parseCorrNotes}
\title{Parse CORR Notes}
\usage{
parseCorrNotes(timeSeries, timezone)
}
\arguments{
\item{timeSeries}{The time series to get notes for}

\item{timezone}{The timezone to parse data into}
}
\description{
Retrieves and formats notes for the CORR report
}
