% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrProcessingCorrections}
\alias{parseCorrProcessingCorrections}
\title{Parse CORR Processing getCorrectionsLabels}
\usage{
parseCorrProcessingCorrections(reportObject, processOrder, timezone)
}
\description{
Parse CORR Processing getCorrectionsLabels
}

