% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrQualifiers}
\alias{parseCorrQualifiers}
\title{Parse CORR Qualifiers}
\usage{
parseCorrQualifiers(timeSeries, timezone)
}
\description{
Reads and formats the qualifiers for use on the CORR report
}

