% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrThresholds}
\alias{parseCorrThresholds}
\title{Parse CORR Thresholds}
\usage{
parseCorrThresholds(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
Retrieves and formats thresholds for the CORR report
}
