% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{parseCustomDataElementsForTemplate}
\alias{parseCustomDataElementsForTemplate}
\title{Peform pulls}
\usage{
parseCustomDataElementsForTemplate(reportData)
}
\arguments{
\item{reportData}{typed report json, the type should be the name of the report}
}
\value{
list of named data elements that should be available to your templates
}
\description{
Peform pulls
}
