% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivationchain-data.R
\name{parseCustomDataElementsForTemplateForDerivationChain}
\alias{parseCustomDataElementsForTemplateForDerivationChain}
\title{parseCustomDataElementsForTemplateForDerivationChain}
\usage{
parseCustomDataElementsForTemplateForDerivationChain(reportData)
}
\arguments{
\item{reportData}{}
}
\value{
list of data elements for template
}
\description{
Will return the derivations array as a json fragment
}
