% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseCustomDataElementsForTemplateForTimeSeriesSummary}
\alias{parseCustomDataElementsForTemplateForTimeSeriesSummary}
\title{parseCustomDataElementsForTemplateForTimeSeriesSummary}
\usage{
parseCustomDataElementsForTemplateForTimeSeriesSummary(reportData)
}
\arguments{
\item{reportData}{full report data structure}
}
\value{
list of data elements for template
}
\description{
Will return a list of tables for the renderer
}
