% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVApprovals}
\alias{parseDVApprovals}
\title{Parse DV Approvals}
\usage{
parseDVApprovals(timeSeries, timezone)
}
\arguments{
\item{timeSeries}{the time series to get approvals for}

\item{timezone}{the timezone to parse the approval times into}
}
\description{
Given a time series read and format the approvals
data from it into apprpoval bars to put on the plot.
}

