% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVData}
\alias{parseDVData}
\title{Parse DV Data}
\usage{
parseDVData(reportObject)
}
\arguments{
\item{reportObject}{The raw report JSON object}
}
\description{
Parses the DV Hydrograph Data from the report
JSON and then furhter processes and formats that data for
plotting. Reutrns a list with 2 parts:
1). dvData - data that is ready for plotting
2). dvInfo - supplemental data that is used in plotting logic
but should not be plotted.
}

