% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVFieldVisitMeasurements}
\alias{parseDVFieldVisitMeasurements}
\title{Parse DV Field Visit Measurements}
\usage{
parseDVFieldVisitMeasurements(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object, reads the field
visit measurements and handles read errors.
}

