% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVGroundWaterLevels}
\alias{parseDVGroundWaterLevels}
\title{Parse DV Ground Water Levels}
\usage{
parseDVGroundWaterLevels(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the ground
water levels and handles read errors.
}

