% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVTimeSeries}
\alias{parseDVTimeSeries}
\title{Parse DV Time Series}
\usage{
parseDVTimeSeries(reportObject, seriesField, descriptionField, timezone,
  excludeZeroNegativeFlag, estimated = FALSE)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{seriesField}{the JSON field name for the TS data}

\item{descriptionField}{the JSON field name for the TS legend name}

\item{timezone}{The timezone to parse the TS points into}

\item{excludeZeroNegativeFlag}{whether or not to exclude zero and negative values}
}
\description{
Given a full report object and series name,
reads and formats the estimated and non-estaimted time
series for plotting.
}

