% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseExcludedControlConditions}
\alias{parseExcludedControlConditions}
\title{Parse Excluded Control Conditions (VDI)}
\usage{
parseExcludedControlConditions(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Default wrapper for the readExcludedControlConditions function
that handles errors thrown and returns the proper data.
}
