% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseExtremesSeriesQualifiers}
\alias{parseExtremesSeriesQualifiers}
\title{Parse Extremes Primary Series Qualifiers}
\usage{
parseExtremesSeriesQualifiers(reportObject, qualType)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{qualType}{the type of qualifier we're looking to read}
}
\description{
Default wrapper for the readPrimarySeriesQualifiers function
that handles errors thrown and returns the proper data.
}
