% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{parseFieldMeasurementData}
\alias{parseFieldMeasurementData}
\title{Parse Field Measurement Data}
\usage{
parseFieldMeasurementData(reportObject)
}
\arguments{
\item{reportObject}{An R object with the raw data required for a V Diagram}
}
\value{
A list containing the field visit measurement data
}
\description{
Takes in a V diagram report and returns the field visit measurement data
}

