% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseFieldVisitReadings}
\alias{parseFieldVisitReadings}
\title{Parse Field Visit Readings}
\usage{
parseFieldVisitReadings(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object, reads the field
visit readings and handles read errors.
}
