% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseGaps}
\alias{parseGaps}
\title{Parse Gaps}
\usage{
parseGaps(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
Default wrapper for the readGaps function
that handles errors thrown and returns the proper data
}
