% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseGroundWaterLevels}
\alias{parseGroundWaterLevels}
\title{Parse Ground Water Levels}
\usage{
parseGroundWaterLevels(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the ground
water levels and handles read errors.
}
