% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseMinMaxIV}
\alias{parseMinMaxIV}
\title{Parse the Min or Max IV Data}
\usage{
parseMinMaxIV(reportObject, stat, timezone, tsType, inverted)
}
\arguments{
\item{reportObject}{the full report data}

\item{stat}{the stat to look up (MAX or MIN)}

\item{timezone}{the timezone to parse the times into}

\item{tsType}{the type of the TS (to use for the legend name)}

\item{inverted}{whether or not the TS is inverted}
}
\value{
the first min or max IV data point from the list of min max IVs
}
\description{
Reads the min or max IV Data from the reportObject then takes the
first entry and formats it properly for use on the DV Hydrograph report.
}

