% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parsePrimaryDvList}
\alias{parsePrimaryDvList}
\title{Parse primary dv lists}
\usage{
parsePrimaryDvList(reportObject, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
subset list of DV points. Each point is named with or "approved_dv", "analyzed_dv", "working_dv"
}
\description{
Read daily values subsetted by month and named by approval level for UV Hydrograph
}
