% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parsePrimarySeriesApprovals}
\alias{parsePrimarySeriesApprovals}
\title{Parse Primary Series Approvals (Five YR)}
\usage{
parsePrimarySeriesApprovals(reportObject, startDate, endDate)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{startDate}{the start date of the report}

\item{endDate}{the end date of the report}
}
\description{
Default wrapper for the readPrimarySeriesApprovals function
that handles errors thrown and returns the proper data.
}
