% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parsePrimarySeriesList}
\alias{parsePrimarySeriesList}
\title{Parse Secondary Series List}
\usage{
parsePrimarySeriesList(reportObject, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
named list of series to be included on secondary plot
}
\description{
Depending on conditions, might be ref series or upchain series
}

