% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parsePrimarySeriesQualifiers}
\alias{parsePrimarySeriesQualifiers}
\title{Parse Primary Series Qualifiers (DV & Five YR)}
\usage{
parsePrimarySeriesQualifiers(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Default wrapper for the readPrimarySeriesQualifiers function
that handles errors thrown and returns the proper data.
}
