% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parsePrimarySeriesQualifiersDV}
\alias{parsePrimarySeriesQualifiersDV}
\title{Parse Primary Series Qualifiers (DV)}
\usage{
parsePrimarySeriesQualifiersDV(reportObject, filterCode = NULL)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{filterCode}{The code to filter read qualifiers to}
}
\description{
Default wrapper for the readPrimarySeriesQualifiers function
that handles errors thrown and returns the proper data.
}
