% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseRatingShiftsAsTable}
\alias{parseRatingShiftsAsTable}
\title{Rating Shifts as a table}
\usage{
parseRatingShiftsAsTable(ratingShifts)
}
\arguments{
\item{ratingShifts}{list of ratingShifts}
}
\value{
table structure with unique row entries for each rating shift found
}
\description{
Given a list of rating shifts, will create a table with unique entries for all rating shifts that appear in the list
}
