% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseRatingShiftsByMonth}
\alias{parseRatingShiftsByMonth}
\title{Parse Rating Shifts}
\usage{
parseRatingShiftsByMonth(reportObject, fieldName, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{fieldName}{the field name of the rating shifts to pull from report}

\item{month}{subset only into this month}
}
\value{
rating shifts subset by month
}
\description{
Read rating shifts for a given series
}
