% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{parseRatingShiftsData}
\alias{parseRatingShiftsData}
\title{Parse Rating Shifts Data}
\usage{
parseRatingShiftsData(reportObject)
}
\arguments{
\item{reportObject}{An R object with the raw data required for a V Diagram}
}
\value{
A list containing rating shift information
}
\description{
Takes in a V diagram report and returns the rating shift information
}
