% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseRefData}
\alias{parseRefData}
\title{Parse Reference Data}
\usage{
parseRefData(reportObject, series, description)
}
\arguments{
\item{reportObject}{The raw report JSON object}

\item{series}{The series to fetch (secondary, tertiary, or quaternary)}

\item{description}{The name of the descriptionField to fetch}
}
\description{
Parses the specified refernec series data from 
the report JSON and then furhter processes and formats that 
data for plotting.
}

