% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseSecondaryCorrectionsByMonth}
\alias{parseSecondaryCorrectionsByMonth}
\title{Parse Secondary Corrections}
\usage{
parseSecondaryCorrectionsByMonth(reportObject, month)
}
\arguments{
\item{reportObject}{the report to render}
}
\value{
corrections list from the correct series
}
\description{
depending on the report configuration, corrections might come from a reference or upchain series
}
