% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{parseSRSQualifiers}
\alias{parseSRSQualifiers}
\title{Parse SRS Qualifiers}
\usage{
parseSRSQualifiers(reportData, timezone, qualifierMetadata)
}
\arguments{
\item{reportData}{the readings JSON object}

\item{timezone}{the timezone of the report}

\item{qualifierMetadata}{the code and display name of qualifiers 
in the report to be joined with the readings qualifier identifier}
}
\description{
Given a readings JSON object reads the
qualifiers and handles read errors.
}
