% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseThresholds}
\alias{parseThresholds}
\title{Parse Thresholds}
\usage{
parseThresholds(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
Default wrapper for the readThresholds function
that handles errors thrown and returns the proper data
}
