% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseTimeSeries}
\alias{parseTimeSeries}
\title{Parse Time Series}
\usage{
parseTimeSeries(reportObject, seriesField, descriptionField, timezone,
  estimated = FALSE, isDV = FALSE, requiredFields = NULL)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{seriesField}{the JSON field name for the TS data}

\item{descriptionField}{the JSON field name for the TS legend name}

\item{timezone}{The timezone to parse the TS points into}

\item{estimated}{whether or not the retrieved time series should be estimated or non-estimated}

\item{isDV}{true to treat the series as a DV series (and parse dates accordingly), defaults to FALSE}

\item{requiredFields}{An optional list of names of required JSON fields to overwrite the default}
}
\value{
The requested time series or NULL if the request time series was not found.
}
\description{
Default wrapper for the readTimeSeries functions that handles
errors thrown by those functions if the specified time series is
not found and throws a warning message. Also handles time series that
are returned without any point data and treats them as NULL.
}
