% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSGrades}
\alias{parseTSSGrades}
\title{Parse TSS Grades}
\usage{
parseTSSGrades(reportData, timezone)
}
\arguments{
\item{timezone}{the timezone of the report}

\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the
grades and handles read errors.
}
