% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSMethods}
\alias{parseTSSMethods}
\title{Parse TSS Methods}
\usage{
parseTSSMethods(reportData, timezone)
}
\arguments{
\item{reportData}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
TSS wrapper for the readMethods function
that handles errors thrown and returns the proper data
}
