% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSNotes}
\alias{parseTSSNotes}
\title{Parse TSS Notes}
\usage{
parseTSSNotes(reportData, timezone)
}
\arguments{
\item{timezone}{the timezone of the report}

\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the
notes and handles read errors.
}
