% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSPrimaryTsMetadata}
\alias{parseTSSPrimaryTsMetadata}
\title{Parse TSS Primary TS Metadata}
\usage{
parseTSSPrimaryTsMetadata(reportData)
}
\arguments{
\item{reportData}{The full report JSON object}
}
\description{
TSS wrapper for the readPrimaryTSMetadata function
that handles errors thrown and returns the proper data
}
