% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSProcessors}
\alias{parseTSSProcessors}
\title{Parse TSS Processors}
\usage{
parseTSSProcessors(reportData, timezone)
}
\arguments{
\item{reportData}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
TSS wrapper for the readProcessors function
that handles errors thrown and returns the proper data
}
