% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSQualifiers}
\alias{parseTSSQualifiers}
\title{Parse TSS Qualifiers}
\usage{
parseTSSQualifiers(reportData, timezone)
}
\arguments{
\item{timezone}{the timezone of the report}

\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the
ratings curves and handles read errors.
}
