% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSRelatedSeries}
\alias{parseTSSRelatedSeries}
\title{Parse TSS Related Series}
\usage{
parseTSSRelatedSeries(reportData)
}
\arguments{
\item{reportData}{the full report JSON object}
}
\description{
Given the full report JSON object reads the
related series and handles read errors.
}
