% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSThresholds}
\alias{parseTSSThresholds}
\title{Parse TSS Thresholds}
\usage{
parseTSSThresholds(reportData, timezone)
}
\arguments{
\item{timezone}{the timezone of the report}

\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the
thresholds and handles read errors.
}
