% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseUvComparisonSeriesByMonth}
\alias{parseUvComparisonSeriesByMonth}
\title{Parse UV Comparison Series}
\usage{
parseUvComparisonSeriesByMonth(reportObject, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
comparison series points subset by month
}
\description{
Read entire comparison series
}
