% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseUvEstimatedSeries}
\alias{parseUvEstimatedSeries}
\title{Parse UV Estimated Series}
\usage{
parseUvEstimatedSeries(reportObject, seriesName, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{seriesName}{series to read}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
series points subste by month
}
\description{
Read estimated portion of a series
}

