% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseUvNonEstimatedSeries}
\alias{parseUvNonEstimatedSeries}
\title{Parse UV Non-Estimated Series}
\usage{
parseUvNonEstimatedSeries(reportObject, seriesName, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{seriesName}{series to read}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
series points subset by month
}
\description{
Read non-estimated portion of a series
}

