% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseUvTimeInformationFromLims}
\alias{parseUvTimeInformationFromLims}
\title{Get UV time information from lims
Returns time information for the lims}
\usage{
parseUvTimeInformationFromLims(lims, timezone)
}
\arguments{
\item{lims}{the lims to describe}

\item{timezone}{the timezone to convert all data to}
}
\value{
named list of metadata about the lims (dates, days, start, end)
}
\description{
Get UV time information from lims
Returns time information for the lims
}
