% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{parseVDiagramData}
\alias{parseVDiagramData}
\title{Parse Vdiagram Data}
\usage{
parseVDiagramData(reportObject)
}
\arguments{
\item{reportObject}{An R object with the raw data required for a V Diagram}
}
\value{
A list containing the relevant data for a V Diagram
}
\description{
Takes in a V diagram report and returns the necessary data for a V Diagram
}

