% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseVerticalFlagLabelSpacing}
\alias{parseVerticalFlagLabelSpacing}
\title{Parse vertical flag label spacing}
\usage{
parseVerticalFlagLabelSpacing(flags, limits)
}
\arguments{
\item{flags}{list of vertical flags (time/comment pairs)}

\item{limits}{the lims that the vertical flag labels should not leave}
}
\value{
list of named items (x, xorigin, y, r, label) which desribes where/how to to place each label
}
\description{
each vertical flag is a time/comment pair. This will deterimine how to place labels so they do not overlap each other.
}
