% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseWaterQualityMeasurements}
\alias{parseWaterQualityMeasurements}
\title{Parse Water Quality Measurements}
\usage{
parseWaterQualityMeasurements(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Given the full report JSON object reads the water
quality measurements and handles read errors.
}
