% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{plotDVHydroLegend}
\alias{plotDVHydroLegend}
\title{Plot DV Hydro Legend}
\usage{
plotDVHydroLegend(plot_object, startDate, endDate, timezone, initialOffset,
  modOffset = 1)
}
\arguments{
\item{plot_object}{The gsplot object to add the legend to}

\item{startDate}{The start date of the report}

\item{endDate}{The end date of the report}

\item{timezone}{The timezone of the report}

\item{initialOffset}{The initial amount to offset the legend by}

\item{modOffset}{[Default: 1] An optional amount to multiply the final calculated offset by}
}
\description{
Given the plot object and additional necessary parameters, calculates the
legend offset and then adds the legend to the plot.
}

