% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plotItem}
\alias{plotItem}
\title{Plot Item}
\usage{
plotItem(plot_object, item, configFunction, configFunctionParams,
  isDV = FALSE)
}
\arguments{
\item{plot_object}{The gsplot object to plot onto}

\item{item}{The item to plot}

\item{configFunction}{The function to use for fetching the style and config data for this TS}

\item{configFunctionParams}{list of params to call pall configFunctionWith}

\item{isDV}{Whether or not the plot is a daily value plot (defulat: FALSE)}
}
\description{
Function that takes a gsplot object and the necessary item data and
then formats the item properly for plotting and plots it.
}
