% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-render.R
\name{plotLanes}
\alias{plotLanes}
\title{Plot Lanes}
\usage{
plotLanes(gsplotObject, laneData, laneName, dateRange, rectHeight)
}
\arguments{
\item{gsplotObject}{The gsplot object to plot the lane onto}

\item{laneData}{The lane data to plot}

\item{laneName}{The variable name of the lane being plotted}

\item{dateRange}{The date range of the report}

\item{rectHeight}{The height to use for rendering the lane rectangles}
}
\value{
The gsplot object with the lane data plotted onto it
}
\description{
Given the gsplot object this plots the specified lane data using the
provided additional parameters.
}

