% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot Time Series}
\usage{
plotTimeSeries(plot_object, ts, name, timezone, configFunction,
  configFunctionAdditionalParams, excludeZeroNegativeFlag = FALSE,
  isDV = FALSE)
}
\arguments{
\item{plot_object}{The gsplot object to plot onto}

\item{ts}{The Time Series to plot}

\item{name}{The variable name to use for the time series (used for style and config matching)}

\item{timezone}{The timezone of the time series (used for calculating gaps)}

\item{configFunction}{The function to use for fetching the style and config data for this TS}

\item{configFunctionAdditionalParams}{additional params to pas into configFunction, config function will be called with (subsetOfSeries, name, and configFunctionAdditionalParams added at end)}

\item{excludeZeroNegativeFlag}{Whether or not to remove zero and negative values from the ts}

\item{isDV}{Whether or not the plot is a daily value plot (default: FALSE)}
}
\description{
Function that takes a gsplot object and the necessary time series data and
formats the time series properly for plotting and then plots it.
}
