% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\docType{methods}
\name{+,repgendate,Period-method}
\alias{+,repgendate,Period-method}
\title{Addition method for class repgendate}
\usage{
\S4method{+}{repgendate,Period}(e1, e2)
}
\arguments{
\item{e1}{object of class "repgendate" (from repgen)}

\item{e2}{object of class "Period" (from lubridate)}
}
\description{
Having this special class limits what arithmetic can be done thanks
to lubridate's "Period" class (e.g. time + hours(4)) because adding
the special class to "Period" won't work unless you define an 
addition method for `+`. This is the addition method for adding Period
and repgendate classes, but other arithmetic operations will need to 
do the same.
}
