% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{print.repgendate}
\alias{print.repgendate}
\title{Setup special repgen date print method}
\usage{
\method{print}{repgendate}(x, ...)
}
\arguments{
\item{x}{a date vector with class "repgendate"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Need this so that HH:SS are not dropped when there is only one
value in a vector and it has 00:00 as the time. print.POSIXct
drops the time for those cases. A class of repgendate is added
in flexibleTimeParse.
}
