% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{printReportFeature}
\alias{printReportFeature}
\title{Used by UV Hydrograph report to organize and fit plots and tables in report}
\usage{
printReportFeature(feature, isTable = FALSE, m = NULL, mar_values = c(8,
  3, 4, 3))
}
\arguments{
\item{feature}{the plot or table to include}

\item{isTable}{TRUE or FALSE if the feature is a table. defaults to FALSE}

\item{m}{the month name for the feature}

\item{mar_values}{the margin values for the feature to ensure space for plot and legend}
}
\value{
the formatted feature requested
}
\description{
sets margins and lays out report in an attractive way for UV Hydrograph
}

