% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{printWithFourthYAxis}
\alias{printWithFourthYAxis}
\title{printWithFourthYAxis}
\usage{
printWithFourthYAxis(plot)
}
\arguments{
\item{plot}{A gsplot, plot object.}
}
\description{
Plot a time series on a fourth axis
}
\seealso{
\code{\link{log_tick_marks}}
}
