% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{printWithMultipleYAxes}
\alias{printWithMultipleYAxes}
\title{printWithMultipleYAxes}
\usage{
printWithMultipleYAxes(plot)
}
\arguments{
\item{plot}{A gsplot, plot object.}
}
\description{
Plot time series on additional y-axes
}
\seealso{
\code{\link{log_tick_marks}}
}
