% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{printWithThirdYAxis}
\alias{printWithThirdYAxis}
\title{printWithThirdYAxis}
\usage{
printWithThirdYAxis(plot)
}
\arguments{
\item{plot}{A gsplot, plot object.}
}
\description{
Plot a time series on a third axis
}
\seealso{
\code{\link{log_tick_marks}}
}
