% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rating.R
\docType{methods}
\name{rating}
\alias{rating}
\alias{rating,character,character-method}
\alias{rating,list,character-method}
\alias{rating,list,missing-method}
\title{rating-diagram report}
\usage{
rating(data, output, ...)

\S4method{rating}{list,character}(data, output)

\S4method{rating}{character,character}(data, output, ...)

\S4method{rating}{list,missing}(data, output, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{additional params passed to GET or authenticateUser}
}
\description{
rating-diagram report
}
\examples{
library(jsonlite)
data <- fromJSON(system.file('extdata','rating.json', package = 'repgen'))
#rating(data)
#rating(data, 'html')
}

