% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readAllFieldVisitQualifiers}
\alias{readAllFieldVisitQualifiers}
\title{Read all qualifiers from field visit readings}
\usage{
readAllFieldVisitQualifiers(visitReadings)
}
\arguments{
\item{visitReadings}{the object representing the parsed field visit readings}
}
\description{
Given a full report object of parsed field visit readings, 
returns all deduplicated qualifiers from the field visit readings formatted 
as a data frame
}
