% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readAllUvReadings}
\alias{readAllUvReadings}
\title{Read Readings}
\usage{
readAllUvReadings(reportObject, month, jsonFieldName)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}

\item{jsonFieldName}{report field to read readings from}
}
\value{
named list of readings by type and subsetted by month
}
\description{
Read readings subsetted by month and separated by type for UV Hydrograph
}
