% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalBar}
\alias{readApprovalBar}
\title{Read Approval Bars}
\usage{
readApprovalBar(ts, timezone, legend_nm, snapToDayBoundaries = FALSE)
}
\arguments{
\item{ts}{the timeseries to get approval bars for}

\item{timezone}{the timezone to convert all times to}

\item{legend_nm}{the name to be assigned to the legend entries (as a suffix)}

\item{snapToDayBoundaries}{true to shift all bar edges to the closest end/beginning of the days}

\item{returns}{a list of approval bar ranges, lists should contain the possible named items appr_working_uv, appr_inreview_uv, appr_approved_uv}
}
\description{
for a timeseries, will return a list of approval bars to be plotted
}

