% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalBarUv}
\alias{readApprovalBarUv}
\title{Read Approval Bars Uv}
\usage{
readApprovalBarUv(ts, timezone, legend_nm, snapToDayBoundaries = FALSE)
}
\arguments{
\item{ts}{the timeseries to get approval bars for, *ts must be parsed by readTimeseries*}

\item{timezone}{the timezone to convert all times to}

\item{legend_nm}{the name to be assigned to the legend entries (as a suffix)}

\item{snapToDayBoundaries}{true to shift all bar edges to the closest end/beginning of the days}
}
\value{
list of approval bar ranges, lists should contain the possible named items appr_working_uv, appr_analyzed_uv, appr_approved_uv
}
\description{
for a timeseries, will return a list of approval bars to be plotted
}
