% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalIndex}
\alias{readApprovalIndex}
\title{Read Approval index}
\usage{
readApprovalIndex(points, approvals, approvalLevel, timezone)
}
\arguments{
\item{points}{the points to apply approvals against}

\item{approvals}{list of approvals to read from}

\item{approvalLevel}{the approval level to read and apply to points}

\item{timezone}{the timezone to convert all datetimes to (in approvals)}
}
\description{
Given a list of points, a set of approvals, and the approvalLevel to apply, will return the indexes of all points to be assigned the approval level
}
