% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalPoints}
\alias{readApprovalPoints}
\title{Read Approval Points}
\usage{
readApprovalPoints(approvals, points, timezone, legend_nm, appr_var_all,
  point_type = NULL)
}
\arguments{
\item{approvals}{list of approvals}

\item{points}{list of points to apply approvals to}

\item{timezone}{the timezone to convert everything to}

\item{legend_nm}{the name of the series to put in label (suffix)}

\item{appr_var_all}{the ordered variable names to map the approval levels (Approved, In Review, Working) to (Eg: c("appr_approved_dv", "appr_inreview_dv", "appr_working_dv") )}

\item{point_type}{the symbol to attach to each point}
}
\value{
named list of data frames. Each frame wil be named according to appr_var_all and data frame will have respective points
}
\description{
given a list of approvals and points, will return the points divided up into separate lists for the different approval levels
}
