% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalRanges}
\alias{readApprovalRanges}
\title{Read Approval Ranges}
\usage{
readApprovalRanges(approvals, approvalLevel, timezone)
}
\arguments{
\item{approvals}{the approvals list object to read from}

\item{approvalLevel}{the approval level to read, typically "Working", "Analyzed", or "Approved"}

\item{timezone}{the timezone to parse times to}
}
\value{
data frame of start and end times for each approval range
}
\description{
Read Approval Ranges
}
