% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readCorrections}
\alias{readCorrections}
\title{Read corrections}
\usage{
readCorrections(reportObject, seriesCorrName)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\description{
Given a full report object and the name of a time series,
returns the corrections list for that time series
}

