% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readEstimatedTimeSeries}
\alias{readEstimatedTimeSeries}
\title{Read an estaimted time series}
\usage{
readEstimatedTimeSeries(reportObject, timezone, seriesName,
  shiftTimeToNoon = FALSE)
}
\arguments{
\item{reportObject}{the full JSON report object}

\item{timezone}{the timezone to parse times to}

\item{seriesName}{the name of the time series to extract}

\item{shiftTimeToNoon}{[DEFAULT: FALSE] whether or not to shift DV times to noon}
}
\description{
Reads and formats a time series from the provided full report object
}

