% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readExcludedControlConditions}
\alias{readExcludedControlConditions}
\title{Read Excluded Control Conditions (VDI)}
\usage{
readExcludedControlConditions(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Reads and formats the excluded control condition data
}
