% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readExtremesSeriesQualifiers}
\alias{readExtremesSeriesQualifiers}
\title{Read Extremes Series Qualifiers}
\usage{
readExtremesSeriesQualifiers(reportObject, qualType)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{qualType}{the type of qualifier we're looking to read}
}
\description{
Reads and formats the extremes qualifiers.
}
