% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readFetchedQualifiers}
\alias{readFetchedQualifiers}
\title{Read field visit reading qualifiers}
\usage{
readFetchedQualifiers(reportObject, inQualifiers, time = NULL)
}
\arguments{
\item{inQualifiers}{list of associated Instantaneous Value qualifiers}

\item{time}{associated Instantaneous Value date and time (optional, defaults to NULL)}
}
\description{
Given an associated Instantaneous Value date and time and qualifiers, 
returns the qualifiers formatted as a data frame
}
