% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readFieldVisitMeasurementsQPoints}
\alias{readFieldVisitMeasurementsQPoints}
\title{Read field visit measurements}
\usage{
readFieldVisitMeasurementsQPoints(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\value{
data frame
}
\description{
Given a full report object, returns the field visit 
measurement discharge points formatted as a time series point set
}
