% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readFieldVisitMeasurementsShifts}
\alias{readFieldVisitMeasurementsShifts}
\title{Read field visit measurements shifts}
\usage{
readFieldVisitMeasurementsShifts(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\description{
Given a full report object, returns the field visit
measurement shifts data formatted as a time series point set
}

