% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readFieldVisitReadings}
\alias{readFieldVisitReadings}
\title{Read field visit readings}
\usage{
readFieldVisitReadings(reportObject, commentFlag)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\description{
Given a full report object, returns the field visit 
readings formatted as a data frame
}

