% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readFieldVists}
\alias{readFieldVists}
\title{Read Field Vists (CORR)}
\usage{
readFieldVists(reportObject, timezone)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{timezone}{the timezone}
}
\description{
Reads and formats the field Vists
}

