% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readGroundWaterLevels}
\alias{readGroundWaterLevels}
\title{Read ground water levels}
\usage{
readGroundWaterLevels(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\description{
Given a full report object, returns the ground water levels
measurements formatted as a time series point set.
}
