% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readIvDifference}
\alias{readIvDifference}
\title{Calculate the difference between the field visit measurement value and the associated
Instantaneous Value}
\usage{
readIvDifference(readingVal, ivVal)
}
\arguments{
\item{readingVal}{field visit measurement value}

\item{ivVal}{associated Instantaneous Value}
}
\description{
Given a field visit measurement value and an associated Instantaneous Value date,
returns the difference.
}

