% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readMethods}
\alias{readMethods}
\title{Read Methods (TSS)}
\usage{
readMethods(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone of the report}
}
\description{
Reads and formats the primary TS methods
}
