% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readMinMaxIVsDV}
\alias{readMinMaxIVsDV}
\title{Read Min/Max IV Data DVHydro}
\usage{
readMinMaxIVsDV(reportObject, stat, timezone, inverted)
}
\arguments{
\item{reportObject}{the full JSON report object}

\item{stat}{the stat to pull (MAX or MIN)}

\item{timezone}{the timezone to parse times into}

\item{inverted}{whether or not the time series is inverted}
}
\description{
Reads and formats Min/Max IV Data from the provided full report object
}
