% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readNonEstimatedTimeSeries}
\alias{readNonEstimatedTimeSeries}
\title{Read a non-estaimted time series}
\usage{
readNonEstimatedTimeSeries(reportObject, seriesName, timezone,
  descriptionField = NULL, shiftTimeToNoon = FALSE, isDV = FALSE,
  requiredFields = NULL)
}
\arguments{
\item{reportObject}{the full JSON report object}

\item{seriesName}{the name of the time series to extract}

\item{timezone}{the timezone to parse times to}

\item{shiftTimeToNoon}{[DEFAULT: FALSE] whether or not to shift DV times to noon}

\item{requiredFields}{optional overriding of required fields for a time series}
}
\description{
Reads and formats a time series from the provided full report object
}

