% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readNonEstimatedTimeSeriesUv}
\alias{readNonEstimatedTimeSeriesUv}
\title{Read a non-estimated time series Uv}
\usage{
readNonEstimatedTimeSeriesUv(reportObject, seriesName, timezone,
  descriptionField = NULL, shiftTimeToNoon = FALSE, isDV = FALSE,
  requiredFields = NULL, onlyMonth = NULL)
}
\arguments{
\item{reportObject}{the full JSON report object}

\item{seriesName}{the name of the time series to extract}

\item{timezone}{the timezone to parse times to}

\item{descriptionField}{The JSON field name to fetch description inofmration from}

\item{shiftTimeToNoon}{[DEFAULT: FALSE] whether or not to shift DV times to noon}

\item{isDV}{whether or not the specified time series is a daily value time series}

\item{requiredFields}{optional overriding of required fields for a time series}

\item{onlyMonth}{4 character month code to limit points to (EG: "1608" only includes August 2016 points)}
}
\value{
ts with only points which are not in the estimated range
}
\description{
Reads and formats a time series from the provided full report object
}
