% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readPrimarySeriesApprovals}
\alias{readPrimarySeriesApprovals}
\title{Read Primary Series Approvals (DV and Five YR)}
\usage{
readPrimarySeriesApprovals(reportObject, startTime, endTime)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{startTime}{the start time of the report}

\item{endTime}{the end time of the report}
}
\description{
Reads and formats the primarySeriesApprovals as a time series
with no points and only approvals. Used to have DV Hydro and Five YR GW
base their approval bars off of the primary (upchain) series approvals instead
of the stat derived approvals.
}

