% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readPrimarySeriesQualifiersDV}
\alias{readPrimarySeriesQualifiersDV}
\title{Read Primary Series Qualifiers (DV)}
\usage{
readPrimarySeriesQualifiersDV(reportObject, filterCode = NULL)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{filterCode}{The qualifier code to filter read qualifiers to}
}
\description{
Reads and formats the primarySeriesQualifiers. Used to
allow DV Hydro to format their max/min UV colors.
}
