% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readPrimaryTSMetadata}
\alias{readPrimaryTSMetadata}
\title{Read Primary TS Metadata (TSS)}
\usage{
readPrimaryTSMetadata(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Reads and validates the primary TS metadata
}
