% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readPrimaryUvHydroApprovalBars}
\alias{readPrimaryUvHydroApprovalBars}
\title{Read Primary Uv Hydro Approval Bars}
\usage{
readPrimaryUvHydroApprovalBars(reportObject, timezone, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{timezone}{timezone to parse all data into}

\item{month}{filter data to this month}
}
\value{
approval bar plotting info for primary series
}
\description{
will read the relevant approval bar from primary series
}
