% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readProcessingCorrections}
\alias{readProcessingCorrections}
\title{Read Processing Corrections (CORR)}
\usage{
readProcessingCorrections(reportObject, processOrder, timezone)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{processOrder}{The processing order to get corrections for. Valid choices: "pre", "post", and "normal"}

\item{timezone}{target timezone to parse data into}
}
\description{
Reads and formats the corrections data for
the specified processing order.
}
