% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readQualifiers}
\alias{readQualifiers}
\title{Read field visit readings qualifiers}
\usage{
readQualifiers(time, inQualifiers)
}
\arguments{
\item{time}{associated Instantaneous Value date and time}

\item{inQualifiers}{list of associated Instantaneous Value qualifiers}
}
\description{
Given an associated Instantaneous Value date and time and qualifiers, 
returns the qualifiers formatted as a data frame
}

