% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readRatingCurves}
\alias{readRatingCurves}
\title{Read Rating Curves (TSS)}
\usage{
readRatingCurves(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone of the report}
}
\description{
Reads and formats the rating curves
}
