% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readRatingShifts}
\alias{readRatingShifts}
\title{Read Rating Shifts}
\usage{
readRatingShifts(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\value{
data frame of rating shift information
}
\description{
given a list of rating shifts returns shift data
}
