% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readReadings}
\alias{readReadings}
\title{Read Readings}
\usage{
readReadings(reportObject, filter = "")
}
\arguments{
\item{filter}{optional filter to restrict to reading types (reference, crestStage, or waterMark)}
}
\value{
data frame of reading records
}
\description{
get the list of readings attached to a report. Will include a year+month field as a month identifier for each record.
}

