% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readReportMetadataField}
\alias{readReportMetadataField}
\title{Read report metadata field}
\usage{
readReportMetadataField(reportObject, field, defaultValue)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}

\item{field}{the field name to read from the metadata}
}
\description{
Given a full report object and field name, returns the
metadata value for the provided field.
}

