% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readSecondaryTimeSeriesUvInfo}
\alias{readSecondaryTimeSeriesUvInfo}
\title{Read Secondary Time Series UV INFO
Decides where or not the report wants reference or upchain series as it's main secondary series and returns metadata for that.}
\usage{
readSecondaryTimeSeriesUvInfo(reportObject)
}
\arguments{
\item{reportObject}{full UV Hydro object}
}
\value{
timeseries metadata see readTimeSeriesUvInfo
}
\description{
Read Secondary Time Series UV INFO
Decides where or not the report wants reference or upchain series as it's main secondary series and returns metadata for that.
}
