% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readSRSQualifiers}
\alias{readSRSQualifiers}
\title{Read Qualifiers (SRS)}
\usage{
readSRSQualifiers(reportObject, timezone, qualifierMetadata)
}
\arguments{
\item{reportObject}{The reading JSON object}

\item{timezone}{The timezone of the report}

\item{qualifierMetadata}{The metadata of the qualifiers found in the readings}
}
\description{
Reads and formats the qualifiers
}
