% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readThresholds}
\alias{readThresholds}
\title{Read Thresholds (CORR)}
\usage{
readThresholds(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Reads and formats the Thresholds data
}

