% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readTimeSeriesUvInfo}
\alias{readTimeSeriesUvInfo}
\title{Read Time Series UV Info}
\usage{
readTimeSeriesUvInfo(reportObject, seriesName)
}
\arguments{
\item{reportObject}{full UV Hydro object}

\item{seriesName}{name of ts field to get info from}
}
\value{
named list of metadata (label, unit, type)
}
\description{
Returns metadata for a TS.
}
