% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readTSSThresholds}
\alias{readTSSThresholds}
\title{Read Thresholds (TSS)}
\usage{
readTSSThresholds(reportObject)
}
\arguments{
\item{reportObject}{the full report JSON object}
}
\description{
Reads and formats the Thresholds data
}
