% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readUpchainSeries}
\alias{readUpchainSeries}
\title{Read Upchain Series (TSS)}
\usage{
readUpchainSeries(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Reads and formats the related upchain series
}
