% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readUvGageHeight}
\alias{readUvGageHeight}
\title{Read Gage Height}
\usage{
readUvGageHeight(reportObject, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}
}
\value{
subset of gage height data
}
\description{
Read gage height subsetted by month for UV Hydrograph
}
