% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readUvMeasurementShifts}
\alias{readUvMeasurementShifts}
\title{Read Measured Shifts}
\usage{
readUvMeasurementShifts(reportObject, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}
}
\value{
subset of measured shifts data
}
\description{
Read measured shifts subsetted by month for UV Hydrograph
}

