% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{readUvWq}
\alias{readUvWq}
\title{Read Water Quality Measurements}
\usage{
readUvWq(reportObject, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}
}
\value{
subset of WQ data, default to empty frame if none found
}
\description{
Read WQ measurements subsetted by month for UV Hydrograph
}

