% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readWaterQualityMeasurements}
\alias{readWaterQualityMeasurements}
\title{Read water quality measurements}
\usage{
readWaterQualityMeasurements(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\value{
data frame
}
\description{
Given a full report object, returns the water quality
measurements formatted as a time series point set.
}
