% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{removeZeroNegative}
\alias{removeZeroNegative}
\title{Remove Zero or Negative Values}
\usage{
removeZeroNegative(df)
}
\arguments{
\item{df}{A data frame with a column called \code{value} that it will check
for negative or zeroes.}
}
\description{
If negative or zero values are passed into function, turn them into NA 
so we can log the plot.
}
\details{
Make sure you have the column called \code{value} included in data
  frame.
}

