% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{renderCustomFragments}
\alias{renderCustomFragments}
\title{Peform the rendering of subfragments for a given report, should return a list of html fragments to 
be available to inject into mustach templates}
\usage{
renderCustomFragments(reportData)
}
\arguments{
\item{reportData}{typed report json, the type should be the name of the report}
}
\value{
list of rendered HTML fragments that will be included in the data available to the final report template
}
\description{
Peform the rendering of subfragments for a given report, should return a list of html fragments to 
be available to inject into mustach templates
}
