% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-renderReport.R
\docType{methods}
\name{renderReport}
\alias{renderReport}
\alias{renderReport,list,character,character-method}
\alias{renderReport}
\title{Render Report}
\usage{
renderReport(data, reportType, author, ...)

\S4method{renderReport}{list,character,character}(data, reportType, author, ...)
}
\arguments{
\item{data}{raw json data}

\item{reportType}{type of report, will error if not supported}

\item{author}{label for person generating report}

\item{...}{remaining parameters}
}
\description{
Render Report
}
\examples{
library(jsonlite)
renderReport(
  fromJSON(
    system.file(
      'extdata', 'example-report.json', package = 'repgen'
    )
  ), 
'examplereport', 
'Author Name')
}
