% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{renderVDiagram}
\alias{renderVDiagram}
\title{Called from V diagram R Markdown files.}
\usage{
renderVDiagram(reportObject)
}
\arguments{
\item{reportObject}{V diagram report data.}
}
\description{
Called from V diagram R Markdown files.
}

