% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{RescaleYTop}
\alias{RescaleYTop}
\title{Rescale the Top of y-axis to Create an Approximately 4% Margin Between the
Vertical Top Extent of Plot Objects and the Top Edge of Plot}
\usage{
RescaleYTop(object)
}
\arguments{
\item{object}{A gsplot, plot object.}
}
\value{
The passed-in gsplot object, with the y-axis top augmented upwards
        appropriately.
}
\description{
This function is an inaccurate emulation of (the top-end-of-plot
  behavior of) R \code{graphics::par}'s \code{yaxs = 'r'} state, because we have
  to use \code{yaxs = 'i'} in spots, but still want the ~4% margin at the top
  of the plot, so we adjust the y-axis endpoint accordingly after we do what
  we need.
}

