% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveyeargwsum-render.R
\name{resortTimeSeries}
\alias{resortTimeSeries}
\title{Resort Time Series}
\usage{
resortTimeSeries(reportObject)
}
\arguments{
\item{reportObject}{the JSON to inspect and reassign/re-order time series}
}
\value{
reportObject with reassigned time series
}
\description{
Before we start assigning time series and rendering plot components, resort the timeseries so we don't have any empty firstStatDerived slots
otherwise the min/max and approval bars dont draw correctly. 
Also, assign the prefixes from the original data so the user doesn't know we're shuffling things around behind the scenes
}
