% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-legend.R
\name{rmDuplicateLegendItems}
\alias{rmDuplicateLegendItems}
\title{Remove Duplicate Legend Items}
\usage{
rmDuplicateLegendItems(object)
}
\arguments{
\item{object}{The plot object which contains the legend.}
}
\value{
The changed object without duplicate legend keys.
}
\description{
Checks a plot object for duplicate keys in the legend.
}

