% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-sitevisitpeak.R
\docType{methods}
\name{sitevisitpeak}
\alias{sitevisitpeak}
\alias{sitevisitpeak,character,character-method}
\alias{sitevisitpeak,list,character-method}
\alias{sitevisitpeak,list,missing-method}
\title{sitevisitpeak report}
\usage{
sitevisitpeak(data, ...)

\S4method{sitevisitpeak}{list,character}(data, ...)

\S4method{sitevisitpeak}{character,character}(data)

\S4method{sitevisitpeak}{list,missing}(data)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{everythingn else}
}
\examples{
library(jsonlite)
data <- fromJSON(system.file('extdata','sitevisitpeak',"sitevisitpeak-example.json",package = 'repgen'))
sitevisitpeak(data, 'Author Name')
}

