% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{sizeOf}
\alias{sizeOf}
\title{Get the size of a dataframe.}
\usage{
sizeOf(df)
}
\arguments{
\item{df}{the data frame to get the size of}
}
\description{
Will throw an error if data frame is NULL or NA.
}

