% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{sortDataAndSides}
\alias{sortDataAndSides}
\title{Sort data and sides}
\usage{
sortDataAndSides(primarySeriesList, uvInfo, refInfo, compInfo, isDischarge)
}
\arguments{
\item{primarySeriesList}{list of timeseries available for reporting}

\item{uvInfo}{main timeseries information for the plot}

\item{refInfo}{timeseries information for reference series}

\item{compInfo}{timeseries information for comparios series}

\item{isDischarge}{whether or not the primary series is discharge}
}
\value{
named list with two items, sides and ylims. Each item has a named list with items for each timeseries. (EG: side for reference would be returnedObject$sides$reference)
}
\description{
Depending on the configuration, reference and comparison series will be plotted on different sides. This constructs ylims and what side each series should be on
}
