% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{splitDataGaps}
\alias{splitDataGaps}
\title{If there are gaps in the time series, don't connect them. This creates
multiple line/point calls if there are gaps.}
\usage{
splitDataGaps(data, ts, isDV)
}
\arguments{
\item{data}{The original list format of JSON.}

\item{ts}{Current time series data.}

\item{isDV}{Logic for whether this plot uses daily values or not.}
}
\description{
If there are gaps in the time series, don't connect them. This creates
multiple line/point calls if there are gaps.
}

