% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{splitDataGapsList}
\alias{splitDataGapsList}
\title{Split multiple time series when there are gaps}
\usage{
splitDataGapsList(allVars, timezone, flagZeroNeg, isDV = FALSE)
}
\arguments{
\item{allVars}{contains all ts/vars that are not empty (equals allVars in the *-data.R script)}

\item{timezone}{string giving the timezone}

\item{flagZeroNeg}{logical indicating whether or not the zeros & negatives can be 
removed for logging the axis}

\item{isDV}{logical saying whether or not the time series is made of daily values; default is FALSE}
}
\description{
Given a list of time series, this will return a list where gaps in data
have been applied to each. The names of the time series will be duplicated if there
were gaps that caused the individual time series to break into multiple.
}
