% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{splitDataGapsTimeSeries}
\alias{splitDataGapsTimeSeries}
\title{Split a time series if there are gaps.}
\usage{
splitDataGapsTimeSeries(timeSeries, timeSeriesName, timezone, flagZeroNeg,
  isDV = FALSE)
}
\arguments{
\item{timeSeries}{current timeseries data as a list}

\item{timeSeriesName}{string giving name that should be given to each list item returned}

\item{timezone}{string giving the timezone}

\item{flagZeroNeg}{logical indicating whether or not the zeros & negatives can be 
removed for logging the axis}

\item{isDV}{logical saying whether or not the time series is made of daily values; default is FALSE}
}
\description{
If there are gaps in the timeseries, don't connect them.
This returns a list of data that can be used for multiple line/point 
calls in the plotting functions.
}

