% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{splitShiftedLabels}
\alias{splitShiftedLabels}
\title{Split Shifted Labels}
\usage{
splitShiftedLabels(inputLabels, startLabelIndex)
}
\arguments{
\item{inputLabels}{The full list of labels from the lane}

\item{startLabelIndex}{The row number to start labeling removed labels from}
}
\value{
A list containing the labels for the table, the update plot labels,
and the last row number that was labeled (to start the next lane labels from)
}
\description{
Given the full list labels for a data lane splits out the
labels that are marked as being shifted and stores them in a list to be
put into the label table. The labels on the plot are replaced by numbers
representing the table row that is holding the relevant label.
}

