% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{srsQualifiersTable}
\alias{srsQualifiersTable}
\title{Create flat text, "qualifiers table" type output table.}
\usage{
srsQualifiersTable(data, table)
}
\arguments{
\item{data}{Report data.}

\item{table}{A vector from which to derive qualifiers from.}
}
\value{
A vector of strings.
}
\description{
Create flat text, "qualifiers table" type output table.
}

