% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fiveyeargwsum-render.R
\name{startFiveYearRender}
\alias{startFiveYearRender}
\title{Starting point, creates RMD and runs rendering}
\usage{
startFiveYearRender(data, output, author)
}
\arguments{
\item{data}{coming in to create a plot}

\item{output}{format (either html or pdf)}

\item{author}{name of person generating the report}
}
\description{
Starting point, creates RMD and runs rendering
}

