% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{startRender}
\alias{startRender}
\title{Starting point, creates RMD and runs rendering}
\usage{
startRender(data, output, author, reportName)
}
\arguments{
\item{data}{coming in to create a plot}

\item{output}{format (html)}

\item{author}{name of person generating the report}

\item{reportName}{name of report being generated (current options: dvhydrograph, fiveyruvhydrograph, vdiagram)}
}
\description{
Starting point, creates RMD and runs rendering
}

