% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{startTemplatedRender}
\alias{startTemplatedRender}
\title{Starting point for template based rendering, this will load common templates and render a shared header/footer and then load
the templates for the report (the reportJson type) and perform execute the renderFragments function to retrieve report-defined
fragments.}
\usage{
startTemplatedRender(reportJson, author)
}
\arguments{
\item{reportJson}{data coming in to create the report, this needs to be typed and that type needs to have the "renderFragments" function defined for it. The type will also be the name of the report.}

\item{author}{string to identifier for person calling report}
}
\value{
the html file and other report bits
}
\description{
this function orchestrates the creation of the html reports using whisker, the final result is written to a file, and that file name is returned.
}
