% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{startUvhydrographRender}
\alias{startUvhydrographRender}
\title{Starting point, creates RMD and runs rendering}
\usage{
startUvhydrographRender(data, output, author)
}
\arguments{
\item{data}{coming in to create a plot}

\item{output}{format (either html or pdf)}

\item{author}{name of person generating the report}
}
\description{
Starting point, creates RMD and runs rendering
}

