% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{subsetByMonth}
\alias{subsetByMonth}
\title{Extract time value points list by a single month}
\usage{
subsetByMonth(pts, onlyMonth)
}
\arguments{
\item{pts}{a list of time and value points}

\item{onlyMonth}{the numeric month to subset the list of time and value points by}
}
\description{
Subset the data passed into the function by the month passed in
}
