% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-limits.R
\name{testCallouts}
\alias{testCallouts}
\title{Test Callouts}
\usage{
testCallouts(plot_obj, xlimits)
}
\arguments{
\item{plot_obj}{The plot object whose callouts will be checked}

\item{xlimits}{The maximum x coordinates available on the graph.}
}
\value{
The plot object changed so that all callouts fit and show on the plot.
}
\description{
A function to ensure that callouts on a plot fit within the x-limits.
}

