% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-limits.R
\name{testCalloutsByView}
\alias{testCalloutsByView}
\title{Test Callouts by View}
\usage{
testCalloutsByView(plot_obj, view_num, xlimits_real, width_char, xrange)
}
\arguments{
\item{plot_obj}{The overarching plot object that contains these objects.}

\item{view_num}{The axis which the callouts are being compared to. Starts at the bottom x-axis
as 1, and goes clockwise from there. (bottom, left, top, right)}

\item{xlimits_real}{The x-axis limits (lower and upper limits) at which point callouts are cut off.}

\item{width_char}{The width of a character on the plot. Used to determine if a string is cut off.}

\item{xrange}{The limits of the graph not taking into account a small amount of buffer room.}
}
\value{
The plot object with the callouts changed to not overlap with the x-axis limits.
}
\description{
A function that calculates if a callout goes past an x-axis limit.
}

