% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{toStartOfDay}
\alias{toStartOfDay}
\title{given a datetime, will remove time and set 0000 as start}
\usage{
toStartOfDay(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
given a datetime, will remove time and set 0000 as start
}

