% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{toStartOfMonth}
\alias{toStartOfMonth}
\title{given a datetime, will change the date to the 1st, plus remove time and set 0000
this is used to set the labels for Five Year Groundwater Summary reports}
\usage{
toStartOfMonth(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
given a datetime, will change the date to the 1st, plus remove time and set 0000
this is used to set the labels for Five Year Groundwater Summary reports
}

