% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{toEndOfTime}
\alias{toEndOfTime}
\title{hacky fix for 9999 year issue which prevents the rectangles from displaying on the graphs 
apologies to the people of 2100 who have to revisit this}
\usage{
toEndOfTime(time)
}
\arguments{
\item{time}{datetime to shift to "the end of time"}
}
\description{
hacky fix for 9999 year issue which prevents the rectangles from displaying on the graphs 
apologies to the people of 2100 who have to revisit this
}

