% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{toSentenceCase}
\alias{toSentenceCase}
\title{Camel Case String}
\usage{
toSentenceCase(string)
}
\arguments{
\item{string}{The string to format}
}
\description{
Formats the supplied string in CamelCasing
}
