% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{toEndOfDay}
\alias{toEndOfDay}
\title{given a datetime, will remove time and set 2359}
\usage{
toEndOfDay(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
given a datetime, will remove time and set 2359
}

