% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{toEndOfMonth}
\alias{toEndOfMonth}
\title{given a datetime, will change the date to the 30th, plus remove time and set 2359
this is used to set the labels for Five Year Groundwater Summary reports}
\usage{
toEndOfMonth(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
given a datetime, will change the date to the 30th, plus remove time and set 2359
this is used to set the labels for Five Year Groundwater Summary reports
}

