% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{translateDateTimes}
\alias{translateDateTimes}
\title{Translate date/times}
\usage{
translateDateTimes(series, timezone)
}
\arguments{
\item{series}{The series to translate into format that can be used for comparison}

\item{timezone}{The timezone for the data}
}
\value{
series with new fields for comparison
}
\description{
Add date/times that is in a format we can use for comparison
}
