% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{unNestCorrectionParameters}
\alias{unNestCorrectionParameters}
\title{Unnest the correction parameters}
\usage{
unNestCorrectionParameters(corrections, timezone)
}
\arguments{
\item{corrections}{the corrections for the report}

\item{timezone}{the timezone to parse data into}
}
\value{
corrections formatted better for report display
}
\description{
Takes the corrections and unnests them so they can 
more easily be processed for each type
}
