% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{useSecondaryPlot}
\alias{useSecondaryPlot}
\title{Use Secondary Plot}
\usage{
useSecondaryPlot(reportObject)
}
\arguments{
\item{reportObject}{the report to be rendered}
}
\value{
boolean true or false of whether a secondary series is appropriate for report
}
\description{
Determines where or not a report should use a secondary plot
}

