% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{uvhydrographPlot}
\alias{uvhydrographPlot}
\title{Create a uvhydrograph}
\usage{
uvhydrographPlot(data)
}
\arguments{
\item{data}{coming in to create a plot}
}
\description{
Create a uvhydrograph
}

