% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validateFetchedData}
\alias{validateFetchedData}
\title{Validate to ensure it is not null or empty and has all required fields}
\usage{
validateFetchedData(data, name, requiredFields)
}
\description{
Given some data and required fields, will check to ensure
that the supplied data is not null or empty and has all required fields.
Will throw an error if either of these checks fail.
}

