% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validParam}
\alias{validParam}
\title{validParam}
\usage{
validParam(val, param, required = FALSE, as.numeric = FALSE)
}
\arguments{
\item{val}{R object.}

\item{param}{string.}

\item{required}{boolean, defaults to FALSE}

\item{as.numeric, }{defaults to FALSE}
}
\description{
Check if a value is NULL.  If NULL and required, returns error message.  
if not required, forces a return of NA.  If not NULL, returns \code{val}.
}
\details{
Checks for valid parameter
}
