% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-vdiagram.R
\docType{methods}
\name{vdiagram}
\alias{vdiagram}
\alias{vdiagram,list-method}
\title{V Diagram Report}
\usage{
vdiagram(data, ...)

\S4method{vdiagram}{list}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{addtional params passed to GET}
}
\description{
V Diagram Report
}
\examples{
library(jsonlite)
library(gsplot)
json_file <- system.file('extdata','vdiagram','vdiagram-example.json', package = 'repgen')
data <-fromJSON(json_file)
vdiagram(data, 'Author Name')
}

