% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{vdiagramTable}
\alias{vdiagramTable}
\title{V diagram table from data inputs.}
\usage{
vdiagramTable(data)
}
\arguments{
\item{data}{A list of properly formatted V diagram data.}
}
\value{
A string properly formatted for HTML.
}
\description{
V diagram table from data inputs.
}

