% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{XAxisLabelStyle}
\alias{XAxisLabelStyle}
\title{X-Axis Label style}
\usage{
XAxisLabelStyle(object, start, end, timezone, plotDates)
}
\arguments{
\item{object}{the plot object to create labels for}

\item{start}{the start date of the date range}

\item{end}{the end date of the date range}

\item{timezone}{the timezone of the date range}

\item{plotDates}{the dates to create the labels at}
}
\description{
Given a plot object and date range parameters,
creates proper X-Axis labels based on the duration of the
date range. Including Year and Month subsets.
}

