% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{YAxisInterval}
\alias{YAxisInterval}
\title{Compute the y-axis real interval, based on a heuristic.}
\usage{
YAxisInterval(corr.value.sequence, uncorr.value.sequence)
}
\arguments{
\item{A}{sequence of corrected time series points.}

\item{A}{sequence of uncorrected time series points.}
}
\value{
The y-axis real interval, as ordered-pair vector.
}
\description{
Compute the y-axis real interval, based on a heuristic.
}

