% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{YEndpoint}
\alias{YEndpoint}
\title{Compute the y-axis endpoint, based on a heuristic.}
\usage{
YEndpoint(corr.value.sequence, uncorr.value.sequence)
}
\arguments{
\item{corr.value.sequence}{A sequence of corrected time series points.}

\item{uncorr.value.sequence}{A sequence of uncorrected time series points.}
}
\value{
The y-axis real interval endpoint.
}
\description{
Compute the y-axis endpoint, based on a heuristic.
}

