% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{YOrigin}
\alias{YOrigin}
\title{Compute the y-axis origin, based on a heuristic.}
\usage{
YOrigin(corr.value.sequence, uncorr.value.sequence)
}
\arguments{
\item{corr.value.sequence}{A sequence of corrected time series points.}

\item{uncorr.value.sequence}{A sequence of uncorrected time series points.}
}
\value{
The y-axis real interval origin.
}
\description{
Compute the y-axis origin, based on a heuristic.
}

