% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{zeroValues}
\alias{zeroValues}
\title{zeroValues}
\usage{
zeroValues(timeSeriesData)
}
\arguments{
\item{timeSeriesData}{The time series that is checked for data containing zero}
}
\value{
A logical value that is true if there is any data that is zero.
}
\description{
Function that returns the data equal to zero.
}
