% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{AddToGsplot}
\alias{AddToGsplot}
\title{Add To gsplot}
\usage{
AddToGsplot(gsplot, plotConfig)
}
\arguments{
\item{gsplot}{A gsplot, plot object.}

\item{plotConfig}{of gsplot calls to make}
}
\value{
A modified gsplot, plot object, with everything in the plot config included.
}
\description{
helper function to add arbitrary list of gsplot calls to a gsplot. Has added feature of extending
the Y limits for any errors bars found.
}

