% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{anyDataExist}
\alias{anyDataExist}
\title{anyDataExist}
\usage{
anyDataExist(data)
}
\arguments{
\item{data}{dataframe, list}
}
\description{
Check if data is NULL, has zero length or has no rows
}
\details{
Checks that any data exists
}
