% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{ApplyApprovalBarStyles}
\alias{ApplyApprovalBarStyles}
\title{Apply styles (and some properties) to approval bar rectangles.}
\usage{
ApplyApprovalBarStyles(object, data)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{data}{A list of gsplot objects to display on the plot.}
}
\value{
gsplot object with approval bar rectangle styles applied.
}
\description{
Apply styles (and some properties) to approval bar rectangles.
}
