% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{applyDataGaps}
\alias{applyDataGaps}
\title{use splitDataGaps and format the resulting data correctly}
\usage{
applyDataGaps(data, relevantData)
}
\description{
use splitDataGaps and format the resulting data correctly
}

