% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{applyDataGaps}
\alias{applyDataGaps}
\title{Use \code{splitDataGaps} and Format the Resulting Data Correctly}
\usage{
applyDataGaps(data, relevantData, isDV = FALSE)
}
\arguments{
\item{data}{The original list format of JSON.}

\item{relevantData}{All time-series/variables that are not empty (equals
\code{allVars} in the \code{*-data.R} script).}

\item{isDV}{Logic for whether this plot uses daily values or not.}
}
\description{
Use \code{splitDataGaps} and Format the Resulting Data Correctly
}

