% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{applyDataGaps}
\alias{applyDataGaps}
\title{use splitDataGaps and format the resulting data correctly}
\usage{
applyDataGaps(data, relevantData, isDV = FALSE)
}
\arguments{
\item{data}{original list format of JSON}

\item{relevantData}{contains all ts/vars that are not empty (equals allVars in the *-data.R script)}

\item{isDV}{logic for whether this plot uses daily values or not}
}
\description{
use splitDataGaps and format the resulting data correctly
}

