% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{applyQualifiers}
\alias{applyQualifiers}
\title{Apply qualifiers}
\usage{
applyQualifiers(reportObject)
}
\arguments{
\item{reportObject}{the extremes report object}
}
\value{
the same reportObject, but with all values updated with qualifiers prefixed as CSV
}
\description{
Will apply all qualifiers to all values in the report object
}
