% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{ApprovalBarY}
\alias{ApprovalBarY}
\title{Compute the Top or Bottom Vertical Position of Approval Bars}
\usage{
ApprovalBarY(lim, ylog = NULL, reverse, ratio)
}
\arguments{
\item{lim}{The y-axis real interval, as two element vector.}

\item{ylog}{A Boolean truth value, indicating whether the y-axis is log_10 scale:
TRUE => log_10; FALSE => linear.}

\item{reverse}{A Boolean truth value, indicating whether the y-axis is inverted:
TRUE => inverted y-axis; FALSE => not inverted.}

\item{ratio}{A scaling ratio to adjust the top or bottom of approval bar
rectangle.}
}
\value{
The appropriate approval bar top or bottom y-axis coordinate, in
        world coordinates.
}
\description{
Compute the Top or Bottom Vertical Position of Approval Bars
}

