% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{approvalBarYBottom}
\alias{approvalBarYBottom}
\title{Compute bottom position of approval bars.}
\usage{
approvalBarYBottom(lim, ylog)
}
\arguments{
\item{lim}{The y-axis real interval, as two element vector.}

\item{ylog}{A Boolean, indicating whether the y-axis is log_10 scale:
TRUE => log_10; FALSE => linear.}
}
\value{
Approval bar, vertical bottom extent, in world coordinates.
}
\description{
Compute bottom position of approval bars.
}
