% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{ApprovalBarYTop}
\alias{ApprovalBarYTop}
\title{Compute the Top Position of Approval Bars}
\usage{
ApprovalBarYTop(lim, ylog, reverse)
}
\arguments{
\item{lim}{The y-axis real interval, as two element vector.}

\item{ylog}{A Boolean truth value indicating whether the y-axis is log_10
scale: TRUE => log_10; FALSE => linear.}

\item{reverse}{A Boolean truth value indicating whether the y-axis is
inverted: TRUE => inverted y-axis; FALSE => not inverted.}
}
\value{
The appropriate approval bar vertical top extent, in world
        coordinates.
}
\description{
Compute the Top Position of Approval Bars
}

