% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{approvalBarYTop}
\alias{approvalBarYTop}
\title{Compute Top Position of Approval Bars}
\usage{
approvalBarYTop(lim, ylog, reverse)
}
\arguments{
\item{lim}{The \emph{y}-axis real interval, as two element vector.}

\item{ylog}{A Boolean, indicating whether the \emph{y}-axis is log10
scale: \code{TRUE} => log10; \code{FALSE} => linear.}

\item{reverse}{A Boolean, indicating whether the \emph{y}-axis is inverted:
\code{TRUE} => inverted \emph{y}-axis; \code{FALSE} => not inverted.}
}
\value{
Approval bar, vertical top extent, in world coordinates.
}
\description{
Compute Top Position of Approval Bars
}
\author{
Andrew Halper
}

