% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{calcEndSeq}
\alias{calcEndSeq}
\title{Get a list of end dates}
\usage{
calcEndSeq(startSeq, endD)
}
\arguments{
\item{startSeq}{the start date}

\item{endD}{the end date}
}
\value{
a list of end dates for the corr report sections
}
\description{
Given a startSeq date list and end date and whether 
or not the end date is the first of the month, provides a list of 
YYYY-MM-DD used in other functions
}
