% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{calcStartSeq}
\alias{calcStartSeq}
\title{Get a list of start dates}
\usage{
calcStartSeq(startD, endD, firstOfMonth)
}
\arguments{
\item{startD}{the start date}

\item{endD}{the end date}

\item{firstOfMonth}{TRUE or FALSE whether the startD is the first of the month}
}
\value{
a list of start dates for the corr report sections
}
\description{
Given a start date and end date and whether 
or not the start date is the first of the month, provides a list of 
YYYY-MM-DD used in other functions
}

