% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{convertStringToTableDisplay}
\alias{convertStringToTableDisplay}
\title{Convert the string to the equivalent HTML code}
\usage{
convertStringToTableDisplay(characters)
}
\arguments{
\item{characters}{The string to convert}
}
\value{
The equivalent HTML codes for that string
}
\description{
Convert the string to the equivalent HTML code
}

