% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{convertTableDisplayToString}
\alias{convertTableDisplayToString}
\title{Convert the String from HTML code to the equivalent raw characters}
\usage{
convertTableDisplayToString(characters)
}
\arguments{
\item{characters}{The characters to convert}
}
\value{
The equivalent string for the HTML codes
}
\description{
Convert the String from HTML code to the equivalent raw characters
}

