% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-correctionsataglance.R
\docType{methods}
\name{correctionsataglance}
\alias{correctionsataglance}
\alias{correctionsataglance,character-method}
\alias{correctionsataglance,list-method}
\title{Corrections-At-A-Glance report}
\usage{
correctionsataglance(data, ...)

\S4method{correctionsataglance}{list}(data, ...)

\S4method{correctionsataglance}{character}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET or authenticateUser}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)
Sys.setenv(TZ = "UTC")

data <-
 fromJSON(
   system.file(
     'extdata', 'correctionsataglance', 'correctionsataglance-example2.json',
     package = 'repgen'
   )
 )
correctionsataglance(data, 'Author Name')
}

