% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-correctionsataglance.R
\docType{methods}
\name{correctionsataglance}
\alias{correctionsataglance}
\alias{correctionsataglance,list-method}
\title{Corrections-at-a-Glance report}
\usage{
correctionsataglance(data, ...)

\S4method{correctionsataglance}{list}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{Additional parameters passed to GET.}
}
\description{
Corrections-at-a-Glance report
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)
Sys.setenv(TZ = "UTC")

data <-
  fromJSON(
    system.file(
      'extdata', 'correctionsataglance', 'correctionsataglance-example2.json',
      package = 'repgen'
    )
  )
correctionsataglance(data, 'Author Name')
}
