% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-correctionsataglance.R
\docType{methods}
\name{correctionsataglance}
\alias{correctionsataglance}
\alias{correctionsataglance,character,character-method}
\alias{correctionsataglance,list,character-method}
\alias{correctionsataglance,list,missing-method}
\title{Corrections-At-A-Glance report}
\usage{
correctionsataglance(data, output, ...)

\S4method{correctionsataglance}{list,character}(data, output, ...)

\S4method{correctionsataglance}{character,character}(data, output, ...)

\S4method{correctionsataglance}{list,missing}(data, output, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{additional params passed to GET or authenticateUser}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)
data <- fromJSON(system.file('extdata','correctionsataglance','correctionsataglance-example2.json', package = 'repgen'))
correctionsataglance(data, 'html', 'Author Name')

}

