% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{createDataRows}
\alias{createDataRows}
\title{create a set of rows for one data parameter}
\usage{
createDataRows(data, param, rowName, isUpchain, includeRelated = TRUE,
  doMerge = TRUE)
}
\arguments{
\item{data}{a set of extremes report data for either upchain, primary, or dv}

\item{param}{either "min" or "max" to specify if we are generating minimum or maximum rows}

\item{rowName}{the name to use for the specified row}

\item{isUpchain}{whether or not this is the upchain dataset and it needs to compare to primary}

\item{includeRelated}{whether or not there is a second column of corresponding data}

\item{doMerge}{whether or not we should merge duplicate rows}
}
\value{
list dataRows
}

