% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{createPlotLanes}
\alias{createPlotLanes}
\title{Create Plot lanes}
\usage{
createPlotLanes(approvalData, requiredData, requiredNames, optionalData,
  optionalNames, dateRange, startSeq, endSeq)
}
\description{
Given lists of required and optional data, creates the
data lanes for the plot and returns them as a list.
}

