% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{createSecondaryPlot}
\alias{createSecondaryPlot}
\title{TODO documentation}
\usage{
createSecondaryPlot(uvInfo, timeInformation, secondarySeriesList, approvalBars,
  effective_shift_pts, meas_shift, gage_height, corrections, lims, invertPlot,
  tertiary_label = "")
}
\description{
TODO documentation
}

