% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-data.R
\name{defaultHistFlags}
\alias{defaultHistFlags}
\title{Set default historic flags}
\usage{
defaultHistFlags(histFlag)
}
\arguments{
\item{histFlag}{The historic flag field from a V Diagram.}
}
\value{
FALSE if " " is passed in or if the field is empty, otherwise it returns what was passed in.
}
\description{
A function that sets the default historic flag if " " is passed in.
}

