% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{DelineateYearBoundaries}
\alias{DelineateYearBoundaries}
\title{DelineateYearBoundaries}
\usage{
DelineateYearBoundaries(object, years)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{years}{A sequence of year begin dates to draw the lines at.}
}
\value{
The passed-in gsplot object, with year boundaries delineated.
}
\description{
Delineate year boundaries on five year GW summary plots, and DV hydrographs
having time intervals of one year or more.
}
\details{
Delineate year boundaries
}
