% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-dvhydrograph.R
\docType{methods}
\name{dvhydrograph}
\alias{dvhydrograph}
\alias{dvhydrograph,list-method}
\title{DV Hydrograph report}
\usage{
dvhydrograph(data, ...)

\S4method{dvhydrograph}{list}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}
}

