% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-dvhydrograph.R
\docType{methods}
\name{dvhydrograph}
\alias{dvhydrograph}
\alias{dvhydrograph,character-method}
\alias{dvhydrograph,list-method}
\title{DV Hydrograph report}
\usage{
dvhydrograph(data, ...)

\S4method{dvhydrograph}{list}(data, ...)

\S4method{dvhydrograph}{character}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET or authenticateUser}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)
Sys.setenv(TZ = "UTC")
data <- fromJSON(system.file('extdata','dvhydrograph','dvhydro-example.json', package = 'repgen'))
dvhydrograph(data, 'Author Name')
\dontrun{
url <- paste0('https://nwissddvasvis01.cr.usgs.gov/service/timeseries/reports/swdvhydrograph/',
'?station=05421682&dischargeIdentifier=Discharge.ft\%5E3\%2Fs&stageIdentifier=',
'Gage+height.ft.Work&dailyDischargeIdentifier=Discharge.ft\%5E3\%2Fs.Mean',
'&ratingModelIdentifier=Gage+height-Discharge.STGQ&waterYear=2011')

# pass in additional params to authenticateUser
dvhydrograph(url, verbose = TRUE, username = 'bbadger', password = '12345')
dvhydrograph(url, 'Author Name')
}
}

