% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{emptyData}
\alias{emptyData}
\title{Returns whether or not all required columns are present}
\usage{
emptyData(data, name, requiredFields)
}
\description{
Returns whether or not all required columns are present
}

