% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{extendStep}
\alias{extendStep}
\title{Use the last point plus 2400 to extend step
the points do not have times, but the x limit is extended with a time to show the whole day
the step needs to be extended to meet this time}
\usage{
extendStep(toPlot)
}
\arguments{
\item{toPlot}{list of items that will be called in the do.call}
}
\description{
Use the last point plus 2400 to extend step
the points do not have times, but the x limit is extended with a time to show the whole day
the step needs to be extended to meet this time
}

