% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-extremes.R
\docType{methods}
\name{extremes}
\alias{extremes}
\alias{extremes,character-method}
\alias{extremes,list-method}
\title{extremes report}
\usage{
extremes(data, ...)

\S4method{extremes}{list}(data, ...)

\S4method{extremes}{character}(data)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{everything else}
}
\examples{
library(jsonlite)
library(dplyr)
Sys.setenv(TZ = "UTC")
data <-
 fromJSON(
   system.file(
     'extdata', 'extremes', 'extremes-example-site-train.json', package = 'repgen'
   )
 )
extremes(data, 'Author Name')
}

