% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-extremes.R
\docType{methods}
\name{extremes}
\alias{extremes}
\alias{extremes,list-method}
\alias{extremes}
\title{Extremes report.}
\usage{
extremes(data, ...)

\S4method{extremes}{list}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{Everything else.}
}
\description{
Extremes report.
}
\examples{
library(jsonlite)
library(dplyr)
data <-
  fromJSON(
    system.file(
      'extdata', 'extremes', 'extremes-example-site-train.json', package = 'repgen'
    )
  )
extremes(data, 'Author Name')
}
