% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-extremes.R
\docType{methods}
\name{extremes}
\alias{extremes}
\alias{extremes,character,character-method}
\alias{extremes,list,character-method}
\alias{extremes,list,missing-method}
\title{extremes report}
\usage{
extremes(data, output, ...)

\S4method{extremes}{list,character}(data, output, ...)

\S4method{extremes}{character,character}(data, output)

\S4method{extremes}{list,missing}(data, output)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{everything else}
}
\examples{
library(jsonlite)
data <- fromJSON(system.file('extdata','extremes', 'extremes-example.json',package = 'repgen'))
extremes(data, 'html', 'Author Name')
}

