% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extremes.R
\name{extremes}
\alias{extremes}
\title{extremes report}
\usage{
extremes(data, output, token)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{token}{an auth token (see \code{\link{authenticate_user}})}
}
\description{
extremes report
}

