% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesQualifiersTable}
\alias{extremesQualifiersTable}
\title{Create a Flat Text "qualifiers table" Type Output Table}
\usage{
extremesQualifiersTable(data, table)
}
\arguments{
\item{data}{Report data.}

\item{table}{A vector to derive qualifiers from.}
}
\value{
A vector of qualifiers.
}
\description{
Create a Flat Text "qualifiers table" Type Output Table
}

