% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesQualifiersTable}
\alias{extremesQualifiersTable}
\title{Create flat text "qualifiers table" type output table.}
\usage{
extremesQualifiersTable(data, table)
}
\arguments{
\item{data}{Report data.}

\item{table}{A vector to derive qualifiers from.}
}
\value{
A vector of qualifiers.
}
\description{
Create flat text "qualifiers table" type output table.
}

