% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesQualifiersTable}
\alias{extremesQualifiersTable}
\title{Create a Flat Text "qualifiers table" Type Output Table}
\usage{
extremesQualifiersTable(reportObject, table, primaryHeaderTerm,
  upchainHeaderTerm)
}
\arguments{
\item{reportObject}{Report data.}

\item{table}{A vector to derive qualifiers from.}

\item{primaryHeaderTerm}{a unique term to help determine which column title is for primary data}

\item{upchainHeaderTerm}{a unique term to help determine which column title is for upchain data}
}
\value{
A vector of qualifiers.
}
\description{
Create a Flat Text "qualifiers table" Type Output Table
}

