% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-data.R
\name{extremesTable}
\alias{extremesTable}
\title{Create a Flat Text "extremes table" Type Output Table}
\usage{
extremesTable(rawData)
}
\arguments{
\item{rawData}{Extremes report JSON string.}
}
\value{
A vector of strings.
}
\description{
Create a Flat Text "extremes table" Type Output Table
}

