% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extremesTable.R
\name{extremesTable}
\alias{extremesTable}
\title{create a flat text 'extremes table' type output table}
\usage{
extremesTable(ts)
}
\arguments{
\item{ts}{a timeseries list that comes from valid extremes json}
}
\value{
string table
}
\description{
create a flat text 'extremes table' type output table
}

