% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchQualifierMetadata}
\alias{fetchQualifierMetadata}
\title{Fetch Qualifier Metadata (TSS)}
\usage{
fetchQualifierMetadata(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}
}
\description{
Given a full report object this will extract the qualifier metadata
}
