% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R
\name{fetchWaterQualityMeasurements}
\alias{fetchWaterQualityMeasurements}
\title{Fetch water quality measurements}
\usage{
fetchWaterQualityMeasurements(reportObject)
}
\arguments{
\item{reportObject}{the full report data}
}
\description{
Given a report object, will pull the water quality measurements
}

