% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{findZeroNegativeGaps}
\alias{findZeroNegativeGaps}
\title{Add periods of zero or negative data to the gaps field of the specified time
series.}
\usage{
findZeroNegativeGaps(field, data, isDV)
}
\arguments{
\item{field}{A field name.}

\item{data}{Time series data structure, as list of fields.}

\item{isDV}{Context is daily values when TRUE; non-daily-values otherwise.}
}
\description{
Add periods of zero or negative data to the gaps field of the specified time
series.
}

