% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{findZeroNegativeGaps}
\alias{findZeroNegativeGaps}
\title{Add Periods of Zero or Negative Data to the Gaps Field of the Specified Time
Series}
\usage{
findZeroNegativeGaps(field, data, isDV)
}
\arguments{
\item{field}{A field name.}

\item{data}{Full report data structure.}

\item{isDV}{Context is daily values when \code{TRUE}; not-daily-values
otherwise.}
}
\description{
Add Periods of Zero or Negative Data to the Gaps Field of the Specified Time
Series
}
\author{
Zack Moore

Andrew Halper
}

