% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{findZeroNegativeGaps}
\alias{findZeroNegativeGaps}
\title{Identify the start and end dates for gaps due to zeroes and negatives}
\usage{
findZeroNegativeGaps(timeValueDF, flagZeroNeg, isVolumetricFlow, isDV)
}
\arguments{
\item{timeValueDF}{data.frame to split with at least two columns, value and time}

\item{flagZeroNeg}{logical indicating whether or not the zeros & negatives can be 
removed for logging the axis}

\item{isVolumetricFlow}{logical indicating whether the values represent volumetric 
flow or not (e.g. FALSE could indicate water level data)}

\item{isDV}{logical saying whether or not the time series is made of daily values}
}
\description{
Finds periods of zero or negative data to the gaps field of 
the specified time series. Only identifies gaps for zeroes and negatives if 
they are going to be removed from the time series for logging (flagZeroNeg and
isLogged are TRUE). Returns a data frame with two columns, startTime and endTime.
}

