% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sig-fiveyeargwsum.R
\docType{methods}
\name{fiveyeargwsum}
\alias{fiveyeargwsum}
\alias{fiveyeargwsum,character,character-method}
\alias{fiveyeargwsum,list,character-method}
\alias{fiveyeargwsum,list,missing-method}
\title{Five-year Groundwater Summary report}
\usage{
fiveyeargwsum(data, output, ...)

\S4method{fiveyeargwsum}{list,character}(data, output, ...)

\S4method{fiveyeargwsum}{character,character}(data, output, ...)

\S4method{fiveyeargwsum}{list,missing}(data, output, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{additional params passed to GET or authenticateUser}
}
\description{
Five-year Groundwater Summary report
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
data <- fromJSON(system.file('extdata','fiveyeargwsum-example.json', package = 'repgen'))
fiveyeargwsum(data, 'html', 'Author Name')
}

