% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-fiveyeargwsum.R
\docType{methods}
\name{fiveyeargwsum}
\alias{fiveyeargwsum}
\alias{fiveyeargwsum,character-method}
\alias{fiveyeargwsum,list-method}
\title{Five year groundwater summary report.}
\usage{
fiveyeargwsum(data, ...)

\S4method{fiveyeargwsum}{list}(data, ...)

\S4method{fiveyeargwsum}{character}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{Additional params passed to GET or \code{authenticateUser}.}
}
\description{
Five year groundwater summary report.
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)

data <-
 fromJSON(
   system.file(
     'extdata', 'fiveyeargwsum', 'fiveyeargwsum-example.json', package = 'repgen'
   )
 )
fiveyeargwsum(data, 'Author Name')
}

