% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-fiveyeargwsum.R
\docType{methods}
\name{fiveyeargwsum}
\alias{fiveyeargwsum}
\alias{fiveyeargwsum,list-method}
\title{Five-year Groundwater Summary report}
\usage{
fiveyeargwsum(data, ...)

\S4method{fiveyeargwsum}{list}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
Sys.setenv(TZ = "UTC")
data <- fromJSON(system.file('extdata','fiveyeargwsum','fiveyeargwsum-example.json', package = 'repgen'))
fiveyeargwsum(data, 'Author Name')

}

