% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{flexibleTimeParse}
\alias{flexibleTimeParse}
\title{will attempt to parse a DV, UTC time, or offset time}
\usage{
flexibleTimeParse(x, timezone, shiftTimeToNoon = TRUE)
}
\arguments{
\item{x}{character vector of the date/time}

\item{timezone}{character vector of length one indicating a timezone code}
}
\value{
time vector
}
\description{
convienence that will attempt to parse a DV, UTC time, or offset time
extremes json
}

