% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{flexibleTimeParse}
\alias{flexibleTimeParse}
\title{Will attempt to parse a DV, UTC time, or offset time.}
\usage{
flexibleTimeParse(x, timezone, shiftTimeToNoon = TRUE)
}
\arguments{
\item{x}{A vector of character vectors of the dates/times.}

\item{timezone}{A character vector of length one, indicating a time zone code.}

\item{shiftTimeToNoon}{Reference time to 12:00 p.m. if TRUE; interpret
literally when FALSE.}
}
\value{
time vector
}
\description{
A convienence function that will attempt to parse a (day
  point-type) date, UTC time, or offset time extremes JSON.
}
