% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{flexibleTimeParse}
\alias{flexibleTimeParse}
\title{will attempt to parse a DV, UTC time, or offset time}
\usage{
flexibleTimeParse(x, timezone, keepTime = TRUE)
}
\arguments{
\item{x}{character vector of the date/time}

\item{timezone}{character vector of length one indicating a timezone code}

\item{keepTime}{logical to indicate whether the final value should include HH:MM:SS or just be a date}
}
\value{
time vector
}
\description{
convienence that will attempt to parse a DV, UTC time, or offset time
extremes json
}

