% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{formatTimeSeriesForPlotting}
\alias{formatTimeSeriesForPlotting}
\title{Format time series for plotting}
\usage{
formatTimeSeriesForPlotting(series, removeZeroNegativeFlag = NULL)
}
\arguments{
\item{series}{The time series data to format for plotting}

\item{removeZeroNegativeFlag}{Whether or not to remove zero and negative values}
}
\value{
A data frame representing the time series for plotting
}
\description{
Helper function that primes a time series for plotting
by extracing the points data frame from the list, adding the legend
name to that data frame, and removing zero/negative value rows if necessary
}

