% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{getApprovalBarConfig}
\alias{getApprovalBarConfig}
\title{Get Approval Bar Configuration}
\usage{
getApprovalBarConfig(approvals, ylim, ylog, reverse)
}
\arguments{
\item{approvals}{A list of data objects relevant to plotting approval
bars. Each list item must be named one of appr_approved_uv, appr_inreview_uv, 
or appr_working_uv. Any other names will result in an error.}

\item{ylim}{The \emph{y}-axis interval, as ordered pair vector.}

\item{ylog}{A Boolean truth value. \code{TRUE} indicates the \emph{y}-axis is
referenced to log10; linear \emph{y}-axis otherwise.}

\item{reverse}{A Boolean truth value. \code{TRUE} indicates the \emph{y}-axis
is inverted; not inverted otherwise.}
}
\value{
a list of configs (data + style) ready to be added to the gsplot object
 for each approval bar. Each config is a list named "rect" with appropriate \code{rect}
 arguments and any styles (likely color and border) defined.
}
\description{
Given approval bars, return a named list of gsplot elements to
  call.
}
\author{
Andrew Halper
}
