% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval-bar-styles.R, R/utils-approvals.R
\name{getApprovalBarStyle}
\alias{getApprovalBarStyle}
\title{Associate a List of Styles (and some properties) with Specified Type of
Approval Bar}
\usage{
getApprovalBarStyle(data, ybottom, ytop)

getApprovalBarStyle(data, ybottom, ytop)
}
\arguments{
\item{data}{Plot data.}

\item{ybottom}{Approval bar rectangle, vertical bottom extent.}

\item{ytop}{Approval bar rectangle, vertical top extent.}

\item{data}{Plot data.}

\item{ybottom}{Approval bar rectangle, vertical bottom extent.}

\item{ytop}{Approval bar rectangle, vertical top extent.}
}
\value{
A list of styles.

A list of styles.
}
\description{
<<<<<<< HEAD
Associate a List of Styles (and some properties) with Specified Type of
Approval Bar
=======
Associate a list of styles (and some properties) with specified type of
approval bar.

Associate a list of styles (and some properties) with specified type of
approval bar.
>>>>>>> refs/remotes/GitHub/master
}

