% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approvals.R
\name{getApprovalBarStyles}
\alias{getApprovalBarStyles}
\title{Get Approval Bar Styles}
\usage{
getApprovalBarStyles(approvals)
}
\arguments{
\item{approvals}{A list of data objects relevant to plotting approval
bars.}
}
\value{
A list of styling elements.
}
\description{
Get styling information for approval bar elements.
}
\author{
Andrew Halper
}
