% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getCorrections}
\alias{getCorrections}
\title{Extract and Restructure Corrections}
\usage{
getCorrections(ts, field)
}
\arguments{
\item{ts}{A list, containing a time series data structure.}

\item{field}{A field name.}
}
\description{
Extract and Restructure Corrections
}

