% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-render.R
\name{getDVHydrographPlotConfig}
\alias{getDVHydrographPlotConfig}
\title{Get DV Hydrograph Plot Config}
\usage{
getDVHydrographPlotConfig(plotItem, plotItemName, legendName, ...)
}
\arguments{
\item{plotItem}{the item to fetch styles and plot features for}

\item{...}{any additional parameters to pass into the function}
}
\description{
Given an item to plot, fetch the associated
plot feature(s) and their styles.
}

