% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{getEstimatedEdges}
\alias{getEstimatedEdges}
\title{Create vertical step edges between estimated and non-estimated series}
\usage{
getEstimatedEdges(stat, est)
}
\arguments{
\item{stat}{the parsed non-estimated time series}

\item{est}{the parsed estimated time series}
}
\value{
a list of vertical lines connecting steps between stat and est
}
\description{
Create vertical step edges between estimated and non-estimated series
}

