% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getFieldVisitMeasurementsShifts}
\alias{getFieldVisitMeasurementsShifts}
\title{Extract and Restructure Field Visit Measurement Shifts}
\usage{
getFieldVisitMeasurementsShifts(ts)
}
\arguments{
\item{ts}{A list, containing a time series data structure.}
}
\description{
Extract and Restructure Field Visit Measurement Shifts
}

