% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getGroundWaterLevels}
\alias{getGroundWaterLevels}
\title{Extract groundwater levels from a time series data structure.}
\usage{
getGroundWaterLevels(ts, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{...}{Optional arguments.}
}
\description{
Extract groundwater levels from a time series data structure.
}

