% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-web.R
\name{getJSON}
\alias{getJSON}
\title{Get JSON to list from an authenticated endpoint.}
\usage{
getJSON(url, ..., verbose = FALSE)
}
\arguments{
\item{url}{The JSON endpoint.}

\item{...}{Additional parameters passed to \code{checkAuth}.}

\item{verbose}{Send output to screen during GET request when TRUE; do not send
output to screen during GET request otherwise.}
}
\description{
Get JSON to list from an authenticated endpoint.
}
\keyword{internal}

