% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-web.R
\name{getJSON}
\alias{getJSON}
\title{get JSON to list from an authenticated endpoint}
\usage{
getJSON(url, ..., verbose = FALSE)
}
\arguments{
\item{url}{the json endpoint}

\item{...}{additional parameters passed to checkAuth}

\item{verbose}{output to screen during GET}
}
\keyword{internal}

