% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{getLaneYData}
\alias{getLaneYData}
\title{Get Lane Y Data}
\usage{
getLaneYData(data, height, initialHeight, overlapInfo = NULL)
}
\arguments{
\item{data}{The lane data to get Y positional data for}

\item{height}{The height to use for lane rectangles}

\item{overlapInfo}{[DEFAULT: NULL] Calculated rectangle overlap data to use}

\item{intialHeight}{The height to start calculating Y positions from}
}
\description{
Get the Y position data for rectangles for the specified lane
}

