% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getMaxStage}
\alias{getMaxStage}
\title{Extract maximum stage height from a data structure.}
\usage{
getMaxStage(ts, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{...}{Optional parameters.}
}
\description{
Extract maximum stage height from a data structure.
}

