% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getMeasuredShiftPlotConfig}
\alias{getMeasuredShiftPlotConfig}
\title{Get Measured shifts Plot Config}
\usage{
getMeasuredShiftPlotConfig(meas_shift)
}
\arguments{
\item{meas_shift}{list of measured shift objects}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a report object, some information about the plot to build, will return a named list of gsplot elements to call
}
