% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getMinStage}
\alias{getMinStage}
\title{Extract minimum stage height from a data structure.}
\usage{
getMinStage(ts, ...)
}
\arguments{
\item{ts}{A time series data structure, as list of fields.}

\item{...}{Optional parameters.}
}
\description{
Extract minimum stage height from a data structure.
}

