% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getPrimaryPlotConfig}
\alias{getPrimaryPlotConfig}
\title{Get Primary Plot Config}
\usage{
getPrimaryPlotConfig(timeseries, name, primary_lbl, reference_lbl, comp_lbl,
  dataSides, dataLimits)
}
\arguments{
\item{timeseries}{list of TS points relavant to primary plot}

\item{name}{name of item being plotted}

\item{primary_lbl}{label of primary time series}

\item{reference_lbl}{label of refernce time series}

\item{comp_lbl}{label of comparison time series}

\item{dataSides}{named list of integers describing what side of the plot correspondingly named series goes on}

\item{dataLimits}{named list of limits for name series}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a list of TS points, some information about the plot to build, will return a named list of gsplot elements to call
}

