% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getPrimaryPlotConfig}
\alias{getPrimaryPlotConfig}
\title{Get Primary Plot Config}
\usage{
getPrimaryPlotConfig(timeseries, name, label, ylim, dataSide = 0,
  comparisonOnIndependentAxes = TRUE)
}
\arguments{
\item{timeseries}{list of TS points relavant to primary plot}

\item{name}{name of item being plotted}

\item{label}{label of time series}

\item{ylim}{y range that the time series covers}

\item{dataSide}{optional for reference and comparison series, integer of what side the time series' y value is on}

\item{comparisonOnIndependentAxes}{set to false if being plotted on the same axes as another}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a list of TS points, some information about the plot to build, will return a named list of gsplot elements to call
}
