% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getPrimaryReportElements}
\alias{getPrimaryReportElements}
\title{Get Primary Report Elements}
\usage{
getPrimaryReportElements(reportObject, month, timezone, excludeZeroNegativeFlag)
}
\arguments{
\item{reportObject}{UV hydro report}

\item{month}{the month to plot for all data}

\item{timezone}{timezone to parse all data into}

\item{excludeZeroNegativeFlag}{flag to exclude ploting of values <= 0}
}
\value{
named list of report elements (plot, table status_msg)
}
\description{
contructs the gsPlot, a table of correction information, and a status message for data associated with the primary plot
}
