% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getReportMetadata}
\alias{getReportMetadata}
\title{Get a Value From Extremes JSON List}
\usage{
getReportMetadata(data, param, ...)
}
\arguments{
\item{data}{A list; can be the output of \code{\link[jsonlite]{fromJSON}}.}

\item{param}{A field name (e.g., "locationNumber").}

\item{...}{Additional arguments passed to \code{repgen:::validParam}, such as
\code{required} or \code{as.numeric}.}
}
\value{
A value or array corresponding to the field specified by
        \code{param}.
}
\description{
A convienence function for accessing from the "values" block in 
             extremes JSON.
}

