% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch.R, R/utils-json.R
\name{getReportMetadata}
\alias{getReportMetadata}
\title{get value from extremes json list}
\usage{
getReportMetadata(data, param, ...)

getReportMetadata(data, param, ...)
}
\arguments{
\item{data}{a list, can be the output of \code{\link[jsonlite]{fromJSON}}.}

\item{param}{the field name (e.g., 'locationNumber')}

\item{...}{additional arguments passed to \code{repgen:::validParam}, 
such as \code{required}, or \code{as.numeric}}

\item{data}{A list; can be the output of \code{\link[jsonlite]{fromJSON}}.}

\item{param}{A field name (e.g., "locationNumber").}

\item{...}{Additional arguments passed to \code{repgen:::validParam}, such as
\code{required} or \code{as.numeric}.}
}
\value{
a value or array corresponding to the field specified by \code{param}

A value or array corresponding to the field specified by
        \code{param}.
}
\description{
convienence function for accessing from the "values" block in 
extremes json

A convienence function for accessing from the "values" block in 
             extremes JSON.
}

