% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getSecondaryPlotConfig}
\alias{getSecondaryPlotConfig}
\title{Get Secondary Plot Config}
\usage{
getSecondaryPlotConfig(timeseries, name, legend_label, ylim, doLog = FALSE)
}
\arguments{
\item{timeseries}{the timeseries to be plotted}

\item{name}{name of style to be applied to given x/y points (corrected, estimated, or uncorrected)}

\item{legend_label}{label to be applied to points in legend}

\item{ylim}{The y-axis limits of the item to get the configuration for}

\item{doLog}{Whether or not the item should be placed on a logarithmic axis (DEFAULT: FALSE)}
}
\value{
named list of gsplot calls. The name is the plotting call to make, and it points to a list of config params for that call
}
\description{
Given a report object, some information about the plot to build, will return a named list of gsplot elements to call
}
