% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{getSecondaryPlotConfig}
\alias{getSecondaryPlotConfig}
\title{Get Secondary Plot Config}
\usage{
getSecondaryPlotConfig(secondaryPlotData, plotStartDate, plotEndDate,
  secondary_lbl, tertiary_lbl, limits)
}
\arguments{
\item{secondaryPlotData}{list of data objects relavant to secondary plot}

\item{plotStartDate}{start date of this plot}

\item{plotEndDate}{end date of this plot}

\item{secondary_lbl}{label of secondary time series}

\item{tertiary_lbl}{label of tertiary time series}

\item{limits}{list of lims for all of the data which will be on here}
}
\description{
Given a report object, some information about the plot to build, will return a named list of gsplot elements to call
}

