% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveyeargwsum-render.R
\name{getSides}
\alias{getSides}
\title{Sort data and sides}
\usage{
getSides(stat1TimeSeries, stat2TimeSeries, stat3TimeSeries, stat4TimeSeries)
}
\arguments{
\item{stat1TimeSeries}{the stat-derived time series 1 selected}

\item{stat2TimeSeries}{the stat-derived time series 2 selected}

\item{stat3TimeSeries}{the stat-derived time series 3 selected}

\item{stat4TimeSeries}{the stat-derived time series 4 selected}
}
\value{
named list with three items, a list of sides, list of series, and list of lims.)
}
\description{
Depending on the configuration, stat-derived time series will be plotted on different sides. This function compares types and figures out where to plot the data and what lims they will need to use to construct each side.
}
