% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{getSimsUrl}
\alias{getSimsUrl}
\title{Inserts the SIMS url (if it exists) into the base of the report}
\usage{
getSimsUrl(simsUrl)
}
\arguments{
\item{simsUrl}{The simsUrl requested to turn into a link}
}
\value{
the HTML link for SIMS url
}
\description{
takes the SIMS url and formats it 
for including in the report as a link
}

