% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{getSimsUrl}
\alias{getSimsUrl}
\title{Put the SIMS URL (if it exists) Into the Base of the Report}
\usage{
getSimsUrl(data)
}
\arguments{
\item{data}{Coming in to create a plot which may have SIMS info.}
}
\description{
Put the SIMS URL (if it exists) Into the Base of the Report
}

