% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{getSimsUrl}
\alias{getSimsUrl}
\title{Put the SIMS url (if it exists) into the base of the report}
\usage{
getSimsUrl(data)
}
\arguments{
\item{data}{coming in to create a plot which may have sims info}
}
\description{
Put the SIMS url (if it exists) into the base of the report
}

