% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Extract UV time series from a data structure.}
\usage{
getTimeSeries(ts, field, estimatedOnly = FALSE, shiftTimeToNoon = TRUE)
}
\arguments{
\item{ts}{A set of time series data, as list of fields.}

\item{field}{A field name.}

\item{estimatedOnly}{Extract estimated values only when TRUE.}

\item{shiftTimeToNoon}{Reference time to 12:00 p.m. when TRUE.}
}
\description{
Extract UV time series from a data structure.
}

