% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Extract and Restructure a Time Series}
\usage{
getTimeSeries(ts, field, estimatedOnly = FALSE, shiftTimeToNoon = TRUE)
}
\arguments{
\item{ts}{A list, containing a time series data structure.}

\item{field}{A field name.}

\item{estimatedOnly}{Extract only estimated values when \code{TRUE}.}

\item{shiftTimeToNoon}{Reference time to 12:00 p.m. when \code{TRUE}.}
}
\description{
Extract and Restructure a Time Series
}

