% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Extract and Restructure a Time Series}
\usage{
getTimeSeries(ts, field, estimatedOnly = FALSE, shiftTimeToNoon = TRUE)
}
\arguments{
\item{ts}{A list, containing a time series data structure.}

\item{field}{A field name.}

\item{estimatedOnly}{Extract only estimated values when \code{TRUE}.}

\item{shiftTimeToNoon}{Reference time to 12:00 p.m. when \code{TRUE}.}
}
\description{
This function is deprecated. Please switch over to using 
  \code{readTimeSeries} and \code{readEstimatedTimeSeries}. Note that
  \code{readTimeSeries} and \code{readEstimatedTimeSeries} now return a list
  of the full time series object instead of the data frame created and
  returned by this function. This means that downstream calls using this time
  series will need to be updated to pass in the correct parameters. See
  \code{inst/extdata/testsnippets/test-timeSeries.JSON} in the repgen source
  directory for example JSON outlining how a time series returned from 
  \code{readTimeSeries}/\code{readEstimatedTimeSeries} will look.
}

