% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{getWaterDataUrl}
\alias{getWaterDataUrl}
\title{Put the waterdata.usgs.gov url (if it exists) into the base of the report}
\usage{
getWaterDataUrl(waterdataUrl)
}
\arguments{
\item{waterdataUrl}{The waterdata url requested to turn into a link}
}
\value{
The HTML link for waterdata url
}
\description{
takes the waterdata url and formats it for including it
in the report as a link
}
