% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{getWaterDataUrl}
\alias{getWaterDataUrl}
\title{Put the waterdata.usgs.gov URL (if it exists) Into the Base of the Report}
\usage{
getWaterDataUrl(data)
}
\arguments{
\item{data}{Coming in to create a plot which may have waterdata info.}
}
\description{
Put the waterdata.usgs.gov URL (if it exists) Into the Base of the Report
}

