% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{getWaterQualityMeasurements}
\alias{getWaterQualityMeasurements}
\title{Extract and Restructure Water Quality Measurements}
\usage{
getWaterQualityMeasurements(ts, ...)
}
\arguments{
\item{ts}{A list, containing a time series data structure.}

\item{...}{Unknown. Possibly obsolete.}
}
\description{
Extract and Restructure Water Quality Measurements
}

