% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{isEmptyOrBlank}
\alias{isEmptyOrBlank}
\title{Indicate whether a value is NULL, NA, or the empty string; or whether a named
object exists in a list of objects.}
\usage{
isEmptyOrBlank(val = NULL, listObjects = NULL, objectName = NULL)
}
\arguments{
\item{val}{A value.}

\item{listObjects}{A list of objects; can be empty.}

\item{objectName}{An object name.}
}
\description{
Indicate whether a value is NULL, NA, or the empty string; or whether a named
object exists in a list of objects.
}

