% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{isEmptyVar}
\alias{isEmptyVar}
\title{Determine whether a variable is (semantically) empty.}
\usage{
isEmptyVar(variable)
}
\arguments{
\item{variable}{A variable's value.}
}
\description{
Determine whether a variable is (semantically) empty.
}

