% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isEmptyVar}
\alias{isEmptyVar}
\title{isEmptyVar}
\usage{
isEmptyVar(variable)
}
\arguments{
\item{variable}{R object, matrix, vector, array or data frame.}
}
\description{
Check if a variable is NULL or has no rows.
}
