% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isEmptyVar}
\alias{isEmptyVar}
\title{Check Usability of a Variable}
\usage{
isEmptyVar(variable)
}
\arguments{
\item{variable}{R object, matrix, vector, array or data frame.}
}
\description{
Check if a variable is \code{NULL}, has no rows, or has a
  \code{time} field that is \code{NULL} or the empty string.
}

