% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{isLogged}
\alias{isLogged}
\title{Test Whether a Time Series Should be Plotted vs. a log10, y-axis}
\usage{
isLogged(all_data, ts_data, series)
}
\arguments{
\item{all_data}{A report data structure.}

\item{ts_data}{A time series data structure.}

\item{series}{Property name of a time series within \code{all_data}.}
}
\description{
Test Whether a Time Series Should be Plotted vs. a log10, y-axis
}

