% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{isLogged}
\alias{isLogged}
\title{isLogged}
\usage{
isLogged(ts_data, isVolFlow, excludeZeroNegative)
}
\arguments{
\item{ts_data}{The time series data itself. If this is \code{NULL} but the other
params are \code{TRUE}, it returns \code{TRUE}.}

\item{isVolFlow}{A logical value indicating whether the time series measures 
a volumetric flow or not.}

\item{excludeZeroNegative}{A logical value indicating whether the time series
will exclude zero and negative values.}
}
\value{
Returns \code{TRUE} if the time series can be plotted on a logged
  axis, and \code{FALSE} otherwise.
}
\description{
To check if a time series can be plotted on a logged axis.
}

