% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{isLogged}
\alias{isLogged}
\title{isLogged}
\usage{
isLogged(ts_data, isVolFlow, excludeZeroNegative)
}
\arguments{
\item{ts_data}{The timeseries data itself. If this is null but the other params are true, it returns true.}

\item{isVolFlow}{A logical value indicating whether the timeseries measures
a volumetric flow or not.}

\item{excludeZeroNegative}{A logical value indicating whether the timeseries
will exclude zero and negative values.}
}
\value{
Return true if the timeseries can be plotted on a logged axis and false otherwise.
}
\description{
To check if a timeseries can be plotted on a logged axis.
}

