% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{isNullOrFalse}
\alias{isNullOrFalse}
\title{Check for NULL or FALSE values}
\usage{
isNullOrFalse(variable)
}
\arguments{
\item{variable}{R object.}
}
\description{
Check if a value is NULL or has a value of FALSE.
}

