% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{isTimeSeriesInverted}
\alias{isTimeSeriesInverted}
\title{isTimeSeriesInverted}
\usage{
isTimeSeriesInverted(timeseries)
}
\arguments{
\item{timeseries}{The timseries to potentially be inverted.}
}
\value{
Returns a logical value that is true if the timeseries should be inverted.
}
\description{
Checks if a time series is to be plotted on a revers \emph{y}-axis or not.
}
