% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataFlags.R
\name{negValues}
\alias{negValues}
\title{negValues}
\usage{
negValues(timeSeriesData)
}
\arguments{
\item{timeSeriesData}{The time series that is checked for negative data}
}
\value{
A logical value that is true if there are any negative values.
}
\description{
Function that returns negative data
}
