% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrApprovals}
\alias{parseCorrApprovals}
\title{Parse CORR Approvals}
\usage{
parseCorrApprovals(timeSeries, timezone, dateSeq)
}
\description{
Reads and formats the approvals for use on the CORR report
}

