% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parseCorrectionsByMonth}
\alias{parseCorrectionsByMonth}
\title{Parse Corrections}
\usage{
parseCorrectionsByMonth(reportObject, seriesName, month)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}
}
\value{
corrections subset by month
}
\description{
Read corrections for a given series
}

