% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrProcessingCorrections}
\alias{parseCorrProcessingCorrections}
\title{Parse CORR Processing getCorrectionsLabels}
\usage{
parseCorrProcessingCorrections(reportObject, processOrder, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{processOrder}{The processing order to fetch data for}

\item{timezone}{The timezone to parse data into}
}
\description{
Returns the processing order corrections for the specified processing order.
These lanes should still show up even if they are empty, so this will always return a list
even if it has no data, unlike the other parse functions for lane data.
}

