% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionsataglance-data.R
\name{parseCorrQualifiers}
\alias{parseCorrQualifiers}
\title{Parse CORR Qualifiers}
\usage{
parseCorrQualifiers(timeSeries, timezone)
}
\arguments{
\item{timeSeries}{The time series to get qualifiers for}

\item{timezone}{The timezone to parse data into}
}
\description{
Retrieves and formats qualifiers for the CORR report
}
