% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-templating-class.R,
%   R/derivationchain-data.R, R/examplereport-data.R, R/timeseriessummary-data.R
\docType{methods}
\name{parseCustomDataElementsForTemplate}
\alias{parseCustomDataElementsForTemplate}
\alias{parseCustomDataElementsForTemplate,derivationchain-method}
\alias{parseCustomDataElementsForTemplate,examplereport-method}
\alias{parseCustomDataElementsForTemplate,timeseriessummary-method}
\title{parseCustomDataElementsForTemplate}
\usage{
parseCustomDataElementsForTemplate(reportData)

\S4method{parseCustomDataElementsForTemplate}{derivationchain}(reportData)

\S4method{parseCustomDataElementsForTemplate}{examplereport}(reportData)


  \S4method{parseCustomDataElementsForTemplate}{timeseriessummary}(reportData)
}
\arguments{
\item{reportData}{typed report json, the type should be the name of the report}
}
\value{
list of named data elements that should be available to your templates. Data will be placed in "reportData"
}
\description{
call custom data parsing, this will allow reports to specify the data structure they want available to the templates
}
