% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-render.R
\name{parseCustomDataElementsForTemplate}
\alias{parseCustomDataElementsForTemplate}
\title{parseCustomDataElementsForTemplate}
\usage{
parseCustomDataElementsForTemplate(reportData)
}
\arguments{
\item{reportData}{typed report json, the type should be the name of the report}
}
\value{
list of named data elements that should be available to your templates. Data will be placed in "reportData"
}
\description{
call custom data parsing, this will allow reports to specify the data structure they want available to the templates
}
