% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvhydrograph-data.R
\name{parseDVTimeSeries}
\alias{parseDVTimeSeries}
\title{Parse DV Time Series}
\usage{
parseDVTimeSeries(reportObject, seriesField, descriptionField, timezone,
  excludeZeroNegativeFlag, estimated = FALSE)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{seriesField}{the JSON field name for the TS data}

\item{descriptionField}{the JSON field name for the TS legend name}

\item{timezone}{The timezone to parse the TS points into}

\item{excludeZeroNegativeFlag}{whether or not to exclude zero and negative values}

\item{estimated}{whether or not the retrieved time series should be estimated or non-estimated}
}
\value{
The requested time series or NULL if the request time series was not found.
}
\description{
A wrapper for the readTimeSeries functions that handles
errors thrown by those functions if the specified time series is
not found and throws a warning message. Also handles time series that
are returned without any point data and treats them as NULL.
}

