% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseFieldVisitMeasurements}
\alias{parseFieldVisitMeasurements}
\title{Parse Field Visit Measurements}
\usage{
parseFieldVisitMeasurements(reportObject, excludeZeroNegativeFlag)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{excludeZeroNegativeFlag}{whether or not zero/negative values are included}
}
\description{
Given the full report JSON object, reads the field
visit measurements and handles read errors.
}
