% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{parsePrimarySeriesList}
\alias{parsePrimarySeriesList}
\title{Parse Primary Series List}
\usage{
parsePrimarySeriesList(reportObject, month, timezone)
}
\arguments{
\item{reportObject}{entire UV Hydro report object}

\item{month}{subset only into this month}

\item{timezone}{timezone to parse all data into}
}
\value{
named list of timeseries objects (NULL if not in report object) as well as inverted and loggedAxis flags. loggedAxis is set so that all series are supported on the same axis.
}
\description{
Read all series including reference or comparison series
}
