% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parsePrimarySeriesQualifiers}
\alias{parsePrimarySeriesQualifiers}
\title{Parse Primary Series Qualifiers}
\usage{
parsePrimarySeriesQualifiers(reportObject, filterCode = NULL)
}
\arguments{
\item{reportObject}{the full report JSON object}

\item{filterCode}{The code to filter read qualifiers to}
}
\description{
Default wrapper for the readPrimarySeriesQualifiers function
that handles errors thrown and returns the proper data.
}
