% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.r
\name{parseRatingShiftsData}
\alias{parseRatingShiftsData}
\title{Parse Rating Shifts Data}
\usage{
parseRatingShiftsData(reportObject)
}
\arguments{
\item{reportObject}{An R object with the raw data required for rating shifts}
}
\value{
A list containing rating shift information
}
\description{
Takes in a report object and returns the rating shift information
}
