% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSApprovals}
\alias{parseTSSApprovals}
\title{Parse Approvals}
\usage{
parseTSSApprovals(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone to parse data into}
}
\description{
Default wrapper for the readApprovals function
that handles errors thrown and returns the proper data
}
