% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSGaps}
\alias{parseTSSGaps}
\title{Parse Gaps}
\usage{
parseTSSGaps(reportData, timezone, isStatDerived)
}
\arguments{
\item{reportData}{The full report JSON object}

\item{timezone}{The timezone to parse data into}

\item{isStatDerived}{boolean if the ts is stat-derived}
}
\description{
TSS wrapper for the readGaps function
that handles errors thrown and returns the proper data
}
