% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessummary-data.R
\name{parseTSSProcessingCorrections}
\alias{parseTSSProcessingCorrections}
\title{Parse Processing Corrections}
\usage{
parseTSSProcessingCorrections(reportObject, processOrder, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{processOrder}{The processing order to fetch data for}

\item{timezone}{The timezone to parse data into}
}
\description{
Default wrapper for the readProcessingCorrections function
that handles errors thrown and returns the proper data
}
