% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plotItem}
\alias{plotItem}
\title{Plot Item}
\usage{
plotItem(plot_object, item, name, configFunction, yLabel = "", isDV = FALSE)
}
\arguments{
\item{plot_object}{The GS Plot object to plot onto}

\item{item}{The item to plot}

\item{name}{The variable name to use for the item (used for style and config matching)}

\item{configFunction}{The function to use for fetching the style and config data for this TS}

\item{yLabel}{The label to put onto the Y-Axis for this item (default: "")}

\item{isDV}{Whether or not the plot is a daily value plot (defulat: FALSE)}
}
\description{
Function that takes a GS plot object and the necessary item data and
then formats the item properly for plotting and plots it.
}

