% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{PlotUVHydrographObject}
\alias{PlotUVHydrographObject}
\title{Add UV hydrograph objects to a plot object.}
\usage{
PlotUVHydrographObject(object, plotConfig)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{list}{of gsplot calls to make}
}
\value{
A modified gsplot, plot object, with UV hydrograph objects included.
}
\description{
Add UV hydrograph objects to a plot object.
}

