% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{printReportFeature}
\alias{printReportFeature}
\title{Organize and Fit, Plots and Tables in UV Hydrograph Report}
\usage{
printReportFeature(feature, isTable = FALSE, m = NULL, mar_values = c(8,
  3, 4, 3))
}
\arguments{
\item{feature}{the plot or table to include}

\item{isTable}{\code{TRUE} or \code{FALSE} if the feature is a table;
defaults to \code{FALSE}.}

\item{m}{The month name for the feature.}

\item{mar_values}{The margin values for the feature to ensure space for plot
and legend.}
}
\value{
The formatted feature requested.
}
\description{
sets margins and lays out report in an attractive way for UV Hydrograph
}

