% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readApprovalRanges}
\alias{readApprovalRanges}
\title{Get Approval Ranges}
\usage{
readApprovalRanges(ts, approval, timezone)
}
\arguments{
\item{ts}{the timeseries object}

\item{approval}{the approval level to read, typically "Working", "In Review", or "Approved"}

\item{timezone}{the timezone to parse times to}
}
\value{
data frame of start and end times for each approval range
}
\description{
Get Approval Ranges
}

