% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readDownchainSeries}
\alias{readDownchainSeries}
\title{Read Downchain Series (TSS)}
\usage{
readDownchainSeries(reportObject)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone of the report}
}
\description{
Reads and formats the related downchain series
}
