% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readRatingShifts}
\alias{readRatingShifts}
<<<<<<< HEAD
\title{Read Rating Shifts}
\usage{
readRatingShifts(reportObject)
}
\arguments{
\item{reportObject}{the object representing the full report JSON}
}
\value{
data frame of rating shift information
}
\description{
given a list of rating shifts returns shift data
=======
\title{Read Rating Shifts (TSS)}
\usage{
readRatingShifts(reportObject, timezone)
}
\arguments{
\item{reportObject}{The full report JSON object}

\item{timezone}{The timezone of the report}
}
\description{
Reads and formats the rating shifts
>>>>>>> upstream/master
}
