% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{readTimeSeries}
\alias{readTimeSeries}
\title{Read time series}
\usage{
readTimeSeries(reportObject, seriesName, timezone, descriptionField = NULL,
  shiftTimeToNoon = FALSE, isDV = FALSE, estimated = FALSE,
  requiredFields = NULL)
}
\arguments{
\item{reportObject}{the full JSON report object}

\item{seriesName}{the name of the time series to extract}

\item{timezone}{the timezone to parse times to}

\item{shiftTimeToNoon}{[DEFAULT: FALSE] whether or not to shift DV times to noon}

\item{isDV}{whether or not the specified time series is a daily value time series}

\item{requiredFields}{optional overriding of required fields for a time series}
}
\description{
Reads and formats a time series from the provided full report object
}

