% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{removeZeroNegative}
\alias{removeZeroNegative}
\title{Remove zero or negative values}
\usage{
removeZeroNegative(df, column = "value")
}
\arguments{
\item{df}{any data frame that it will check for negative or zeroes}

\item{column}{(Optional) [DEFAULT: 'value'] The column to look for zero/negative values in}
}
\description{
If negative or zero values are passed into function, turn them into NA 
so we can log the plot
}
