% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{removeZeroNegative}
\alias{removeZeroNegative}
\title{Remove zero or negative values}
\usage{
removeZeroNegative(df)
}
\arguments{
\item{df}{any data frame with a column called value that it will check for negative or zeroes}
}
\description{
If negative or zero values are passed into function, turn them into NA 
so we can log the plot
}
\details{
make sure you have the column called value included in data frame
}

