% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{removeZeroNegative}
\alias{removeZeroNegative}
\title{User specified option to treat negative/zero values as NA in order to have
the plot logged.}
\usage{
removeZeroNegative(df)
}
\arguments{
\item{df}{A time series, as a sequence of ordinal (time,value) pairs.}
}
\description{
User specified option to treat negative/zero values as NA in order to have
the plot logged.
}

