% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-templating-class.R,
%   R/derivationchain-render.R, R/examplereport-render.R,
%   R/timeseriessummary-render.R
\docType{methods}
\name{renderCustomFragments}
\alias{renderCustomFragments}
\alias{renderCustomFragments,derivationchain-method}
\alias{renderCustomFragments}
\alias{renderCustomFragments,examplereport-method}
\alias{renderCustomFragments}
\alias{renderCustomFragments,timeseriessummary-method}
\alias{renderCustomFragments}
\title{renderCustomFragments}
\usage{
renderCustomFragments(reportData)

\S4method{renderCustomFragments}{derivationchain}(reportData)

\S4method{renderCustomFragments}{examplereport}(reportData)

\S4method{renderCustomFragments}{timeseriessummary}(reportData)
}
\arguments{
\item{reportData}{typed report json, the type should be the name of the report}
}
\value{
vector of rendered HTML fragments that will be included in the data available to the final report template. The fragments will be
available to the templates as renderedFragments.html
}
\description{
Peform the rendering of subfragments for a given report, should return a list of html fragments to 
be available to inject into mustach templates
}
