% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-sensorreading.R
\docType{methods}
\name{sensorreadingsummary}
\alias{sensorreadingsummary}
\alias{sensorreadingsummary,character-method}
\alias{sensorreadingsummary,list-method}
\title{Sensor reading summary report.}
\usage{
sensorreadingsummary(data, ...)

\S4method{sensorreadingsummary}{list}(data, ...)

\S4method{sensorreadingsummary}{character}(data)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{everything else}
}
\description{
Sensor reading summary report.
}
\examples{
library(jsonlite)

data <-
 fromJSON(
   system.file(
     'extdata', 'sensorreadingsummary', "sensorReadingSummary-example.json", package = 'repgen'
   )
 )
sensorreadingsummary(data, 'Author Name')
}

