% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sig-sensorreading.R
\docType{methods}
\name{sensorreadingsummary}
\alias{sensorreadingsummary}
\alias{sensorreadingsummary,character,character-method}
\alias{sensorreadingsummary,list,character-method}
\alias{sensorreadingsummary,list,missing-method}
\title{sensorreadingsummary report}
\usage{
sensorreadingsummary(data, output, ...)

\S4method{sensorreadingsummary}{list,character}(data, output, ...)

\S4method{sensorreadingsummary}{character,character}(data, output)

\S4method{sensorreadingsummary}{list,missing}(data, output)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{everything else}
}
\description{
sensorreadingsummary report
}
\examples{
library(jsonlite)
data <- fromJSON(system.file('extdata',"sensorReadingSummary-example.json",package = 'repgen'))
sensorreadingsummary(data, 'pdf', 'Author Name')
sensorreadingsummary(data, 'html', 'Author Name')
}

