% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{sensorreadingTable}
\alias{sensorreadingTable}
\title{Create a flat text, "sensorreading table" type output table.}
\usage{
sensorreadingTable(data)
}
\arguments{
\item{data}{Sensor reading report JSON string.}
}
\value{
data.frame table
}
\description{
Create a flat text, "sensorreading table" type output table.
}

