% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{sensorreadingTable}
\alias{sensorreadingTable}
\title{Create a Flat Text, "sensorreading table" Type Output Table}
\usage{
sensorreadingTable(reportData)
}
\arguments{
\item{reportData}{A sensorreading report JSON string.}
}
\value{
a table of data suitable for including in the html report
}
\description{
Takes a JSON string and extracts and formats readings for the report
}

