% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{sensorreadingTable}
\alias{sensorreadingTable}
\title{create a flat text 'sensorreading table' type output table}
\usage{
sensorreadingTable(data)
}
\arguments{
\item{data}{sensorreading report json string}
}
\value{
data.frame table
}

