% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{sensorreadingTable}
\alias{sensorreadingTable}
\title{Create a Flat Text, "sensorreading table" Type Output Table}
\usage{
sensorreadingTable(data)
}
\arguments{
\item{data}{Sensor reading report JSON string.}
}
\value{
data.frame table
}
\description{
Create a Flat Text, "sensorreading table" Type Output Table
}

