% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-auth.R
\name{setBaseURL}
\alias{setBaseURL}
\title{Set AQCU endpoint}
\usage{
setBaseURL(endpoint = "prod")
}
\arguments{
\item{endpoint}{Indicate which AQCU endpoint
you want to use options: \code{c('prod','qa','dev')}}
}
\description{
Sets the internal URLS used to either the production, QA, or dev server.
URLS are stored internally to the package
}
\examples{
\dontrun{
setBaseURL('prod')
setBaseURL('qa')
setBaseURL('dev')
}
}
\author{
Luke Winslow, Jordan S Read
}

