% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-sitevisitpeak.R
\docType{methods}
\name{sitevisitpeak}
\alias{sitevisitpeak}
\alias{sitevisitpeak,character-method}
\alias{sitevisitpeak,list-method}
\title{sitevisitpeak report}
\usage{
sitevisitpeak(data, ...)

\S4method{sitevisitpeak}{list}(data, ...)

\S4method{sitevisitpeak}{character}(data)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{everythingn else}
}
\examples{
library(jsonlite)

 data <-
  fromJSON(
    system.file(
      'extdata', 'sitevisitpeak', "sitevisitpeak-example.json", package = 'repgen'
    )
  )
sitevisitpeak(data, 'Author Name')
}

