% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-sitevisitpeak.R
\docType{methods}
\name{sitevisitpeak}
\alias{sitevisitpeak}
\alias{sitevisitpeak,list-method}
\alias{sitevisitpeak}
\title{Site Visit Peak Report}
\usage{
sitevisitpeak(data, ...)

\S4method{sitevisitpeak}{list}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{everythingn else}
}
\description{
Site Visit Peak Report
}
\examples{
library(jsonlite)
library(dplyr)

 data <-
  fromJSON(
    system.file(
      'extdata', 'sitevisitpeak', "sitevisitpeak-example.json", package = 'repgen'
    )
  )
sitevisitpeak(data, 'Author Name')
}
