% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sig-sitevisitpeak.R
\docType{methods}
\name{sitevisitpeak}
\alias{sitevisitpeak}
\alias{sitevisitpeak,character,character-method}
\alias{sitevisitpeak,list,character-method}
\alias{sitevisitpeak,list,missing-method}
\title{sitevisitpeak report}
\usage{
sitevisitpeak(data, output, ...)

\S4method{sitevisitpeak}{list,character}(data, output, ...)

\S4method{sitevisitpeak}{character,character}(data, output)

\S4method{sitevisitpeak}{list,missing}(data, output)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}
}
\description{
sitevisitpeak report
}
\examples{
library(jsonlite)
data <- fromJSON(system.file('extdata',"sitevisitpeak-example.json",package = 'repgen'))
sitevisitpeak(data, 'pdf', 'Author Name')
sitevisitpeak(data, 'html', 'Author Name')
}

