% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{sitevisitpeakTable}
\alias{sitevisitpeakTable}
\title{Create a Flat Text, "sitevisitpeak table" Type Output Table}
\usage{
sitevisitpeakTable(data)
}
\arguments{
\item{data}{Site visit peak report JSON string.}
}
\value{
data.frame table
}
\description{
Create a Flat Text, "sitevisitpeak table" Type Output Table
}

