% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{sitevisitpeakTable}
\alias{sitevisitpeakTable}
\title{create a flat text 'sitevisitpeak table' type output table}
\usage{
sitevisitpeakTable(data)
}
\arguments{
\item{data}{sitevisitpeak report json string}
}
\value{
data.frame table
}

