% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{sitevisitpeakTable}
\alias{sitevisitpeakTable}
\title{Create a Flat Text, "sitevisitpeak table" Type Output Table}
\usage{
sitevisitpeakTable(readingsData, excludeComments)
}
\arguments{
\item{readingsData}{data frame of parsed field visit readings data.}

\item{excludeComments}{boolean from report metadata of whether comments are included.}
}
\value{
data frame of site visit peak data
}
\description{
Create a Flat Text, "sitevisitpeak table" Type Output Table
}

