% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{sitevisitpeakTable}
\alias{sitevisitpeakTable}
\title{create a flat text 'sitevisitpeak table' type output table}
\usage{
sitevisitpeakTable(rawData)
}
\arguments{
\item{rawData}{sitevisitpeak report json string}

\item{rawData}{sitevisitpeak report json string}
}
\value{
data.frame table

string table
}
\description{
create a flat text 'sitevisitpeak table' type output table
}

