% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{splitDataGaps}
\alias{splitDataGaps}
\title{if there are gaps in the timeseries, don't connect them
this creates multiple line/point calls if there are gaps}
\usage{
splitDataGaps(data, ts)
}
\description{
if there are gaps in the timeseries, don't connect them
this creates multiple line/point calls if there are gaps
}

