% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gaps.R
\name{splitDataGaps}
\alias{splitDataGaps}
\title{If There are Gaps in the Time Series, Don't Connect Them}
\usage{
splitDataGaps(data, ts, isDV)
}
\arguments{
\item{data}{Full report data structure.}

\item{ts}{Current time series data.}

\item{isDV}{Logic for whether this plot uses daily values or not.}
}
\description{
This creates multiple line/point calls if there are gaps.
}

