% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{srsQualifiersTable}
\alias{srsQualifiersTable}
\title{Create a table of qualifiers that are used in the report and prepare
them in a table for use in the bottom of the report}
\usage{
srsQualifiersTable(reportObject, table)
}
\arguments{
\item{reportObject}{The JSON data for the report requested}

\item{table}{A vector from which to derive qualifiers from that
were displayed in the report itself.}
}
\value{
A table to print at the bottom of the report or an empty
data.frame if there are no qualifiers
}
\description{
Create a table of qualifiers that are used in the report and prepare
them in a table for use in the bottom of the report
}

