% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{srsQualifiersTable}
\alias{srsQualifiersTable}
\title{create flat text 'qualifiers table' type output table}
\usage{
srsQualifiersTable(data, table)
}
\arguments{
\item{data}{report data}
}
\value{
string table
}

