% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorreading-data.R
\name{srsQualifiersTable}
\alias{srsQualifiersTable}
\title{Create Flat Text, "qualifiers table" Type Output Table}
\usage{
srsQualifiersTable(reportData, table)
}
\arguments{
\item{table}{A vector from which to derive qualifiers from.}

\item{data}{Report data.}
}
\value{
A vector of strings.
}
\description{
Create Flat Text, "qualifiers table" Type Output Table
}

