% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{startRender}
\alias{startRender}
\title{Starting point, creates RMD and runs rendering}
\usage{
startRender(data, author, reportName)
}
\arguments{
\item{data}{coming in to create a plot}

\item{author}{name of person generating the report}

\item{reportName}{name of report being generated}
}
\value{
the html file and other report bits
}
\description{
this function orchestrates the creation of the html reports
}

