% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shared.R
\name{startRender}
\alias{startRender}
\title{Create R Markdown and Run Rendering}
\usage{
startRender(data, author, reportName)
}
\arguments{
\item{data}{Input data to create a plot.}

\item{author}{The name of person generating the report.}

\item{reportName}{The name of report being generated (current options:
"dvhydrograph", "fiveyruvhydrograph", "vdiagram").}
}
\description{
Create R Markdown and Run Rendering
}

