% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{svpQualifiersTable}
\alias{svpQualifiersTable}
\title{create flat text 'qualifiers table' type output table}
\usage{
svpQualifiersTable(data, table)
}
\arguments{
\item{data}{report data}
}
\value{
string table
}

