% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitevisitpeak-data.R
\name{svpQualifiersTable}
\alias{svpQualifiersTable}
\title{Create Flat Text, "qualifiers table" Type Output Table}
\usage{
svpQualifiersTable(data, table)
}
\arguments{
\item{data}{Report data.}

\item{table}{A vector from which to derive qualifiers from.}
}
\value{
A vector of strings.
}
\description{
Create Flat Text, "qualifiers table" Type Output Table
}

