% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{timeFormatting}
\alias{timeFormatting}
\title{timeFormatting}
\usage{
timeFormatting(timeVals, dateFormatMask, splitChar = "[T]")
}
\arguments{
\item{timeVals}{String with format of "YYYY-MM-DDTHH:MM:SS.SSS-UTC offset".}

\item{dateFormatMask}{String with preferred output date format}

\item{splitChar}{String with the character the function should split the date and time, defaults to "[T]"}
}
\value{
list with date in first position, time in second position.
}
\description{
Formats date to passed-in format mask, and time to "(UTC [offset] )"
}
