% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{toStartOfMonth}
\alias{toStartOfMonth}
\title{Given a datetime, will change the date to the 1st, plus remove time and set
00:00. This is used to set the labels for Five Year Groundwater Summary
reports.}
\usage{
toStartOfMonth(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
Given a datetime, will change the date to the 1st, plus remove time and set
00:00. This is used to set the labels for Five Year Groundwater Summary
reports.
}
