% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{toStartOfDay}
\alias{toEndOfDay}
\alias{toStartOfDay}
\title{given a datetime, will remove time and set 0000 as start}
\usage{
toStartOfDay(time)

toEndOfDay(time)
}
\arguments{
\item{time}{time to shift to start}

\item{time}{time to shift to start}
}
\description{
given a datetime, will remove time and set 0000 as start

given a datetime, will remove time and set 2359
}

