% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{toEndOfDay}
\alias{toEndOfDay}
\title{Given a datetime, will remove time and set 23:59.}
\usage{
toEndOfDay(time)
}
\arguments{
\item{time}{datetime to shift to start}
}
\description{
Given a datetime, will remove time and set 23:59.
}
