% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{toEndOfMonth}
\alias{toEndOfMonth}
\title{Given a datetime, will change the date to the 30th, remove time, and set 
23:59. This is used to set the labels for Five Year Groundwater Summary 
reports.}
\usage{
toEndOfMonth(time)
}
\arguments{
\item{time}{A datetime to shift to start.}
}
\description{
Given a datetime, will change the date to the 30th, remove time, and set 
23:59. This is used to set the labels for Five Year Groundwater Summary 
reports.
}

