% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-uvhydrograph.R
\docType{methods}
\name{uvhydrograph}
\alias{uvhydrograph}
\alias{uvhydrograph,list-method}
\title{UV Hydrograph report}
\usage{
uvhydrograph(data, ...)

\S4method{uvhydrograph}{list}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET}
}

