% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-uvhydrograph.R
\docType{methods}
\name{uvhydrograph}
\alias{uvhydrograph}
\alias{uvhydrograph,character-method}
\alias{uvhydrograph,list-method}
\title{UV Hydrograph report}
\usage{
uvhydrograph(data, ...)

\S4method{uvhydrograph}{list}(data, ...)

\S4method{uvhydrograph}{character}(data, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{...}{additional params passed to GET or authenticateUser}
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)
Sys.setenv(TZ = "UTC")
gwData <- fromJSON(system.file('extdata','uvhydrograph','uvhydro-groundwater.json', package = 'repgen'))
uvhydrograph(gwData, 'Author Name')
}

