% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-uvhydrograph.R
\docType{methods}
\name{uvhydrograph}
\alias{uvhydrograph}
\alias{uvhydrograph,list-method}
\alias{uvhydrograph}
\title{UV Hydrograph Report}
\usage{
uvhydrograph(data, ...)

\S4method{uvhydrograph}{list}(data, ...)
}
\arguments{
\item{data}{Local data (as list), or URL.}

\item{...}{Additional parameters passed to GET.}
}
\description{
UV Hydrograph Report
}
\examples{
library(gsplot)
library(jsonlite)
library(lubridate)
library(dplyr)

gwData <-
  fromJSON(
    system.file(
      'extdata', 'uvhydrograph', 'uvhydro-groundwater.json', package = 'repgen'
    )
  )
 uvhydrograph(gwData, 'Author Name')
}
