% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{uvhydrographPlot}
\alias{uvhydrographPlot}
\title{Create a Unit Values Hydrograph}
\usage{
uvhydrographPlot(data)
}
\arguments{
\item{data}{Unit values data structure.}
}
\description{
Create a Unit Values Hydrograph
}

