% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-render.R
\name{uvhydrographPlot}
\alias{uvhydrographPlot}
\title{Create a Unit Values Hydrograph}
\usage{
uvhydrographPlot(reportObject)
}
\arguments{
\item{reportObject}{full UV hydro report data structure.}
}
\value{
list of all elements to be individually rendered
}
\description{
Create a Unit Values Hydrograph
}
