% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vdiagram.R
\docType{methods}
\name{vdiagram}
\alias{vdiagram}
\alias{vdiagram,character,character-method}
\alias{vdiagram,list,character-method}
\alias{vdiagram,list,missing-method}
\title{v-diagram report}
\usage{
vdiagram(data, output)

\S4method{vdiagram}{list,character}(data, output)

\S4method{vdiagram}{character,character}(data, output)

\S4method{vdiagram}{list,missing}(data)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}
}
\description{
v-diagram report
}
\examples{
library(jsonlite)
json_file <- system.file('extdata','vdiagram-v6.json', package = 'repgen')
data <-fromJSON(json_file)
vdiagram(data, 'html')
vdiagram(data, 'pdf')
}

