% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sig-vdiagram.R
\docType{methods}
\name{vdiagram}
\alias{vdiagram}
\alias{vdiagram,character,character-method}
\alias{vdiagram,character,missing-method}
\alias{vdiagram,list,character-method}
\alias{vdiagram,list,missing-method}
\title{v-diagram report}
\usage{
vdiagram(data, output, ...)

\S4method{vdiagram}{list,character}(data, output, ...)

\S4method{vdiagram}{list,missing}(data, output, ...)

\S4method{vdiagram}{character,character}(data, output, ...)

\S4method{vdiagram}{character,missing}(data, output, ...)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}

\item{...}{addtional params passed to GET or authenticateUser}
}
\description{
v-diagram report
}
\examples{
library(jsonlite)
json_file <- system.file('extdata','vdiagram-example.json', package = 'repgen')
data <-fromJSON(json_file)
vdiagram(data, 'html', 'Author Name')
vdiagram(data, 'pdf', 'Author Name')
\dontrun{
url <- paste0('http://nwissddvasvis01.cr.usgs.gov/service/timeseries/reports/swreviewvdiagram/?',
'station=01350000&dischargeIdentifier=Discharge.ft\%5E3\%2Fs&stageIdentifier=',
 'Gage+height.ft.Work.DD002&dailyDischargeIdentifier=',
 'Discharge.ft\%5E3\%2Fs.Mean&ratingModelIdentifier=Gage+height-Discharge.STGQ&waterYear=2014')
vdiagram(data = url, verbose = TRUE) # plot to screen with auth
}
}

