% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vdiagram.R
\name{vdiagram}
\alias{vdiagram}
\title{v-diagram report}
\usage{
vdiagram(data, output)
}
\arguments{
\item{data}{local data (as list) or URL}

\item{output}{a supported pandoc output format (see \code{system("pandoc -h")} for options)}
}
\description{
v-diagram report
}
\examples{
library(jsonlite)
json_file <- system.file('extdata','vdiagram_example.json', package = 'repgen')
data <-fromJSON(json_file)
vdiagram(data, 'html')
vdiagram(data, 'pdf')
}

