% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{vdiagramTable}
\alias{vdiagramTable}
\title{V diagram table from data inputs}
\usage{
vdiagramTable(data)
}
\arguments{
\item{data}{a list of properly formatted v-diagram data}
}
\value{
a string properly formatted for html
}

