% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiagram-render.R
\name{vdiagramTable}
\alias{vdiagramTable}
\title{V diagram table from data inputs}
\usage{
vdiagramTable(data, output)
}
\arguments{
\item{data}{a list of properly formatted v-diagram data}

\item{output}{output type for table. ('html', others supported by \code{\link[knitr]{kable}]})}
}
\value{
a string properly formatted for the specified output type
}

