% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vdiagram-fnc.R
\name{vdiagramTable}
\alias{vdiagramTable}
\title{v-diagram table from data inputs}
\usage{
vdiagramTable(data, output)
}
\arguments{
\item{data}{a list of properly formatted v-diagram data}

\item{output}{output type for table. ('html','pdf', others supported by \code{\link[knitr]{kable}]})}
}
\value{
a string properly formatted for the specified output type
}
\description{
v-diagram table from data inputs
}

