% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{XAxisLabels}
\alias{XAxisLabels}
\title{XAxisLabels}
\usage{
XAxisLabels(object, text, at.months, at.years)
}
\arguments{
\item{object}{A gsplot, plot object.}

\item{text}{Vector of month abbreviations.}

\item{at.months}{Vector of month dates to label month abbreviations (in
"text" vector) at.}

\item{at.years}{Vector of dates to label years at.}
}
\value{
The passed-in gsplot object, with x-axis labeled.
}
\description{
Add x-axis labels to five year GW summary plots, and DV hydrographs
having time intervals of one year or more.
}
\details{
Add x-axis labels
}
