% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvhydrograph-data.R
\name{xposGroupValue}
\alias{xposGroupValue}
\title{X Position of Group Value}
\usage{
xposGroupValue(data, prev, r, build_vec, vars)
}
\description{
helper function to figure out the x value (time) of where to place group value ???
TODO
}

