% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{zeroValues}
\alias{zeroValues}
\title{Finds if the plot data has any zero values.}
\usage{
zeroValues(data, val_nm)
}
\arguments{
\item{data}{A time series data structure, as sequence of (time,value) pairs.}

\item{val_nm}{A field name.}
}
\description{
Finds if the plot data has any zero values.
}

